refactor!: 把业务代码搬进 src/boss_zhipin/ 包，为 Phase D standalone bundle 铺路
为什么必须做：Phase D 用 pytauri standalone 把 Python 代码嵌进 .app
工作机制是 ``uv pip install ./src-tauri``，**前提是被装的 Python 代码
是个可安装 package**。当前散布在仓库根目录的 audit/ gui/ models/ utils/
website_oper/ vectorization.py 是 namespace dirs 不是 package，配
``[tool.uv] package = false``，pip install 装不进去。

这个 commit **不写一行新逻辑**，纯结构重构 + import 路径改写：

结构调整
--------
- ``git mv`` 把 audit/ gui/ models/ utils/ website_oper/ vectorization.py
  boss_tauri/ main.py 全部挪到 src/boss_zhipin/ 下；boss_tauri/ 改名
  src/boss_zhipin/tauri/；main.py 改名 src/boss_zhipin/cli.py。
- 新 src/boss_zhipin/__init__.py（package docstring）和 __main__.py
  （python -m boss_zhipin 入口）。
- cli.py 把原本散在模块顶层的 load_dotenv() 和 logging.basicConfig() 收
  进新加的 _cli_main() 函数，纯 import boss_zhipin.cli 不再有 side effect。
- 根目录 main.py 保留为 3 行向后兼容 shim，``uv run main.py`` 仍然能跑。

build 配置
----------
- pyproject.toml 删 [tool.uv] package = false，加 [build-system]
  (setuptools >= 80) + [tool.setuptools.packages.find] where = ["src"]。
  ``uv sync`` 会自动把 boss_zhipin editable install 进 .venv。
- 加 [project.scripts] boss-zhipin = "boss_zhipin.cli:_cli_main"。
- conftest.py 删 sys.path 注入（包装好了不需要）。

import 路径改写
---------------
所有 ``from {audit,gui,models,utils,website_oper,vectorization} import X``
→ ``from boss_zhipin.X import Y``。涉及：
- src/boss_zhipin/* 内部互相 import
- tests/* 全部用例 + 字符串形式的 monkeypatch.setattr("audit.LOG_PATH", ...)
- tests/test_main_helpers.py: ``import main`` → ``from boss_zhipin import cli as main``
- .github/workflows/ci.yml smoke import 全部加 boss_zhipin. 前缀
- tauri-ui/vite.config.ts outDir 从 ../boss_tauri/frontend
  改 ../src/boss_zhipin/tauri/frontend

验证
----
- 87 tests pass
- CI smoke imports 全部 OK（含 ``import main`` shim + ``import boss_zhipin``）
- ``uv run python -m boss_zhipin.tauri`` 桌面 App 起得来、UI 稳定 ≥6s
- ``uv run main.py`` shim 入口 _cli_main 正确委托到 boss_zhipin.cli

BREAKING：
- ``from audit import X`` 之类的直接 import 不再工作；改成
  ``from boss_zhipin.audit import X``。本仓库内已全部同步。
- 如果外部 PR / fork 引用旧路径，需要 rebase 后修 imports。
