+++ src/boss_zhipin/__main__.py
@@ -1,0 +1,5 @@
+"""``python -m boss_zhipin`` 入口——委托给 ``boss_zhipin.cli``。"""
+from boss_zhipin.cli import _cli_main
+
+if __name__ == "__main__":
+    _cli_main()
