+++ src/boss_zhipin/tauri/__main__.py
@@ -1,6 +1,6 @@
 """``python -m boss_tauri`` 入口。"""
 import sys
 
-from boss_tauri import main
+from boss_zhipin.tauri import main
 
 sys.exit(main())
