+++ src/boss_zhipin/tauri/frontend/assets/index-sINFgi6y.js
@@ -1,0 +1,86 @@
+(function(){const y=document.createElement("link").relList;if(y&&y.supports&&y.supports("modulepreload"))return;for(const R of document.querySelectorAll('link[rel="modulepreload"]'))P(R);new MutationObserver(R=>{for(const V of R)if(V.type==="childList")for(const q of V.addedNodes)q.tagName==="LINK"&&q.rel==="modulepreload"&&P(q)}).observe(document,{childList:!0,subtree:!0});function c(R){const V={};return R.integrity&&(V.integrity=R.integrity),R.referrerPolicy&&(V.referrerPolicy=R.referrerPolicy),R.crossOrigin==="use-credentials"?V.credentials="include":R.crossOrigin==="anonymous"?V.credentials="omit":V.credentials="same-origin",V}function P(R){if(R.ep)return;R.ep=!0;const V=c(R);fetch(R.href,V)}})();function Qi(m){return m&&m.__esModule&&Object.prototype.hasOwnProperty.call(m,"default")?m.default:m}var Ii={exports:{}},Cr={},Di={exports:{}},K={};/**
+ * @license React
+ * react.production.min.js
+ *
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */var $a;function ud(){if($a)return K;$a=1;var m=Symbol.for("react.element"),y=Symbol.for("react.portal"),c=Symbol.for("react.fragment"),P=Symbol.for("react.strict_mode"),R=Symbol.for("react.profiler"),V=Symbol.for("react.provider"),q=Symbol.for("react.context"),ne=Symbol.for("react.forward_ref"),$=Symbol.for("react.suspense"),H=Symbol.for("react.memo"),B=Symbol.for("react.lazy"),O=Symbol.iterator;function T(f){return f===null||typeof f!="object"?null:(f=O&&f[O]||f["@@iterator"],typeof f=="function"?f:null)}var Y={isMounted:function(){return!1},enqueueForceUpdate:function(){},enqueueReplaceState:function(){},enqueueSetState:function(){}},G=Object.assign,A={};function W(f,w,Q){this.props=f,this.context=w,this.refs=A,this.updater=Q||Y}W.prototype.isReactComponent={},W.prototype.setState=function(f,w){if(typeof f!="object"&&typeof f!="function"&&f!=null)throw Error("setState(...): takes an object of state variables to update or a function which returns an object of state variables.");this.updater.enqueueSetState(this,f,w,"setState")},W.prototype.forceUpdate=function(f){this.updater.enqueueForceUpdate(this,f,"forceUpdate")};function Z(){}Z.prototype=W.prototype;function ye(f,w,Q){this.props=f,this.context=w,this.refs=A,this.updater=Q||Y}var ke=ye.prototype=new Z;ke.constructor=ye,G(ke,W.prototype),ke.isPureReactComponent=!0;var ie=Array.isArray,Oe=Object.prototype.hasOwnProperty,re={current:null},se={key:!0,ref:!0,__self:!0,__source:!0};function ze(f,w,Q){var X,b={},ee=null,ae=null;if(w!=null)for(X in w.ref!==void 0&&(ae=w.ref),w.key!==void 0&&(ee=""+w.key),w)Oe.call(w,X)&&!se.hasOwnProperty(X)&&(b[X]=w[X]);var le=arguments.length-2;if(le===1)b.children=Q;else if(1<le){for(var pe=Array(le),Ze=0;Ze<le;Ze++)pe[Ze]=arguments[Ze+2];b.children=pe}if(f&&f.defaultProps)for(X in le=f.defaultProps,le)b[X]===void 0&&(b[X]=le[X]);return{$$typeof:m,type:f,key:ee,ref:ae,props:b,_owner:re.current}}function vt(f,w){return{$$typeof:m,type:f.type,key:w,ref:f.ref,props:f.props,_owner:f._owner}}function st(f){return typeof f=="object"&&f!==null&&f.$$typeof===m}function M(f){var w={"=":"=0",":":"=2"};return"$"+f.replace(/[=:]/g,function(Q){return w[Q]})}var Be=/\/+/g;function Ae(f,w){return typeof f=="object"&&f!==null&&f.key!=null?M(""+f.key):w.toString(36)}function Re(f,w,Q,X,b){var ee=typeof f;(ee==="undefined"||ee==="boolean")&&(f=null);var ae=!1;if(f===null)ae=!0;else switch(ee){case"string":case"number":ae=!0;break;case"object":switch(f.$$typeof){case m:case y:ae=!0}}if(ae)return ae=f,b=b(ae),f=X===""?"."+Ae(ae,0):X,ie(b)?(Q="",f!=null&&(Q=f.replace(Be,"$&/")+"/"),Re(b,w,Q,"",function(Ze){return Ze})):b!=null&&(st(b)&&(b=vt(b,Q+(!b.key||ae&&ae.key===b.key?"":(""+b.key).replace(Be,"$&/")+"/")+f)),w.push(b)),1;if(ae=0,X=X===""?".":X+":",ie(f))for(var le=0;le<f.length;le++){ee=f[le];var pe=X+Ae(ee,le);ae+=Re(ee,w,Q,pe,b)}else if(pe=T(f),typeof pe=="function")for(f=pe.call(f),le=0;!(ee=f.next()).done;)ee=ee.value,pe=X+Ae(ee,le++),ae+=Re(ee,w,Q,pe,b);else if(ee==="object")throw w=String(f),Error("Objects are not valid as a React child (found: "+(w==="[object Object]"?"object with keys {"+Object.keys(f).join(", ")+"}":w)+"). If you meant to render a collection of children, use an array instead.");return ae}function yt(f,w,Q){if(f==null)return f;var X=[],b=0;return Re(f,X,"","",function(ee){return w.call(Q,ee,b++)}),X}function We(f){if(f._status===-1){var w=f._result;w=w(),w.then(function(Q){(f._status===0||f._status===-1)&&(f._status=1,f._result=Q)},function(Q){(f._status===0||f._status===-1)&&(f._status=2,f._result=Q)}),f._status===-1&&(f._status=0,f._result=w)}if(f._status===1)return f._result.default;throw f._result}var ge={current:null},E={transition:null},F={ReactCurrentDispatcher:ge,ReactCurrentBatchConfig:E,ReactCurrentOwner:re};function N(){throw Error("act(...) is not supported in production builds of React.")}return K.Children={map:yt,forEach:function(f,w,Q){yt(f,function(){w.apply(this,arguments)},Q)},count:function(f){var w=0;return yt(f,function(){w++}),w},toArray:function(f){return yt(f,function(w){return w})||[]},only:function(f){if(!st(f))throw Error("React.Children.only expected to receive a single React element child.");return f}},K.Component=W,K.Fragment=c,K.Profiler=R,K.PureComponent=ye,K.StrictMode=P,K.Suspense=$,K.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED=F,K.act=N,K.cloneElement=function(f,w,Q){if(f==null)throw Error("React.cloneElement(...): The argument must be a React element, but you passed "+f+".");var X=G({},f.props),b=f.key,ee=f.ref,ae=f._owner;if(w!=null){if(w.ref!==void 0&&(ee=w.ref,ae=re.current),w.key!==void 0&&(b=""+w.key),f.type&&f.type.defaultProps)var le=f.type.defaultProps;for(pe in w)Oe.call(w,pe)&&!se.hasOwnProperty(pe)&&(X[pe]=w[pe]===void 0&&le!==void 0?le[pe]:w[pe])}var pe=arguments.length-2;if(pe===1)X.children=Q;else if(1<pe){le=Array(pe);for(var Ze=0;Ze<pe;Ze++)le[Ze]=arguments[Ze+2];X.children=le}return{$$typeof:m,type:f.type,key:b,ref:ee,props:X,_owner:ae}},K.createContext=function(f){return f={$$typeof:q,_currentValue:f,_currentValue2:f,_threadCount:0,Provider:null,Consumer:null,_defaultValue:null,_globalName:null},f.Provider={$$typeof:V,_context:f},f.Consumer=f},K.createElement=ze,K.createFactory=function(f){var w=ze.bind(null,f);return w.type=f,w},K.createRef=function(){return{current:null}},K.forwardRef=function(f){return{$$typeof:ne,render:f}},K.isValidElement=st,K.lazy=function(f){return{$$typeof:B,_payload:{_status:-1,_result:f},_init:We}},K.memo=function(f,w){return{$$typeof:H,type:f,compare:w===void 0?null:w}},K.startTransition=function(f){var w=E.transition;E.transition={};try{f()}finally{E.transition=w}},K.unstable_act=N,K.useCallback=function(f,w){return ge.current.useCallback(f,w)},K.useContext=function(f){return ge.current.useContext(f)},K.useDebugValue=function(){},K.useDeferredValue=function(f){return ge.current.useDeferredValue(f)},K.useEffect=function(f,w){return ge.current.useEffect(f,w)},K.useId=function(){return ge.current.useId()},K.useImperativeHandle=function(f,w,Q){return ge.current.useImperativeHandle(f,w,Q)},K.useInsertionEffect=function(f,w){return ge.current.useInsertionEffect(f,w)},K.useLayoutEffect=function(f,w){return ge.current.useLayoutEffect(f,w)},K.useMemo=function(f,w){return ge.current.useMemo(f,w)},K.useReducer=function(f,w,Q){return ge.current.useReducer(f,w,Q)},K.useRef=function(f){return ge.current.useRef(f)},K.useState=function(f){return ge.current.useState(f)},K.useSyncExternalStore=function(f,w,Q){return ge.current.useSyncExternalStore(f,w,Q)},K.useTransition=function(){return ge.current.useTransition()},K.version="18.3.1",K}var Ba;function Pr(){return Ba||(Ba=1,Di.exports=ud()),Di.exports}/**
+ * @license React
+ * react-jsx-runtime.production.min.js
+ *
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */var Wa;function id(){if(Wa)return Cr;Wa=1;var m=Pr(),y=Symbol.for("react.element"),c=Symbol.for("react.fragment"),P=Object.prototype.hasOwnProperty,R=m.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED.ReactCurrentOwner,V={key:!0,ref:!0,__self:!0,__source:!0};function q(ne,$,H){var B,O={},T=null,Y=null;H!==void 0&&(T=""+H),$.key!==void 0&&(T=""+$.key),$.ref!==void 0&&(Y=$.ref);for(B in $)P.call($,B)&&!V.hasOwnProperty(B)&&(O[B]=$[B]);if(ne&&ne.defaultProps)for(B in $=ne.defaultProps,$)O[B]===void 0&&(O[B]=$[B]);return{$$typeof:y,type:ne,key:T,ref:Y,props:O,_owner:R.current}}return Cr.Fragment=c,Cr.jsx=q,Cr.jsxs=q,Cr}var Ha;function od(){return Ha||(Ha=1,Ii.exports=id()),Ii.exports}var v=od(),oe=Pr();const ic=Qi(oe);var Ul={},Mi={exports:{}},Ge={},Fi={exports:{}},Ui={};/**
+ * @license React
+ * scheduler.production.min.js
+ *
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */var Qa;function sd(){return Qa||(Qa=1,(function(m){function y(E,F){var N=E.length;E.push(F);e:for(;0<N;){var f=N-1>>>1,w=E[f];if(0<R(w,F))E[f]=F,E[N]=w,N=f;else break e}}function c(E){return E.length===0?null:E[0]}function P(E){if(E.length===0)return null;var F=E[0],N=E.pop();if(N!==F){E[0]=N;e:for(var f=0,w=E.length,Q=w>>>1;f<Q;){var X=2*(f+1)-1,b=E[X],ee=X+1,ae=E[ee];if(0>R(b,N))ee<w&&0>R(ae,b)?(E[f]=ae,E[ee]=N,f=ee):(E[f]=b,E[X]=N,f=X);else if(ee<w&&0>R(ae,N))E[f]=ae,E[ee]=N,f=ee;else break e}}return F}function R(E,F){var N=E.sortIndex-F.sortIndex;return N!==0?N:E.id-F.id}if(typeof performance=="object"&&typeof performance.now=="function"){var V=performance;m.unstable_now=function(){return V.now()}}else{var q=Date,ne=q.now();m.unstable_now=function(){return q.now()-ne}}var $=[],H=[],B=1,O=null,T=3,Y=!1,G=!1,A=!1,W=typeof setTimeout=="function"?setTimeout:null,Z=typeof clearTimeout=="function"?clearTimeout:null,ye=typeof setImmediate<"u"?setImmediate:null;typeof navigator<"u"&&navigator.scheduling!==void 0&&navigator.scheduling.isInputPending!==void 0&&navigator.scheduling.isInputPending.bind(navigator.scheduling);function ke(E){for(var F=c(H);F!==null;){if(F.callback===null)P(H);else if(F.startTime<=E)P(H),F.sortIndex=F.expirationTime,y($,F);else break;F=c(H)}}function ie(E){if(A=!1,ke(E),!G)if(c($)!==null)G=!0,We(Oe);else{var F=c(H);F!==null&&ge(ie,F.startTime-E)}}function Oe(E,F){G=!1,A&&(A=!1,Z(ze),ze=-1),Y=!0;var N=T;try{for(ke(F),O=c($);O!==null&&(!(O.expirationTime>F)||E&&!M());){var f=O.callback;if(typeof f=="function"){O.callback=null,T=O.priorityLevel;var w=f(O.expirationTime<=F);F=m.unstable_now(),typeof w=="function"?O.callback=w:O===c($)&&P($),ke(F)}else P($);O=c($)}if(O!==null)var Q=!0;else{var X=c(H);X!==null&&ge(ie,X.startTime-F),Q=!1}return Q}finally{O=null,T=N,Y=!1}}var re=!1,se=null,ze=-1,vt=5,st=-1;function M(){return!(m.unstable_now()-st<vt)}function Be(){if(se!==null){var E=m.unstable_now();st=E;var F=!0;try{F=se(!0,E)}finally{F?Ae():(re=!1,se=null)}}else re=!1}var Ae;if(typeof ye=="function")Ae=function(){ye(Be)};else if(typeof MessageChannel<"u"){var Re=new MessageChannel,yt=Re.port2;Re.port1.onmessage=Be,Ae=function(){yt.postMessage(null)}}else Ae=function(){W(Be,0)};function We(E){se=E,re||(re=!0,Ae())}function ge(E,F){ze=W(function(){E(m.unstable_now())},F)}m.unstable_IdlePriority=5,m.unstable_ImmediatePriority=1,m.unstable_LowPriority=4,m.unstable_NormalPriority=3,m.unstable_Profiling=null,m.unstable_UserBlockingPriority=2,m.unstable_cancelCallback=function(E){E.callback=null},m.unstable_continueExecution=function(){G||Y||(G=!0,We(Oe))},m.unstable_forceFrameRate=function(E){0>E||125<E?console.error("forceFrameRate takes a positive int between 0 and 125, forcing frame rates higher than 125 fps is not supported"):vt=0<E?Math.floor(1e3/E):5},m.unstable_getCurrentPriorityLevel=function(){return T},m.unstable_getFirstCallbackNode=function(){return c($)},m.unstable_next=function(E){switch(T){case 1:case 2:case 3:var F=3;break;default:F=T}var N=T;T=F;try{return E()}finally{T=N}},m.unstable_pauseExecution=function(){},m.unstable_requestPaint=function(){},m.unstable_runWithPriority=function(E,F){switch(E){case 1:case 2:case 3:case 4:case 5:break;default:E=3}var N=T;T=E;try{return F()}finally{T=N}},m.unstable_scheduleCallback=function(E,F,N){var f=m.unstable_now();switch(typeof N=="object"&&N!==null?(N=N.delay,N=typeof N=="number"&&0<N?f+N:f):N=f,E){case 1:var w=-1;break;case 2:w=250;break;case 5:w=1073741823;break;case 4:w=1e4;break;default:w=5e3}return w=N+w,E={id:B++,callback:F,priorityLevel:E,startTime:N,expirationTime:w,sortIndex:-1},N>f?(E.sortIndex=N,y(H,E),c($)===null&&E===c(H)&&(A?(Z(ze),ze=-1):A=!0,ge(ie,N-f))):(E.sortIndex=w,y($,E),G||Y||(G=!0,We(Oe))),E},m.unstable_shouldYield=M,m.unstable_wrapCallback=function(E){var F=T;return function(){var N=T;T=F;try{return E.apply(this,arguments)}finally{T=N}}}})(Ui)),Ui}var Ka;function ad(){return Ka||(Ka=1,Fi.exports=sd()),Fi.exports}/**
+ * @license React
+ * react-dom.production.min.js
+ *
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */var Ya;function cd(){if(Ya)return Ge;Ya=1;var m=Pr(),y=ad();function c(e){for(var t="https://reactjs.org/docs/error-decoder.html?invariant="+e,n=1;n<arguments.length;n++)t+="&args[]="+encodeURIComponent(arguments[n]);return"Minified React error #"+e+"; visit "+t+" for the full message or use the non-minified dev environment for full errors and additional helpful warnings."}var P=new Set,R={};function V(e,t){q(e,t),q(e+"Capture",t)}function q(e,t){for(R[e]=t,e=0;e<t.length;e++)P.add(t[e])}var ne=!(typeof window>"u"||typeof window.document>"u"||typeof window.document.createElement>"u"),$=Object.prototype.hasOwnProperty,H=/^[:A-Z_a-z\u00C0-\u00D6\u00D8-\u00F6\u00F8-\u02FF\u0370-\u037D\u037F-\u1FFF\u200C-\u200D\u2070-\u218F\u2C00-\u2FEF\u3001-\uD7FF\uF900-\uFDCF\uFDF0-\uFFFD][:A-Z_a-z\u00C0-\u00D6\u00D8-\u00F6\u00F8-\u02FF\u0370-\u037D\u037F-\u1FFF\u200C-\u200D\u2070-\u218F\u2C00-\u2FEF\u3001-\uD7FF\uF900-\uFDCF\uFDF0-\uFFFD\-.0-9\u00B7\u0300-\u036F\u203F-\u2040]*$/,B={},O={};function T(e){return $.call(O,e)?!0:$.call(B,e)?!1:H.test(e)?O[e]=!0:(B[e]=!0,!1)}function Y(e,t,n,r){if(n!==null&&n.type===0)return!1;switch(typeof t){case"function":case"symbol":return!0;case"boolean":return r?!1:n!==null?!n.acceptsBooleans:(e=e.toLowerCase().slice(0,5),e!=="data-"&&e!=="aria-");default:return!1}}function G(e,t,n,r){if(t===null||typeof t>"u"||Y(e,t,n,r))return!0;if(r)return!1;if(n!==null)switch(n.type){case 3:return!t;case 4:return t===!1;case 5:return isNaN(t);case 6:return isNaN(t)||1>t}return!1}function A(e,t,n,r,l,u,i){this.acceptsBooleans=t===2||t===3||t===4,this.attributeName=r,this.attributeNamespace=l,this.mustUseProperty=n,this.propertyName=e,this.type=t,this.sanitizeURL=u,this.removeEmptyString=i}var W={};"children dangerouslySetInnerHTML defaultValue defaultChecked innerHTML suppressContentEditableWarning suppressHydrationWarning style".split(" ").forEach(function(e){W[e]=new A(e,0,!1,e,null,!1,!1)}),[["acceptCharset","accept-charset"],["className","class"],["htmlFor","for"],["httpEquiv","http-equiv"]].forEach(function(e){var t=e[0];W[t]=new A(t,1,!1,e[1],null,!1,!1)}),["contentEditable","draggable","spellCheck","value"].forEach(function(e){W[e]=new A(e,2,!1,e.toLowerCase(),null,!1,!1)}),["autoReverse","externalResourcesRequired","focusable","preserveAlpha"].forEach(function(e){W[e]=new A(e,2,!1,e,null,!1,!1)}),"allowFullScreen async autoFocus autoPlay controls default defer disabled disablePictureInPicture disableRemotePlayback formNoValidate hidden loop noModule noValidate open playsInline readOnly required reversed scoped seamless itemScope".split(" ").forEach(function(e){W[e]=new A(e,3,!1,e.toLowerCase(),null,!1,!1)}),["checked","multiple","muted","selected"].forEach(function(e){W[e]=new A(e,3,!0,e,null,!1,!1)}),["capture","download"].forEach(function(e){W[e]=new A(e,4,!1,e,null,!1,!1)}),["cols","rows","size","span"].forEach(function(e){W[e]=new A(e,6,!1,e,null,!1,!1)}),["rowSpan","start"].forEach(function(e){W[e]=new A(e,5,!1,e.toLowerCase(),null,!1,!1)});var Z=/[\-:]([a-z])/g;function ye(e){return e[1].toUpperCase()}"accent-height alignment-baseline arabic-form baseline-shift cap-height clip-path clip-rule color-interpolation color-interpolation-filters color-profile color-rendering dominant-baseline enable-background fill-opacity fill-rule flood-color flood-opacity font-family font-size font-size-adjust font-stretch font-style font-variant font-weight glyph-name glyph-orientation-horizontal glyph-orientation-vertical horiz-adv-x horiz-origin-x image-rendering letter-spacing lighting-color marker-end marker-mid marker-start overline-position overline-thickness paint-order panose-1 pointer-events rendering-intent shape-rendering stop-color stop-opacity strikethrough-position strikethrough-thickness stroke-dasharray stroke-dashoffset stroke-linecap stroke-linejoin stroke-miterlimit stroke-opacity stroke-width text-anchor text-decoration text-rendering underline-position underline-thickness unicode-bidi unicode-range units-per-em v-alphabetic v-hanging v-ideographic v-mathematical vector-effect vert-adv-y vert-origin-x vert-origin-y word-spacing writing-mode xmlns:xlink x-height".split(" ").forEach(function(e){var t=e.replace(Z,ye);W[t]=new A(t,1,!1,e,null,!1,!1)}),"xlink:actuate xlink:arcrole xlink:role xlink:show xlink:title xlink:type".split(" ").forEach(function(e){var t=e.replace(Z,ye);W[t]=new A(t,1,!1,e,"http://www.w3.org/1999/xlink",!1,!1)}),["xml:base","xml:lang","xml:space"].forEach(function(e){var t=e.replace(Z,ye);W[t]=new A(t,1,!1,e,"http://www.w3.org/XML/1998/namespace",!1,!1)}),["tabIndex","crossOrigin"].forEach(function(e){W[e]=new A(e,1,!1,e.toLowerCase(),null,!1,!1)}),W.xlinkHref=new A("xlinkHref",1,!1,"xlink:href","http://www.w3.org/1999/xlink",!0,!1),["src","href","action","formAction"].forEach(function(e){W[e]=new A(e,1,!1,e.toLowerCase(),null,!0,!0)});function ke(e,t,n,r){var l=W.hasOwnProperty(t)?W[t]:null;(l!==null?l.type!==0:r||!(2<t.length)||t[0]!=="o"&&t[0]!=="O"||t[1]!=="n"&&t[1]!=="N")&&(G(t,n,l,r)&&(n=null),r||l===null?T(t)&&(n===null?e.removeAttribute(t):e.setAttribute(t,""+n)):l.mustUseProperty?e[l.propertyName]=n===null?l.type===3?!1:"":n:(t=l.attributeName,r=l.attributeNamespace,n===null?e.removeAttribute(t):(l=l.type,n=l===3||l===4&&n===!0?"":""+n,r?e.setAttributeNS(r,t,n):e.setAttribute(t,n))))}var ie=m.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED,Oe=Symbol.for("react.element"),re=Symbol.for("react.portal"),se=Symbol.for("react.fragment"),ze=Symbol.for("react.strict_mode"),vt=Symbol.for("react.profiler"),st=Symbol.for("react.provider"),M=Symbol.for("react.context"),Be=Symbol.for("react.forward_ref"),Ae=Symbol.for("react.suspense"),Re=Symbol.for("react.suspense_list"),yt=Symbol.for("react.memo"),We=Symbol.for("react.lazy"),ge=Symbol.for("react.offscreen"),E=Symbol.iterator;function F(e){return e===null||typeof e!="object"?null:(e=E&&e[E]||e["@@iterator"],typeof e=="function"?e:null)}var N=Object.assign,f;function w(e){if(f===void 0)try{throw Error()}catch(n){var t=n.stack.trim().match(/\n( *(at )?)/);f=t&&t[1]||""}return`
+`+f+e}var Q=!1;function X(e,t){if(!e||Q)return"";Q=!0;var n=Error.prepareStackTrace;Error.prepareStackTrace=void 0;try{if(t)if(t=function(){throw Error()},Object.defineProperty(t.prototype,"props",{set:function(){throw Error()}}),typeof Reflect=="object"&&Reflect.construct){try{Reflect.construct(t,[])}catch(h){var r=h}Reflect.construct(e,[],t)}else{try{t.call()}catch(h){r=h}e.call(t.prototype)}else{try{throw Error()}catch(h){r=h}e()}}catch(h){if(h&&r&&typeof h.stack=="string"){for(var l=h.stack.split(`
+`),u=r.stack.split(`
+`),i=l.length-1,o=u.length-1;1<=i&&0<=o&&l[i]!==u[o];)o--;for(;1<=i&&0<=o;i--,o--)if(l[i]!==u[o]){if(i!==1||o!==1)do if(i--,o--,0>o||l[i]!==u[o]){var s=`
+`+l[i].replace(" at new "," at ");return e.displayName&&s.includes("<anonymous>")&&(s=s.replace("<anonymous>",e.displayName)),s}while(1<=i&&0<=o);break}}}finally{Q=!1,Error.prepareStackTrace=n}return(e=e?e.displayName||e.name:"")?w(e):""}function b(e){switch(e.tag){case 5:return w(e.type);case 16:return w("Lazy");case 13:return w("Suspense");case 19:return w("SuspenseList");case 0:case 2:case 15:return e=X(e.type,!1),e;case 11:return e=X(e.type.render,!1),e;case 1:return e=X(e.type,!0),e;default:return""}}function ee(e){if(e==null)return null;if(typeof e=="function")return e.displayName||e.name||null;if(typeof e=="string")return e;switch(e){case se:return"Fragment";case re:return"Portal";case vt:return"Profiler";case ze:return"StrictMode";case Ae:return"Suspense";case Re:return"SuspenseList"}if(typeof e=="object")switch(e.$$typeof){case M:return(e.displayName||"Context")+".Consumer";case st:return(e._context.displayName||"Context")+".Provider";case Be:var t=e.render;return e=e.displayName,e||(e=t.displayName||t.name||"",e=e!==""?"ForwardRef("+e+")":"ForwardRef"),e;case yt:return t=e.displayName||null,t!==null?t:ee(e.type)||"Memo";case We:t=e._payload,e=e._init;try{return ee(e(t))}catch{}}return null}function ae(e){var t=e.type;switch(e.tag){case 24:return"Cache";case 9:return(t.displayName||"Context")+".Consumer";case 10:return(t._context.displayName||"Context")+".Provider";case 18:return"DehydratedFragment";case 11:return e=t.render,e=e.displayName||e.name||"",t.displayName||(e!==""?"ForwardRef("+e+")":"ForwardRef");case 7:return"Fragment";case 5:return t;case 4:return"Portal";case 3:return"Root";case 6:return"Text";case 16:return ee(t);case 8:return t===ze?"StrictMode":"Mode";case 22:return"Offscreen";case 12:return"Profiler";case 21:return"Scope";case 13:return"Suspense";case 19:return"SuspenseList";case 25:return"TracingMarker";case 1:case 0:case 17:case 2:case 14:case 15:if(typeof t=="function")return t.displayName||t.name||null;if(typeof t=="string")return t}return null}function le(e){switch(typeof e){case"boolean":case"number":case"string":case"undefined":return e;case"object":return e;default:return""}}function pe(e){var t=e.type;return(e=e.nodeName)&&e.toLowerCase()==="input"&&(t==="checkbox"||t==="radio")}function Ze(e){var t=pe(e)?"checked":"value",n=Object.getOwnPropertyDescriptor(e.constructor.prototype,t),r=""+e[t];if(!e.hasOwnProperty(t)&&typeof n<"u"&&typeof n.get=="function"&&typeof n.set=="function"){var l=n.get,u=n.set;return Object.defineProperty(e,t,{configurable:!0,get:function(){return l.call(this)},set:function(i){r=""+i,u.call(this,i)}}),Object.defineProperty(e,t,{enumerable:n.enumerable}),{getValue:function(){return r},setValue:function(i){r=""+i},stopTracking:function(){e._valueTracker=null,delete e[t]}}}}function jr(e){e._valueTracker||(e._valueTracker=Ze(e))}function Ki(e){if(!e)return!1;var t=e._valueTracker;if(!t)return!0;var n=t.getValue(),r="";return e&&(r=pe(e)?e.checked?"true":"false":e.value),e=r,e!==n?(t.setValue(e),!0):!1}function zr(e){if(e=e||(typeof document<"u"?document:void 0),typeof e>"u")return null;try{return e.activeElement||e.body}catch{return e.body}}function $l(e,t){var n=t.checked;return N({},t,{defaultChecked:void 0,defaultValue:void 0,value:void 0,checked:n??e._wrapperState.initialChecked})}function Yi(e,t){var n=t.defaultValue==null?"":t.defaultValue,r=t.checked!=null?t.checked:t.defaultChecked;n=le(t.value!=null?t.value:n),e._wrapperState={initialChecked:r,initialValue:n,controlled:t.type==="checkbox"||t.type==="radio"?t.checked!=null:t.value!=null}}function Xi(e,t){t=t.checked,t!=null&&ke(e,"checked",t,!1)}function Bl(e,t){Xi(e,t);var n=le(t.value),r=t.type;if(n!=null)r==="number"?(n===0&&e.value===""||e.value!=n)&&(e.value=""+n):e.value!==""+n&&(e.value=""+n);else if(r==="submit"||r==="reset"){e.removeAttribute("value");return}t.hasOwnProperty("value")?Wl(e,t.type,n):t.hasOwnProperty("defaultValue")&&Wl(e,t.type,le(t.defaultValue)),t.checked==null&&t.defaultChecked!=null&&(e.defaultChecked=!!t.defaultChecked)}function Gi(e,t,n){if(t.hasOwnProperty("value")||t.hasOwnProperty("defaultValue")){var r=t.type;if(!(r!=="submit"&&r!=="reset"||t.value!==void 0&&t.value!==null))return;t=""+e._wrapperState.initialValue,n||t===e.value||(e.value=t),e.defaultValue=t}n=e.name,n!==""&&(e.name=""),e.defaultChecked=!!e._wrapperState.initialChecked,n!==""&&(e.name=n)}function Wl(e,t,n){(t!=="number"||zr(e.ownerDocument)!==e)&&(n==null?e.defaultValue=""+e._wrapperState.initialValue:e.defaultValue!==""+n&&(e.defaultValue=""+n))}var Vn=Array.isArray;function mn(e,t,n,r){if(e=e.options,t){t={};for(var l=0;l<n.length;l++)t["$"+n[l]]=!0;for(n=0;n<e.length;n++)l=t.hasOwnProperty("$"+e[n].value),e[n].selected!==l&&(e[n].selected=l),l&&r&&(e[n].defaultSelected=!0)}else{for(n=""+le(n),t=null,l=0;l<e.length;l++){if(e[l].value===n){e[l].selected=!0,r&&(e[l].defaultSelected=!0);return}t!==null||e[l].disabled||(t=e[l])}t!==null&&(t.selected=!0)}}function Hl(e,t){if(t.dangerouslySetInnerHTML!=null)throw Error(c(91));return N({},t,{value:void 0,defaultValue:void 0,children:""+e._wrapperState.initialValue})}function Zi(e,t){var n=t.value;if(n==null){if(n=t.children,t=t.defaultValue,n!=null){if(t!=null)throw Error(c(92));if(Vn(n)){if(1<n.length)throw Error(c(93));n=n[0]}t=n}t==null&&(t=""),n=t}e._wrapperState={initialValue:le(n)}}function Ji(e,t){var n=le(t.value),r=le(t.defaultValue);n!=null&&(n=""+n,n!==e.value&&(e.value=n),t.defaultValue==null&&e.defaultValue!==n&&(e.defaultValue=n)),r!=null&&(e.defaultValue=""+r)}function qi(e){var t=e.textContent;t===e._wrapperState.initialValue&&t!==""&&t!==null&&(e.value=t)}function bi(e){switch(e){case"svg":return"http://www.w3.org/2000/svg";case"math":return"http://www.w3.org/1998/Math/MathML";default:return"http://www.w3.org/1999/xhtml"}}function Ql(e,t){return e==null||e==="http://www.w3.org/1999/xhtml"?bi(t):e==="http://www.w3.org/2000/svg"&&t==="foreignObject"?"http://www.w3.org/1999/xhtml":e}var Rr,eo=(function(e){return typeof MSApp<"u"&&MSApp.execUnsafeLocalFunction?function(t,n,r,l){MSApp.execUnsafeLocalFunction(function(){return e(t,n,r,l)})}:e})(function(e,t){if(e.namespaceURI!=="http://www.w3.org/2000/svg"||"innerHTML"in e)e.innerHTML=t;else{for(Rr=Rr||document.createElement("div"),Rr.innerHTML="<svg>"+t.valueOf().toString()+"</svg>",t=Rr.firstChild;e.firstChild;)e.removeChild(e.firstChild);for(;t.firstChild;)e.appendChild(t.firstChild)}});function $n(e,t){if(t){var n=e.firstChild;if(n&&n===e.lastChild&&n.nodeType===3){n.nodeValue=t;return}}e.textContent=t}var Bn={animationIterationCount:!0,aspectRatio:!0,borderImageOutset:!0,borderImageSlice:!0,borderImageWidth:!0,boxFlex:!0,boxFlexGroup:!0,boxOrdinalGroup:!0,columnCount:!0,columns:!0,flex:!0,flexGrow:!0,flexPositive:!0,flexShrink:!0,flexNegative:!0,flexOrder:!0,gridArea:!0,gridRow:!0,gridRowEnd:!0,gridRowSpan:!0,gridRowStart:!0,gridColumn:!0,gridColumnEnd:!0,gridColumnSpan:!0,gridColumnStart:!0,fontWeight:!0,lineClamp:!0,lineHeight:!0,opacity:!0,order:!0,orphans:!0,tabSize:!0,widows:!0,zIndex:!0,zoom:!0,fillOpacity:!0,floodOpacity:!0,stopOpacity:!0,strokeDasharray:!0,strokeDashoffset:!0,strokeMiterlimit:!0,strokeOpacity:!0,strokeWidth:!0},sc=["Webkit","ms","Moz","O"];Object.keys(Bn).forEach(function(e){sc.forEach(function(t){t=t+e.charAt(0).toUpperCase()+e.substring(1),Bn[t]=Bn[e]})});function to(e,t,n){return t==null||typeof t=="boolean"||t===""?"":n||typeof t!="number"||t===0||Bn.hasOwnProperty(e)&&Bn[e]?(""+t).trim():t+"px"}function no(e,t){e=e.style;for(var n in t)if(t.hasOwnProperty(n)){var r=n.indexOf("--")===0,l=to(n,t[n],r);n==="float"&&(n="cssFloat"),r?e.setProperty(n,l):e[n]=l}}var ac=N({menuitem:!0},{area:!0,base:!0,br:!0,col:!0,embed:!0,hr:!0,img:!0,input:!0,keygen:!0,link:!0,meta:!0,param:!0,source:!0,track:!0,wbr:!0});function Kl(e,t){if(t){if(ac[e]&&(t.children!=null||t.dangerouslySetInnerHTML!=null))throw Error(c(137,e));if(t.dangerouslySetInnerHTML!=null){if(t.children!=null)throw Error(c(60));if(typeof t.dangerouslySetInnerHTML!="object"||!("__html"in t.dangerouslySetInnerHTML))throw Error(c(61))}if(t.style!=null&&typeof t.style!="object")throw Error(c(62))}}function Yl(e,t){if(e.indexOf("-")===-1)return typeof t.is=="string";switch(e){case"annotation-xml":case"color-profile":case"font-face":case"font-face-src":case"font-face-uri":case"font-face-format":case"font-face-name":case"missing-glyph":return!1;default:return!0}}var Xl=null;function Gl(e){return e=e.target||e.srcElement||window,e.correspondingUseElement&&(e=e.correspondingUseElement),e.nodeType===3?e.parentNode:e}var Zl=null,hn=null,vn=null;function ro(e){if(e=cr(e)){if(typeof Zl!="function")throw Error(c(280));var t=e.stateNode;t&&(t=br(t),Zl(e.stateNode,e.type,t))}}function lo(e){hn?vn?vn.push(e):vn=[e]:hn=e}function uo(){if(hn){var e=hn,t=vn;if(vn=hn=null,ro(e),t)for(e=0;e<t.length;e++)ro(t[e])}}function io(e,t){return e(t)}function oo(){}var Jl=!1;function so(e,t,n){if(Jl)return e(t,n);Jl=!0;try{return io(e,t,n)}finally{Jl=!1,(hn!==null||vn!==null)&&(oo(),uo())}}function Wn(e,t){var n=e.stateNode;if(n===null)return null;var r=br(n);if(r===null)return null;n=r[t];e:switch(t){case"onClick":case"onClickCapture":case"onDoubleClick":case"onDoubleClickCapture":case"onMouseDown":case"onMouseDownCapture":case"onMouseMove":case"onMouseMoveCapture":case"onMouseUp":case"onMouseUpCapture":case"onMouseEnter":(r=!r.disabled)||(e=e.type,r=!(e==="button"||e==="input"||e==="select"||e==="textarea")),e=!r;break e;default:e=!1}if(e)return null;if(n&&typeof n!="function")throw Error(c(231,t,typeof n));return n}var ql=!1;if(ne)try{var Hn={};Object.defineProperty(Hn,"passive",{get:function(){ql=!0}}),window.addEventListener("test",Hn,Hn),window.removeEventListener("test",Hn,Hn)}catch{ql=!1}function cc(e,t,n,r,l,u,i,o,s){var h=Array.prototype.slice.call(arguments,3);try{t.apply(n,h)}catch(S){this.onError(S)}}var Qn=!1,Lr=null,Tr=!1,bl=null,fc={onError:function(e){Qn=!0,Lr=e}};function dc(e,t,n,r,l,u,i,o,s){Qn=!1,Lr=null,cc.apply(fc,arguments)}function pc(e,t,n,r,l,u,i,o,s){if(dc.apply(this,arguments),Qn){if(Qn){var h=Lr;Qn=!1,Lr=null}else throw Error(c(198));Tr||(Tr=!0,bl=h)}}function bt(e){var t=e,n=e;if(e.alternate)for(;t.return;)t=t.return;else{e=t;do t=e,(t.flags&4098)!==0&&(n=t.return),e=t.return;while(e)}return t.tag===3?n:null}function ao(e){if(e.tag===13){var t=e.memoizedState;if(t===null&&(e=e.alternate,e!==null&&(t=e.memoizedState)),t!==null)return t.dehydrated}return null}function co(e){if(bt(e)!==e)throw Error(c(188))}function mc(e){var t=e.alternate;if(!t){if(t=bt(e),t===null)throw Error(c(188));return t!==e?null:e}for(var n=e,r=t;;){var l=n.return;if(l===null)break;var u=l.alternate;if(u===null){if(r=l.return,r!==null){n=r;continue}break}if(l.child===u.child){for(u=l.child;u;){if(u===n)return co(l),e;if(u===r)return co(l),t;u=u.sibling}throw Error(c(188))}if(n.return!==r.return)n=l,r=u;else{for(var i=!1,o=l.child;o;){if(o===n){i=!0,n=l,r=u;break}if(o===r){i=!0,r=l,n=u;break}o=o.sibling}if(!i){for(o=u.child;o;){if(o===n){i=!0,n=u,r=l;break}if(o===r){i=!0,r=u,n=l;break}o=o.sibling}if(!i)throw Error(c(189))}}if(n.alternate!==r)throw Error(c(190))}if(n.tag!==3)throw Error(c(188));return n.stateNode.current===n?e:t}function fo(e){return e=mc(e),e!==null?po(e):null}function po(e){if(e.tag===5||e.tag===6)return e;for(e=e.child;e!==null;){var t=po(e);if(t!==null)return t;e=e.sibling}return null}var mo=y.unstable_scheduleCallback,ho=y.unstable_cancelCallback,hc=y.unstable_shouldYield,vc=y.unstable_requestPaint,Se=y.unstable_now,yc=y.unstable_getCurrentPriorityLevel,eu=y.unstable_ImmediatePriority,vo=y.unstable_UserBlockingPriority,Or=y.unstable_NormalPriority,gc=y.unstable_LowPriority,yo=y.unstable_IdlePriority,Ir=null,gt=null;function wc(e){if(gt&&typeof gt.onCommitFiberRoot=="function")try{gt.onCommitFiberRoot(Ir,e,void 0,(e.current.flags&128)===128)}catch{}}var at=Math.clz32?Math.clz32:kc,Sc=Math.log,xc=Math.LN2;function kc(e){return e>>>=0,e===0?32:31-(Sc(e)/xc|0)|0}var Dr=64,Mr=4194304;function Kn(e){switch(e&-e){case 1:return 1;case 2:return 2;case 4:return 4;case 8:return 8;case 16:return 16;case 32:return 32;case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:return e&4194240;case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:return e&130023424;case 134217728:return 134217728;case 268435456:return 268435456;case 536870912:return 536870912;case 1073741824:return 1073741824;default:return e}}function Fr(e,t){var n=e.pendingLanes;if(n===0)return 0;var r=0,l=e.suspendedLanes,u=e.pingedLanes,i=n&268435455;if(i!==0){var o=i&~l;o!==0?r=Kn(o):(u&=i,u!==0&&(r=Kn(u)))}else i=n&~l,i!==0?r=Kn(i):u!==0&&(r=Kn(u));if(r===0)return 0;if(t!==0&&t!==r&&(t&l)===0&&(l=r&-r,u=t&-t,l>=u||l===16&&(u&4194240)!==0))return t;if((r&4)!==0&&(r|=n&16),t=e.entangledLanes,t!==0)for(e=e.entanglements,t&=r;0<t;)n=31-at(t),l=1<<n,r|=e[n],t&=~l;return r}function Ec(e,t){switch(e){case 1:case 2:case 4:return t+250;case 8:case 16:case 32:case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:return t+5e3;case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:return-1;case 134217728:case 268435456:case 536870912:case 1073741824:return-1;default:return-1}}function _c(e,t){for(var n=e.suspendedLanes,r=e.pingedLanes,l=e.expirationTimes,u=e.pendingLanes;0<u;){var i=31-at(u),o=1<<i,s=l[i];s===-1?((o&n)===0||(o&r)!==0)&&(l[i]=Ec(o,t)):s<=t&&(e.expiredLanes|=o),u&=~o}}function tu(e){return e=e.pendingLanes&-1073741825,e!==0?e:e&1073741824?1073741824:0}function go(){var e=Dr;return Dr<<=1,(Dr&4194240)===0&&(Dr=64),e}function nu(e){for(var t=[],n=0;31>n;n++)t.push(e);return t}function Yn(e,t,n){e.pendingLanes|=t,t!==536870912&&(e.suspendedLanes=0,e.pingedLanes=0),e=e.eventTimes,t=31-at(t),e[t]=n}function Cc(e,t){var n=e.pendingLanes&~t;e.pendingLanes=t,e.suspendedLanes=0,e.pingedLanes=0,e.expiredLanes&=t,e.mutableReadLanes&=t,e.entangledLanes&=t,t=e.entanglements;var r=e.eventTimes;for(e=e.expirationTimes;0<n;){var l=31-at(n),u=1<<l;t[l]=0,r[l]=-1,e[l]=-1,n&=~u}}function ru(e,t){var n=e.entangledLanes|=t;for(e=e.entanglements;n;){var r=31-at(n),l=1<<r;l&t|e[r]&t&&(e[r]|=t),n&=~l}}var ue=0;function wo(e){return e&=-e,1<e?4<e?(e&268435455)!==0?16:536870912:4:1}var So,lu,xo,ko,Eo,uu=!1,Ur=[],Ot=null,It=null,Dt=null,Xn=new Map,Gn=new Map,Mt=[],Nc="mousedown mouseup touchcancel touchend touchstart auxclick dblclick pointercancel pointerdown pointerup dragend dragstart drop compositionend compositionstart keydown keypress keyup input textInput copy cut paste click change contextmenu reset submit".split(" ");function _o(e,t){switch(e){case"focusin":case"focusout":Ot=null;break;case"dragenter":case"dragleave":It=null;break;case"mouseover":case"mouseout":Dt=null;break;case"pointerover":case"pointerout":Xn.delete(t.pointerId);break;case"gotpointercapture":case"lostpointercapture":Gn.delete(t.pointerId)}}function Zn(e,t,n,r,l,u){return e===null||e.nativeEvent!==u?(e={blockedOn:t,domEventName:n,eventSystemFlags:r,nativeEvent:u,targetContainers:[l]},t!==null&&(t=cr(t),t!==null&&lu(t)),e):(e.eventSystemFlags|=r,t=e.targetContainers,l!==null&&t.indexOf(l)===-1&&t.push(l),e)}function Pc(e,t,n,r,l){switch(t){case"focusin":return Ot=Zn(Ot,e,t,n,r,l),!0;case"dragenter":return It=Zn(It,e,t,n,r,l),!0;case"mouseover":return Dt=Zn(Dt,e,t,n,r,l),!0;case"pointerover":var u=l.pointerId;return Xn.set(u,Zn(Xn.get(u)||null,e,t,n,r,l)),!0;case"gotpointercapture":return u=l.pointerId,Gn.set(u,Zn(Gn.get(u)||null,e,t,n,r,l)),!0}return!1}function Co(e){var t=en(e.target);if(t!==null){var n=bt(t);if(n!==null){if(t=n.tag,t===13){if(t=ao(n),t!==null){e.blockedOn=t,Eo(e.priority,function(){xo(n)});return}}else if(t===3&&n.stateNode.current.memoizedState.isDehydrated){e.blockedOn=n.tag===3?n.stateNode.containerInfo:null;return}}}e.blockedOn=null}function Ar(e){if(e.blockedOn!==null)return!1;for(var t=e.targetContainers;0<t.length;){var n=ou(e.domEventName,e.eventSystemFlags,t[0],e.nativeEvent);if(n===null){n=e.nativeEvent;var r=new n.constructor(n.type,n);Xl=r,n.target.dispatchEvent(r),Xl=null}else return t=cr(n),t!==null&&lu(t),e.blockedOn=n,!1;t.shift()}return!0}function No(e,t,n){Ar(e)&&n.delete(t)}function jc(){uu=!1,Ot!==null&&Ar(Ot)&&(Ot=null),It!==null&&Ar(It)&&(It=null),Dt!==null&&Ar(Dt)&&(Dt=null),Xn.forEach(No),Gn.forEach(No)}function Jn(e,t){e.blockedOn===t&&(e.blockedOn=null,uu||(uu=!0,y.unstable_scheduleCallback(y.unstable_NormalPriority,jc)))}function qn(e){function t(l){return Jn(l,e)}if(0<Ur.length){Jn(Ur[0],e);for(var n=1;n<Ur.length;n++){var r=Ur[n];r.blockedOn===e&&(r.blockedOn=null)}}for(Ot!==null&&Jn(Ot,e),It!==null&&Jn(It,e),Dt!==null&&Jn(Dt,e),Xn.forEach(t),Gn.forEach(t),n=0;n<Mt.length;n++)r=Mt[n],r.blockedOn===e&&(r.blockedOn=null);for(;0<Mt.length&&(n=Mt[0],n.blockedOn===null);)Co(n),n.blockedOn===null&&Mt.shift()}var yn=ie.ReactCurrentBatchConfig,Vr=!0;function zc(e,t,n,r){var l=ue,u=yn.transition;yn.transition=null;try{ue=1,iu(e,t,n,r)}finally{ue=l,yn.transition=u}}function Rc(e,t,n,r){var l=ue,u=yn.transition;yn.transition=null;try{ue=4,iu(e,t,n,r)}finally{ue=l,yn.transition=u}}function iu(e,t,n,r){if(Vr){var l=ou(e,t,n,r);if(l===null)_u(e,t,r,$r,n),_o(e,r);else if(Pc(l,e,t,n,r))r.stopPropagation();else if(_o(e,r),t&4&&-1<Nc.indexOf(e)){for(;l!==null;){var u=cr(l);if(u!==null&&So(u),u=ou(e,t,n,r),u===null&&_u(e,t,r,$r,n),u===l)break;l=u}l!==null&&r.stopPropagation()}else _u(e,t,r,null,n)}}var $r=null;function ou(e,t,n,r){if($r=null,e=Gl(r),e=en(e),e!==null)if(t=bt(e),t===null)e=null;else if(n=t.tag,n===13){if(e=ao(t),e!==null)return e;e=null}else if(n===3){if(t.stateNode.current.memoizedState.isDehydrated)return t.tag===3?t.stateNode.containerInfo:null;e=null}else t!==e&&(e=null);return $r=e,null}function Po(e){switch(e){case"cancel":case"click":case"close":case"contextmenu":case"copy":case"cut":case"auxclick":case"dblclick":case"dragend":case"dragstart":case"drop":case"focusin":case"focusout":case"input":case"invalid":case"keydown":case"keypress":case"keyup":case"mousedown":case"mouseup":case"paste":case"pause":case"play":case"pointercancel":case"pointerdown":case"pointerup":case"ratechange":case"reset":case"resize":case"seeked":case"submit":case"touchcancel":case"touchend":case"touchstart":case"volumechange":case"change":case"selectionchange":case"textInput":case"compositionstart":case"compositionend":case"compositionupdate":case"beforeblur":case"afterblur":case"beforeinput":case"blur":case"fullscreenchange":case"focus":case"hashchange":case"popstate":case"select":case"selectstart":return 1;case"drag":case"dragenter":case"dragexit":case"dragleave":case"dragover":case"mousemove":case"mouseout":case"mouseover":case"pointermove":case"pointerout":case"pointerover":case"scroll":case"toggle":case"touchmove":case"wheel":case"mouseenter":case"mouseleave":case"pointerenter":case"pointerleave":return 4;case"message":switch(yc()){case eu:return 1;case vo:return 4;case Or:case gc:return 16;case yo:return 536870912;default:return 16}default:return 16}}var Ft=null,su=null,Br=null;function jo(){if(Br)return Br;var e,t=su,n=t.length,r,l="value"in Ft?Ft.value:Ft.textContent,u=l.length;for(e=0;e<n&&t[e]===l[e];e++);var i=n-e;for(r=1;r<=i&&t[n-r]===l[u-r];r++);return Br=l.slice(e,1<r?1-r:void 0)}function Wr(e){var t=e.keyCode;return"charCode"in e?(e=e.charCode,e===0&&t===13&&(e=13)):e=t,e===10&&(e=13),32<=e||e===13?e:0}function Hr(){return!0}function zo(){return!1}function Je(e){function t(n,r,l,u,i){this._reactName=n,this._targetInst=l,this.type=r,this.nativeEvent=u,this.target=i,this.currentTarget=null;for(var o in e)e.hasOwnProperty(o)&&(n=e[o],this[o]=n?n(u):u[o]);return this.isDefaultPrevented=(u.defaultPrevented!=null?u.defaultPrevented:u.returnValue===!1)?Hr:zo,this.isPropagationStopped=zo,this}return N(t.prototype,{preventDefault:function(){this.defaultPrevented=!0;var n=this.nativeEvent;n&&(n.preventDefault?n.preventDefault():typeof n.returnValue!="unknown"&&(n.returnValue=!1),this.isDefaultPrevented=Hr)},stopPropagation:function(){var n=this.nativeEvent;n&&(n.stopPropagation?n.stopPropagation():typeof n.cancelBubble!="unknown"&&(n.cancelBubble=!0),this.isPropagationStopped=Hr)},persist:function(){},isPersistent:Hr}),t}var gn={eventPhase:0,bubbles:0,cancelable:0,timeStamp:function(e){return e.timeStamp||Date.now()},defaultPrevented:0,isTrusted:0},au=Je(gn),bn=N({},gn,{view:0,detail:0}),Lc=Je(bn),cu,fu,er,Qr=N({},bn,{screenX:0,screenY:0,clientX:0,clientY:0,pageX:0,pageY:0,ctrlKey:0,shiftKey:0,altKey:0,metaKey:0,getModifierState:pu,button:0,buttons:0,relatedTarget:function(e){return e.relatedTarget===void 0?e.fromElement===e.srcElement?e.toElement:e.fromElement:e.relatedTarget},movementX:function(e){return"movementX"in e?e.movementX:(e!==er&&(er&&e.type==="mousemove"?(cu=e.screenX-er.screenX,fu=e.screenY-er.screenY):fu=cu=0,er=e),cu)},movementY:function(e){return"movementY"in e?e.movementY:fu}}),Ro=Je(Qr),Tc=N({},Qr,{dataTransfer:0}),Oc=Je(Tc),Ic=N({},bn,{relatedTarget:0}),du=Je(Ic),Dc=N({},gn,{animationName:0,elapsedTime:0,pseudoElement:0}),Mc=Je(Dc),Fc=N({},gn,{clipboardData:function(e){return"clipboardData"in e?e.clipboardData:window.clipboardData}}),Uc=Je(Fc),Ac=N({},gn,{data:0}),Lo=Je(Ac),Vc={Esc:"Escape",Spacebar:" ",Left:"ArrowLeft",Up:"ArrowUp",Right:"ArrowRight",Down:"ArrowDown",Del:"Delete",Win:"OS",Menu:"ContextMenu",Apps:"ContextMenu",Scroll:"ScrollLock",MozPrintableKey:"Unidentified"},$c={8:"Backspace",9:"Tab",12:"Clear",13:"Enter",16:"Shift",17:"Control",18:"Alt",19:"Pause",20:"CapsLock",27:"Escape",32:" ",33:"PageUp",34:"PageDown",35:"End",36:"Home",37:"ArrowLeft",38:"ArrowUp",39:"ArrowRight",40:"ArrowDown",45:"Insert",46:"Delete",112:"F1",113:"F2",114:"F3",115:"F4",116:"F5",117:"F6",118:"F7",119:"F8",120:"F9",121:"F10",122:"F11",123:"F12",144:"NumLock",145:"ScrollLock",224:"Meta"},Bc={Alt:"altKey",Control:"ctrlKey",Meta:"metaKey",Shift:"shiftKey"};function Wc(e){var t=this.nativeEvent;return t.getModifierState?t.getModifierState(e):(e=Bc[e])?!!t[e]:!1}function pu(){return Wc}var Hc=N({},bn,{key:function(e){if(e.key){var t=Vc[e.key]||e.key;if(t!=="Unidentified")return t}return e.type==="keypress"?(e=Wr(e),e===13?"Enter":String.fromCharCode(e)):e.type==="keydown"||e.type==="keyup"?$c[e.keyCode]||"Unidentified":""},code:0,location:0,ctrlKey:0,shiftKey:0,altKey:0,metaKey:0,repeat:0,locale:0,getModifierState:pu,charCode:function(e){return e.type==="keypress"?Wr(e):0},keyCode:function(e){return e.type==="keydown"||e.type==="keyup"?e.keyCode:0},which:function(e){return e.type==="keypress"?Wr(e):e.type==="keydown"||e.type==="keyup"?e.keyCode:0}}),Qc=Je(Hc),Kc=N({},Qr,{pointerId:0,width:0,height:0,pressure:0,tangentialPressure:0,tiltX:0,tiltY:0,twist:0,pointerType:0,isPrimary:0}),To=Je(Kc),Yc=N({},bn,{touches:0,targetTouches:0,changedTouches:0,altKey:0,metaKey:0,ctrlKey:0,shiftKey:0,getModifierState:pu}),Xc=Je(Yc),Gc=N({},gn,{propertyName:0,elapsedTime:0,pseudoElement:0}),Zc=Je(Gc),Jc=N({},Qr,{deltaX:function(e){return"deltaX"in e?e.deltaX:"wheelDeltaX"in e?-e.wheelDeltaX:0},deltaY:function(e){return"deltaY"in e?e.deltaY:"wheelDeltaY"in e?-e.wheelDeltaY:"wheelDelta"in e?-e.wheelDelta:0},deltaZ:0,deltaMode:0}),qc=Je(Jc),bc=[9,13,27,32],mu=ne&&"CompositionEvent"in window,tr=null;ne&&"documentMode"in document&&(tr=document.documentMode);var ef=ne&&"TextEvent"in window&&!tr,Oo=ne&&(!mu||tr&&8<tr&&11>=tr),Io=" ",Do=!1;function Mo(e,t){switch(e){case"keyup":return bc.indexOf(t.keyCode)!==-1;case"keydown":return t.keyCode!==229;case"keypress":case"mousedown":case"focusout":return!0;default:return!1}}function Fo(e){return e=e.detail,typeof e=="object"&&"data"in e?e.data:null}var wn=!1;function tf(e,t){switch(e){case"compositionend":return Fo(t);case"keypress":return t.which!==32?null:(Do=!0,Io);case"textInput":return e=t.data,e===Io&&Do?null:e;default:return null}}function nf(e,t){if(wn)return e==="compositionend"||!mu&&Mo(e,t)?(e=jo(),Br=su=Ft=null,wn=!1,e):null;switch(e){case"paste":return null;case"keypress":if(!(t.ctrlKey||t.altKey||t.metaKey)||t.ctrlKey&&t.altKey){if(t.char&&1<t.char.length)return t.char;if(t.which)return String.fromCharCode(t.which)}return null;case"compositionend":return Oo&&t.locale!=="ko"?null:t.data;default:return null}}var rf={color:!0,date:!0,datetime:!0,"datetime-local":!0,email:!0,month:!0,number:!0,password:!0,range:!0,search:!0,tel:!0,text:!0,time:!0,url:!0,week:!0};function Uo(e){var t=e&&e.nodeName&&e.nodeName.toLowerCase();return t==="input"?!!rf[e.type]:t==="textarea"}function Ao(e,t,n,r){lo(r),t=Zr(t,"onChange"),0<t.length&&(n=new au("onChange","change",null,n,r),e.push({event:n,listeners:t}))}var nr=null,rr=null;function lf(e){rs(e,0)}function Kr(e){var t=_n(e);if(Ki(t))return e}function uf(e,t){if(e==="change")return t}var Vo=!1;if(ne){var hu;if(ne){var vu="oninput"in document;if(!vu){var $o=document.createElement("div");$o.setAttribute("oninput","return;"),vu=typeof $o.oninput=="function"}hu=vu}else hu=!1;Vo=hu&&(!document.documentMode||9<document.documentMode)}function Bo(){nr&&(nr.detachEvent("onpropertychange",Wo),rr=nr=null)}function Wo(e){if(e.propertyName==="value"&&Kr(rr)){var t=[];Ao(t,rr,e,Gl(e)),so(lf,t)}}function of(e,t,n){e==="focusin"?(Bo(),nr=t,rr=n,nr.attachEvent("onpropertychange",Wo)):e==="focusout"&&Bo()}function sf(e){if(e==="selectionchange"||e==="keyup"||e==="keydown")return Kr(rr)}function af(e,t){if(e==="click")return Kr(t)}function cf(e,t){if(e==="input"||e==="change")return Kr(t)}function ff(e,t){return e===t&&(e!==0||1/e===1/t)||e!==e&&t!==t}var ct=typeof Object.is=="function"?Object.is:ff;function lr(e,t){if(ct(e,t))return!0;if(typeof e!="object"||e===null||typeof t!="object"||t===null)return!1;var n=Object.keys(e),r=Object.keys(t);if(n.length!==r.length)return!1;for(r=0;r<n.length;r++){var l=n[r];if(!$.call(t,l)||!ct(e[l],t[l]))return!1}return!0}function Ho(e){for(;e&&e.firstChild;)e=e.firstChild;return e}function Qo(e,t){var n=Ho(e);e=0;for(var r;n;){if(n.nodeType===3){if(r=e+n.textContent.length,e<=t&&r>=t)return{node:n,offset:t-e};e=r}e:{for(;n;){if(n.nextSibling){n=n.nextSibling;break e}n=n.parentNode}n=void 0}n=Ho(n)}}function Ko(e,t){return e&&t?e===t?!0:e&&e.nodeType===3?!1:t&&t.nodeType===3?Ko(e,t.parentNode):"contains"in e?e.contains(t):e.compareDocumentPosition?!!(e.compareDocumentPosition(t)&16):!1:!1}function Yo(){for(var e=window,t=zr();t instanceof e.HTMLIFrameElement;){try{var n=typeof t.contentWindow.location.href=="string"}catch{n=!1}if(n)e=t.contentWindow;else break;t=zr(e.document)}return t}function yu(e){var t=e&&e.nodeName&&e.nodeName.toLowerCase();return t&&(t==="input"&&(e.type==="text"||e.type==="search"||e.type==="tel"||e.type==="url"||e.type==="password")||t==="textarea"||e.contentEditable==="true")}function df(e){var t=Yo(),n=e.focusedElem,r=e.selectionRange;if(t!==n&&n&&n.ownerDocument&&Ko(n.ownerDocument.documentElement,n)){if(r!==null&&yu(n)){if(t=r.start,e=r.end,e===void 0&&(e=t),"selectionStart"in n)n.selectionStart=t,n.selectionEnd=Math.min(e,n.value.length);else if(e=(t=n.ownerDocument||document)&&t.defaultView||window,e.getSelection){e=e.getSelection();var l=n.textContent.length,u=Math.min(r.start,l);r=r.end===void 0?u:Math.min(r.end,l),!e.extend&&u>r&&(l=r,r=u,u=l),l=Qo(n,u);var i=Qo(n,r);l&&i&&(e.rangeCount!==1||e.anchorNode!==l.node||e.anchorOffset!==l.offset||e.focusNode!==i.node||e.focusOffset!==i.offset)&&(t=t.createRange(),t.setStart(l.node,l.offset),e.removeAllRanges(),u>r?(e.addRange(t),e.extend(i.node,i.offset)):(t.setEnd(i.node,i.offset),e.addRange(t)))}}for(t=[],e=n;e=e.parentNode;)e.nodeType===1&&t.push({element:e,left:e.scrollLeft,top:e.scrollTop});for(typeof n.focus=="function"&&n.focus(),n=0;n<t.length;n++)e=t[n],e.element.scrollLeft=e.left,e.element.scrollTop=e.top}}var pf=ne&&"documentMode"in document&&11>=document.documentMode,Sn=null,gu=null,ur=null,wu=!1;function Xo(e,t,n){var r=n.window===n?n.document:n.nodeType===9?n:n.ownerDocument;wu||Sn==null||Sn!==zr(r)||(r=Sn,"selectionStart"in r&&yu(r)?r={start:r.selectionStart,end:r.selectionEnd}:(r=(r.ownerDocument&&r.ownerDocument.defaultView||window).getSelection(),r={anchorNode:r.anchorNode,anchorOffset:r.anchorOffset,focusNode:r.focusNode,focusOffset:r.focusOffset}),ur&&lr(ur,r)||(ur=r,r=Zr(gu,"onSelect"),0<r.length&&(t=new au("onSelect","select",null,t,n),e.push({event:t,listeners:r}),t.target=Sn)))}function Yr(e,t){var n={};return n[e.toLowerCase()]=t.toLowerCase(),n["Webkit"+e]="webkit"+t,n["Moz"+e]="moz"+t,n}var xn={animationend:Yr("Animation","AnimationEnd"),animationiteration:Yr("Animation","AnimationIteration"),animationstart:Yr("Animation","AnimationStart"),transitionend:Yr("Transition","TransitionEnd")},Su={},Go={};ne&&(Go=document.createElement("div").style,"AnimationEvent"in window||(delete xn.animationend.animation,delete xn.animationiteration.animation,delete xn.animationstart.animation),"TransitionEvent"in window||delete xn.transitionend.transition);function Xr(e){if(Su[e])return Su[e];if(!xn[e])return e;var t=xn[e],n;for(n in t)if(t.hasOwnProperty(n)&&n in Go)return Su[e]=t[n];return e}var Zo=Xr("animationend"),Jo=Xr("animationiteration"),qo=Xr("animationstart"),bo=Xr("transitionend"),es=new Map,ts="abort auxClick cancel canPlay canPlayThrough click close contextMenu copy cut drag dragEnd dragEnter dragExit dragLeave dragOver dragStart drop durationChange emptied encrypted ended error gotPointerCapture input invalid keyDown keyPress keyUp load loadedData loadedMetadata loadStart lostPointerCapture mouseDown mouseMove mouseOut mouseOver mouseUp paste pause play playing pointerCancel pointerDown pointerMove pointerOut pointerOver pointerUp progress rateChange reset resize seeked seeking stalled submit suspend timeUpdate touchCancel touchEnd touchStart volumeChange scroll toggle touchMove waiting wheel".split(" ");function Ut(e,t){es.set(e,t),V(t,[e])}for(var xu=0;xu<ts.length;xu++){var ku=ts[xu],mf=ku.toLowerCase(),hf=ku[0].toUpperCase()+ku.slice(1);Ut(mf,"on"+hf)}Ut(Zo,"onAnimationEnd"),Ut(Jo,"onAnimationIteration"),Ut(qo,"onAnimationStart"),Ut("dblclick","onDoubleClick"),Ut("focusin","onFocus"),Ut("focusout","onBlur"),Ut(bo,"onTransitionEnd"),q("onMouseEnter",["mouseout","mouseover"]),q("onMouseLeave",["mouseout","mouseover"]),q("onPointerEnter",["pointerout","pointerover"]),q("onPointerLeave",["pointerout","pointerover"]),V("onChange","change click focusin focusout input keydown keyup selectionchange".split(" ")),V("onSelect","focusout contextmenu dragend focusin keydown keyup mousedown mouseup selectionchange".split(" ")),V("onBeforeInput",["compositionend","keypress","textInput","paste"]),V("onCompositionEnd","compositionend focusout keydown keypress keyup mousedown".split(" ")),V("onCompositionStart","compositionstart focusout keydown keypress keyup mousedown".split(" ")),V("onCompositionUpdate","compositionupdate focusout keydown keypress keyup mousedown".split(" "));var ir="abort canplay canplaythrough durationchange emptied encrypted ended error loadeddata loadedmetadata loadstart pause play playing progress ratechange resize seeked seeking stalled suspend timeupdate volumechange waiting".split(" "),vf=new Set("cancel close invalid load scroll toggle".split(" ").concat(ir));function ns(e,t,n){var r=e.type||"unknown-event";e.currentTarget=n,pc(r,t,void 0,e),e.currentTarget=null}function rs(e,t){t=(t&4)!==0;for(var n=0;n<e.length;n++){var r=e[n],l=r.event;r=r.listeners;e:{var u=void 0;if(t)for(var i=r.length-1;0<=i;i--){var o=r[i],s=o.instance,h=o.currentTarget;if(o=o.listener,s!==u&&l.isPropagationStopped())break e;ns(l,o,h),u=s}else for(i=0;i<r.length;i++){if(o=r[i],s=o.instance,h=o.currentTarget,o=o.listener,s!==u&&l.isPropagationStopped())break e;ns(l,o,h),u=s}}}if(Tr)throw e=bl,Tr=!1,bl=null,e}function fe(e,t){var n=t[Ru];n===void 0&&(n=t[Ru]=new Set);var r=e+"__bubble";n.has(r)||(ls(t,e,2,!1),n.add(r))}function Eu(e,t,n){var r=0;t&&(r|=4),ls(n,e,r,t)}var Gr="_reactListening"+Math.random().toString(36).slice(2);function or(e){if(!e[Gr]){e[Gr]=!0,P.forEach(function(n){n!=="selectionchange"&&(vf.has(n)||Eu(n,!1,e),Eu(n,!0,e))});var t=e.nodeType===9?e:e.ownerDocument;t===null||t[Gr]||(t[Gr]=!0,Eu("selectionchange",!1,t))}}function ls(e,t,n,r){switch(Po(t)){case 1:var l=zc;break;case 4:l=Rc;break;default:l=iu}n=l.bind(null,t,n,e),l=void 0,!ql||t!=="touchstart"&&t!=="touchmove"&&t!=="wheel"||(l=!0),r?l!==void 0?e.addEventListener(t,n,{capture:!0,passive:l}):e.addEventListener(t,n,!0):l!==void 0?e.addEventListener(t,n,{passive:l}):e.addEventListener(t,n,!1)}function _u(e,t,n,r,l){var u=r;if((t&1)===0&&(t&2)===0&&r!==null)e:for(;;){if(r===null)return;var i=r.tag;if(i===3||i===4){var o=r.stateNode.containerInfo;if(o===l||o.nodeType===8&&o.parentNode===l)break;if(i===4)for(i=r.return;i!==null;){var s=i.tag;if((s===3||s===4)&&(s=i.stateNode.containerInfo,s===l||s.nodeType===8&&s.parentNode===l))return;i=i.return}for(;o!==null;){if(i=en(o),i===null)return;if(s=i.tag,s===5||s===6){r=u=i;continue e}o=o.parentNode}}r=r.return}so(function(){var h=u,S=Gl(n),x=[];e:{var g=es.get(e);if(g!==void 0){var _=au,j=e;switch(e){case"keypress":if(Wr(n)===0)break e;case"keydown":case"keyup":_=Qc;break;case"focusin":j="focus",_=du;break;case"focusout":j="blur",_=du;break;case"beforeblur":case"afterblur":_=du;break;case"click":if(n.button===2)break e;case"auxclick":case"dblclick":case"mousedown":case"mousemove":case"mouseup":case"mouseout":case"mouseover":case"contextmenu":_=Ro;break;case"drag":case"dragend":case"dragenter":case"dragexit":case"dragleave":case"dragover":case"dragstart":case"drop":_=Oc;break;case"touchcancel":case"touchend":case"touchmove":case"touchstart":_=Xc;break;case Zo:case Jo:case qo:_=Mc;break;case bo:_=Zc;break;case"scroll":_=Lc;break;case"wheel":_=qc;break;case"copy":case"cut":case"paste":_=Uc;break;case"gotpointercapture":case"lostpointercapture":case"pointercancel":case"pointerdown":case"pointermove":case"pointerout":case"pointerover":case"pointerup":_=To}var z=(t&4)!==0,xe=!z&&e==="scroll",d=z?g!==null?g+"Capture":null:g;z=[];for(var a=h,p;a!==null;){p=a;var k=p.stateNode;if(p.tag===5&&k!==null&&(p=k,d!==null&&(k=Wn(a,d),k!=null&&z.push(sr(a,k,p)))),xe)break;a=a.return}0<z.length&&(g=new _(g,j,null,n,S),x.push({event:g,listeners:z}))}}if((t&7)===0){e:{if(g=e==="mouseover"||e==="pointerover",_=e==="mouseout"||e==="pointerout",g&&n!==Xl&&(j=n.relatedTarget||n.fromElement)&&(en(j)||j[Et]))break e;if((_||g)&&(g=S.window===S?S:(g=S.ownerDocument)?g.defaultView||g.parentWindow:window,_?(j=n.relatedTarget||n.toElement,_=h,j=j?en(j):null,j!==null&&(xe=bt(j),j!==xe||j.tag!==5&&j.tag!==6)&&(j=null)):(_=null,j=h),_!==j)){if(z=Ro,k="onMouseLeave",d="onMouseEnter",a="mouse",(e==="pointerout"||e==="pointerover")&&(z=To,k="onPointerLeave",d="onPointerEnter",a="pointer"),xe=_==null?g:_n(_),p=j==null?g:_n(j),g=new z(k,a+"leave",_,n,S),g.target=xe,g.relatedTarget=p,k=null,en(S)===h&&(z=new z(d,a+"enter",j,n,S),z.target=p,z.relatedTarget=xe,k=z),xe=k,_&&j)t:{for(z=_,d=j,a=0,p=z;p;p=kn(p))a++;for(p=0,k=d;k;k=kn(k))p++;for(;0<a-p;)z=kn(z),a--;for(;0<p-a;)d=kn(d),p--;for(;a--;){if(z===d||d!==null&&z===d.alternate)break t;z=kn(z),d=kn(d)}z=null}else z=null;_!==null&&us(x,g,_,z,!1),j!==null&&xe!==null&&us(x,xe,j,z,!0)}}e:{if(g=h?_n(h):window,_=g.nodeName&&g.nodeName.toLowerCase(),_==="select"||_==="input"&&g.type==="file")var L=uf;else if(Uo(g))if(Vo)L=cf;else{L=sf;var I=of}else(_=g.nodeName)&&_.toLowerCase()==="input"&&(g.type==="checkbox"||g.type==="radio")&&(L=af);if(L&&(L=L(e,h))){Ao(x,L,n,S);break e}I&&I(e,g,h),e==="focusout"&&(I=g._wrapperState)&&I.controlled&&g.type==="number"&&Wl(g,"number",g.value)}switch(I=h?_n(h):window,e){case"focusin":(Uo(I)||I.contentEditable==="true")&&(Sn=I,gu=h,ur=null);break;case"focusout":ur=gu=Sn=null;break;case"mousedown":wu=!0;break;case"contextmenu":case"mouseup":case"dragend":wu=!1,Xo(x,n,S);break;case"selectionchange":if(pf)break;case"keydown":case"keyup":Xo(x,n,S)}var D;if(mu)e:{switch(e){case"compositionstart":var U="onCompositionStart";break e;case"compositionend":U="onCompositionEnd";break e;case"compositionupdate":U="onCompositionUpdate";break e}U=void 0}else wn?Mo(e,n)&&(U="onCompositionEnd"):e==="keydown"&&n.keyCode===229&&(U="onCompositionStart");U&&(Oo&&n.locale!=="ko"&&(wn||U!=="onCompositionStart"?U==="onCompositionEnd"&&wn&&(D=jo()):(Ft=S,su="value"in Ft?Ft.value:Ft.textContent,wn=!0)),I=Zr(h,U),0<I.length&&(U=new Lo(U,e,null,n,S),x.push({event:U,listeners:I}),D?U.data=D:(D=Fo(n),D!==null&&(U.data=D)))),(D=ef?tf(e,n):nf(e,n))&&(h=Zr(h,"onBeforeInput"),0<h.length&&(S=new Lo("onBeforeInput","beforeinput",null,n,S),x.push({event:S,listeners:h}),S.data=D))}rs(x,t)})}function sr(e,t,n){return{instance:e,listener:t,currentTarget:n}}function Zr(e,t){for(var n=t+"Capture",r=[];e!==null;){var l=e,u=l.stateNode;l.tag===5&&u!==null&&(l=u,u=Wn(e,n),u!=null&&r.unshift(sr(e,u,l)),u=Wn(e,t),u!=null&&r.push(sr(e,u,l))),e=e.return}return r}function kn(e){if(e===null)return null;do e=e.return;while(e&&e.tag!==5);return e||null}function us(e,t,n,r,l){for(var u=t._reactName,i=[];n!==null&&n!==r;){var o=n,s=o.alternate,h=o.stateNode;if(s!==null&&s===r)break;o.tag===5&&h!==null&&(o=h,l?(s=Wn(n,u),s!=null&&i.unshift(sr(n,s,o))):l||(s=Wn(n,u),s!=null&&i.push(sr(n,s,o)))),n=n.return}i.length!==0&&e.push({event:t,listeners:i})}var yf=/\r\n?/g,gf=/\u0000|\uFFFD/g;function is(e){return(typeof e=="string"?e:""+e).replace(yf,`
+`).replace(gf,"")}function Jr(e,t,n){if(t=is(t),is(e)!==t&&n)throw Error(c(425))}function qr(){}var Cu=null,Nu=null;function Pu(e,t){return e==="textarea"||e==="noscript"||typeof t.children=="string"||typeof t.children=="number"||typeof t.dangerouslySetInnerHTML=="object"&&t.dangerouslySetInnerHTML!==null&&t.dangerouslySetInnerHTML.__html!=null}var ju=typeof setTimeout=="function"?setTimeout:void 0,wf=typeof clearTimeout=="function"?clearTimeout:void 0,os=typeof Promise=="function"?Promise:void 0,Sf=typeof queueMicrotask=="function"?queueMicrotask:typeof os<"u"?function(e){return os.resolve(null).then(e).catch(xf)}:ju;function xf(e){setTimeout(function(){throw e})}function zu(e,t){var n=t,r=0;do{var l=n.nextSibling;if(e.removeChild(n),l&&l.nodeType===8)if(n=l.data,n==="/$"){if(r===0){e.removeChild(l),qn(t);return}r--}else n!=="$"&&n!=="$?"&&n!=="$!"||r++;n=l}while(n);qn(t)}function At(e){for(;e!=null;e=e.nextSibling){var t=e.nodeType;if(t===1||t===3)break;if(t===8){if(t=e.data,t==="$"||t==="$!"||t==="$?")break;if(t==="/$")return null}}return e}function ss(e){e=e.previousSibling;for(var t=0;e;){if(e.nodeType===8){var n=e.data;if(n==="$"||n==="$!"||n==="$?"){if(t===0)return e;t--}else n==="/$"&&t++}e=e.previousSibling}return null}var En=Math.random().toString(36).slice(2),wt="__reactFiber$"+En,ar="__reactProps$"+En,Et="__reactContainer$"+En,Ru="__reactEvents$"+En,kf="__reactListeners$"+En,Ef="__reactHandles$"+En;function en(e){var t=e[wt];if(t)return t;for(var n=e.parentNode;n;){if(t=n[Et]||n[wt]){if(n=t.alternate,t.child!==null||n!==null&&n.child!==null)for(e=ss(e);e!==null;){if(n=e[wt])return n;e=ss(e)}return t}e=n,n=e.parentNode}return null}function cr(e){return e=e[wt]||e[Et],!e||e.tag!==5&&e.tag!==6&&e.tag!==13&&e.tag!==3?null:e}function _n(e){if(e.tag===5||e.tag===6)return e.stateNode;throw Error(c(33))}function br(e){return e[ar]||null}var Lu=[],Cn=-1;function Vt(e){return{current:e}}function de(e){0>Cn||(e.current=Lu[Cn],Lu[Cn]=null,Cn--)}function ce(e,t){Cn++,Lu[Cn]=e.current,e.current=t}var $t={},Ie=Vt($t),He=Vt(!1),tn=$t;function Nn(e,t){var n=e.type.contextTypes;if(!n)return $t;var r=e.stateNode;if(r&&r.__reactInternalMemoizedUnmaskedChildContext===t)return r.__reactInternalMemoizedMaskedChildContext;var l={},u;for(u in n)l[u]=t[u];return r&&(e=e.stateNode,e.__reactInternalMemoizedUnmaskedChildContext=t,e.__reactInternalMemoizedMaskedChildContext=l),l}function Qe(e){return e=e.childContextTypes,e!=null}function el(){de(He),de(Ie)}function as(e,t,n){if(Ie.current!==$t)throw Error(c(168));ce(Ie,t),ce(He,n)}function cs(e,t,n){var r=e.stateNode;if(t=t.childContextTypes,typeof r.getChildContext!="function")return n;r=r.getChildContext();for(var l in r)if(!(l in t))throw Error(c(108,ae(e)||"Unknown",l));return N({},n,r)}function tl(e){return e=(e=e.stateNode)&&e.__reactInternalMemoizedMergedChildContext||$t,tn=Ie.current,ce(Ie,e),ce(He,He.current),!0}function fs(e,t,n){var r=e.stateNode;if(!r)throw Error(c(169));n?(e=cs(e,t,tn),r.__reactInternalMemoizedMergedChildContext=e,de(He),de(Ie),ce(Ie,e)):de(He),ce(He,n)}var _t=null,nl=!1,Tu=!1;function ds(e){_t===null?_t=[e]:_t.push(e)}function _f(e){nl=!0,ds(e)}function Bt(){if(!Tu&&_t!==null){Tu=!0;var e=0,t=ue;try{var n=_t;for(ue=1;e<n.length;e++){var r=n[e];do r=r(!0);while(r!==null)}_t=null,nl=!1}catch(l){throw _t!==null&&(_t=_t.slice(e+1)),mo(eu,Bt),l}finally{ue=t,Tu=!1}}return null}var Pn=[],jn=0,rl=null,ll=0,tt=[],nt=0,nn=null,Ct=1,Nt="";function rn(e,t){Pn[jn++]=ll,Pn[jn++]=rl,rl=e,ll=t}function ps(e,t,n){tt[nt++]=Ct,tt[nt++]=Nt,tt[nt++]=nn,nn=e;var r=Ct;e=Nt;var l=32-at(r)-1;r&=~(1<<l),n+=1;var u=32-at(t)+l;if(30<u){var i=l-l%5;u=(r&(1<<i)-1).toString(32),r>>=i,l-=i,Ct=1<<32-at(t)+l|n<<l|r,Nt=u+e}else Ct=1<<u|n<<l|r,Nt=e}function Ou(e){e.return!==null&&(rn(e,1),ps(e,1,0))}function Iu(e){for(;e===rl;)rl=Pn[--jn],Pn[jn]=null,ll=Pn[--jn],Pn[jn]=null;for(;e===nn;)nn=tt[--nt],tt[nt]=null,Nt=tt[--nt],tt[nt]=null,Ct=tt[--nt],tt[nt]=null}var qe=null,be=null,me=!1,ft=null;function ms(e,t){var n=it(5,null,null,0);n.elementType="DELETED",n.stateNode=t,n.return=e,t=e.deletions,t===null?(e.deletions=[n],e.flags|=16):t.push(n)}function hs(e,t){switch(e.tag){case 5:var n=e.type;return t=t.nodeType!==1||n.toLowerCase()!==t.nodeName.toLowerCase()?null:t,t!==null?(e.stateNode=t,qe=e,be=At(t.firstChild),!0):!1;case 6:return t=e.pendingProps===""||t.nodeType!==3?null:t,t!==null?(e.stateNode=t,qe=e,be=null,!0):!1;case 13:return t=t.nodeType!==8?null:t,t!==null?(n=nn!==null?{id:Ct,overflow:Nt}:null,e.memoizedState={dehydrated:t,treeContext:n,retryLane:1073741824},n=it(18,null,null,0),n.stateNode=t,n.return=e,e.child=n,qe=e,be=null,!0):!1;default:return!1}}function Du(e){return(e.mode&1)!==0&&(e.flags&128)===0}function Mu(e){if(me){var t=be;if(t){var n=t;if(!hs(e,t)){if(Du(e))throw Error(c(418));t=At(n.nextSibling);var r=qe;t&&hs(e,t)?ms(r,n):(e.flags=e.flags&-4097|2,me=!1,qe=e)}}else{if(Du(e))throw Error(c(418));e.flags=e.flags&-4097|2,me=!1,qe=e}}}function vs(e){for(e=e.return;e!==null&&e.tag!==5&&e.tag!==3&&e.tag!==13;)e=e.return;qe=e}function ul(e){if(e!==qe)return!1;if(!me)return vs(e),me=!0,!1;var t;if((t=e.tag!==3)&&!(t=e.tag!==5)&&(t=e.type,t=t!=="head"&&t!=="body"&&!Pu(e.type,e.memoizedProps)),t&&(t=be)){if(Du(e))throw ys(),Error(c(418));for(;t;)ms(e,t),t=At(t.nextSibling)}if(vs(e),e.tag===13){if(e=e.memoizedState,e=e!==null?e.dehydrated:null,!e)throw Error(c(317));e:{for(e=e.nextSibling,t=0;e;){if(e.nodeType===8){var n=e.data;if(n==="/$"){if(t===0){be=At(e.nextSibling);break e}t--}else n!=="$"&&n!=="$!"&&n!=="$?"||t++}e=e.nextSibling}be=null}}else be=qe?At(e.stateNode.nextSibling):null;return!0}function ys(){for(var e=be;e;)e=At(e.nextSibling)}function zn(){be=qe=null,me=!1}function Fu(e){ft===null?ft=[e]:ft.push(e)}var Cf=ie.ReactCurrentBatchConfig;function fr(e,t,n){if(e=n.ref,e!==null&&typeof e!="function"&&typeof e!="object"){if(n._owner){if(n=n._owner,n){if(n.tag!==1)throw Error(c(309));var r=n.stateNode}if(!r)throw Error(c(147,e));var l=r,u=""+e;return t!==null&&t.ref!==null&&typeof t.ref=="function"&&t.ref._stringRef===u?t.ref:(t=function(i){var o=l.refs;i===null?delete o[u]:o[u]=i},t._stringRef=u,t)}if(typeof e!="string")throw Error(c(284));if(!n._owner)throw Error(c(290,e))}return e}function il(e,t){throw e=Object.prototype.toString.call(t),Error(c(31,e==="[object Object]"?"object with keys {"+Object.keys(t).join(", ")+"}":e))}function gs(e){var t=e._init;return t(e._payload)}function ws(e){function t(d,a){if(e){var p=d.deletions;p===null?(d.deletions=[a],d.flags|=16):p.push(a)}}function n(d,a){if(!e)return null;for(;a!==null;)t(d,a),a=a.sibling;return null}function r(d,a){for(d=new Map;a!==null;)a.key!==null?d.set(a.key,a):d.set(a.index,a),a=a.sibling;return d}function l(d,a){return d=Zt(d,a),d.index=0,d.sibling=null,d}function u(d,a,p){return d.index=p,e?(p=d.alternate,p!==null?(p=p.index,p<a?(d.flags|=2,a):p):(d.flags|=2,a)):(d.flags|=1048576,a)}function i(d){return e&&d.alternate===null&&(d.flags|=2),d}function o(d,a,p,k){return a===null||a.tag!==6?(a=ji(p,d.mode,k),a.return=d,a):(a=l(a,p),a.return=d,a)}function s(d,a,p,k){var L=p.type;return L===se?S(d,a,p.props.children,k,p.key):a!==null&&(a.elementType===L||typeof L=="object"&&L!==null&&L.$$typeof===We&&gs(L)===a.type)?(k=l(a,p.props),k.ref=fr(d,a,p),k.return=d,k):(k=Rl(p.type,p.key,p.props,null,d.mode,k),k.ref=fr(d,a,p),k.return=d,k)}function h(d,a,p,k){return a===null||a.tag!==4||a.stateNode.containerInfo!==p.containerInfo||a.stateNode.implementation!==p.implementation?(a=zi(p,d.mode,k),a.return=d,a):(a=l(a,p.children||[]),a.return=d,a)}function S(d,a,p,k,L){return a===null||a.tag!==7?(a=dn(p,d.mode,k,L),a.return=d,a):(a=l(a,p),a.return=d,a)}function x(d,a,p){if(typeof a=="string"&&a!==""||typeof a=="number")return a=ji(""+a,d.mode,p),a.return=d,a;if(typeof a=="object"&&a!==null){switch(a.$$typeof){case Oe:return p=Rl(a.type,a.key,a.props,null,d.mode,p),p.ref=fr(d,null,a),p.return=d,p;case re:return a=zi(a,d.mode,p),a.return=d,a;case We:var k=a._init;return x(d,k(a._payload),p)}if(Vn(a)||F(a))return a=dn(a,d.mode,p,null),a.return=d,a;il(d,a)}return null}function g(d,a,p,k){var L=a!==null?a.key:null;if(typeof p=="string"&&p!==""||typeof p=="number")return L!==null?null:o(d,a,""+p,k);if(typeof p=="object"&&p!==null){switch(p.$$typeof){case Oe:return p.key===L?s(d,a,p,k):null;case re:return p.key===L?h(d,a,p,k):null;case We:return L=p._init,g(d,a,L(p._payload),k)}if(Vn(p)||F(p))return L!==null?null:S(d,a,p,k,null);il(d,p)}return null}function _(d,a,p,k,L){if(typeof k=="string"&&k!==""||typeof k=="number")return d=d.get(p)||null,o(a,d,""+k,L);if(typeof k=="object"&&k!==null){switch(k.$$typeof){case Oe:return d=d.get(k.key===null?p:k.key)||null,s(a,d,k,L);case re:return d=d.get(k.key===null?p:k.key)||null,h(a,d,k,L);case We:var I=k._init;return _(d,a,p,I(k._payload),L)}if(Vn(k)||F(k))return d=d.get(p)||null,S(a,d,k,L,null);il(a,k)}return null}function j(d,a,p,k){for(var L=null,I=null,D=a,U=a=0,je=null;D!==null&&U<p.length;U++){D.index>U?(je=D,D=null):je=D.sibling;var te=g(d,D,p[U],k);if(te===null){D===null&&(D=je);break}e&&D&&te.alternate===null&&t(d,D),a=u(te,a,U),I===null?L=te:I.sibling=te,I=te,D=je}if(U===p.length)return n(d,D),me&&rn(d,U),L;if(D===null){for(;U<p.length;U++)D=x(d,p[U],k),D!==null&&(a=u(D,a,U),I===null?L=D:I.sibling=D,I=D);return me&&rn(d,U),L}for(D=r(d,D);U<p.length;U++)je=_(D,d,U,p[U],k),je!==null&&(e&&je.alternate!==null&&D.delete(je.key===null?U:je.key),a=u(je,a,U),I===null?L=je:I.sibling=je,I=je);return e&&D.forEach(function(Jt){return t(d,Jt)}),me&&rn(d,U),L}function z(d,a,p,k){var L=F(p);if(typeof L!="function")throw Error(c(150));if(p=L.call(p),p==null)throw Error(c(151));for(var I=L=null,D=a,U=a=0,je=null,te=p.next();D!==null&&!te.done;U++,te=p.next()){D.index>U?(je=D,D=null):je=D.sibling;var Jt=g(d,D,te.value,k);if(Jt===null){D===null&&(D=je);break}e&&D&&Jt.alternate===null&&t(d,D),a=u(Jt,a,U),I===null?L=Jt:I.sibling=Jt,I=Jt,D=je}if(te.done)return n(d,D),me&&rn(d,U),L;if(D===null){for(;!te.done;U++,te=p.next())te=x(d,te.value,k),te!==null&&(a=u(te,a,U),I===null?L=te:I.sibling=te,I=te);return me&&rn(d,U),L}for(D=r(d,D);!te.done;U++,te=p.next())te=_(D,d,U,te.value,k),te!==null&&(e&&te.alternate!==null&&D.delete(te.key===null?U:te.key),a=u(te,a,U),I===null?L=te:I.sibling=te,I=te);return e&&D.forEach(function(ld){return t(d,ld)}),me&&rn(d,U),L}function xe(d,a,p,k){if(typeof p=="object"&&p!==null&&p.type===se&&p.key===null&&(p=p.props.children),typeof p=="object"&&p!==null){switch(p.$$typeof){case Oe:e:{for(var L=p.key,I=a;I!==null;){if(I.key===L){if(L=p.type,L===se){if(I.tag===7){n(d,I.sibling),a=l(I,p.props.children),a.return=d,d=a;break e}}else if(I.elementType===L||typeof L=="object"&&L!==null&&L.$$typeof===We&&gs(L)===I.type){n(d,I.sibling),a=l(I,p.props),a.ref=fr(d,I,p),a.return=d,d=a;break e}n(d,I);break}else t(d,I);I=I.sibling}p.type===se?(a=dn(p.props.children,d.mode,k,p.key),a.return=d,d=a):(k=Rl(p.type,p.key,p.props,null,d.mode,k),k.ref=fr(d,a,p),k.return=d,d=k)}return i(d);case re:e:{for(I=p.key;a!==null;){if(a.key===I)if(a.tag===4&&a.stateNode.containerInfo===p.containerInfo&&a.stateNode.implementation===p.implementation){n(d,a.sibling),a=l(a,p.children||[]),a.return=d,d=a;break e}else{n(d,a);break}else t(d,a);a=a.sibling}a=zi(p,d.mode,k),a.return=d,d=a}return i(d);case We:return I=p._init,xe(d,a,I(p._payload),k)}if(Vn(p))return j(d,a,p,k);if(F(p))return z(d,a,p,k);il(d,p)}return typeof p=="string"&&p!==""||typeof p=="number"?(p=""+p,a!==null&&a.tag===6?(n(d,a.sibling),a=l(a,p),a.return=d,d=a):(n(d,a),a=ji(p,d.mode,k),a.return=d,d=a),i(d)):n(d,a)}return xe}var Rn=ws(!0),Ss=ws(!1),ol=Vt(null),sl=null,Ln=null,Uu=null;function Au(){Uu=Ln=sl=null}function Vu(e){var t=ol.current;de(ol),e._currentValue=t}function $u(e,t,n){for(;e!==null;){var r=e.alternate;if((e.childLanes&t)!==t?(e.childLanes|=t,r!==null&&(r.childLanes|=t)):r!==null&&(r.childLanes&t)!==t&&(r.childLanes|=t),e===n)break;e=e.return}}function Tn(e,t){sl=e,Uu=Ln=null,e=e.dependencies,e!==null&&e.firstContext!==null&&((e.lanes&t)!==0&&(Ke=!0),e.firstContext=null)}function rt(e){var t=e._currentValue;if(Uu!==e)if(e={context:e,memoizedValue:t,next:null},Ln===null){if(sl===null)throw Error(c(308));Ln=e,sl.dependencies={lanes:0,firstContext:e}}else Ln=Ln.next=e;return t}var ln=null;function Bu(e){ln===null?ln=[e]:ln.push(e)}function xs(e,t,n,r){var l=t.interleaved;return l===null?(n.next=n,Bu(t)):(n.next=l.next,l.next=n),t.interleaved=n,Pt(e,r)}function Pt(e,t){e.lanes|=t;var n=e.alternate;for(n!==null&&(n.lanes|=t),n=e,e=e.return;e!==null;)e.childLanes|=t,n=e.alternate,n!==null&&(n.childLanes|=t),n=e,e=e.return;return n.tag===3?n.stateNode:null}var Wt=!1;function Wu(e){e.updateQueue={baseState:e.memoizedState,firstBaseUpdate:null,lastBaseUpdate:null,shared:{pending:null,interleaved:null,lanes:0},effects:null}}function ks(e,t){e=e.updateQueue,t.updateQueue===e&&(t.updateQueue={baseState:e.baseState,firstBaseUpdate:e.firstBaseUpdate,lastBaseUpdate:e.lastBaseUpdate,shared:e.shared,effects:e.effects})}function jt(e,t){return{eventTime:e,lane:t,tag:0,payload:null,callback:null,next:null}}function Ht(e,t,n){var r=e.updateQueue;if(r===null)return null;if(r=r.shared,(J&2)!==0){var l=r.pending;return l===null?t.next=t:(t.next=l.next,l.next=t),r.pending=t,Pt(e,n)}return l=r.interleaved,l===null?(t.next=t,Bu(r)):(t.next=l.next,l.next=t),r.interleaved=t,Pt(e,n)}function al(e,t,n){if(t=t.updateQueue,t!==null&&(t=t.shared,(n&4194240)!==0)){var r=t.lanes;r&=e.pendingLanes,n|=r,t.lanes=n,ru(e,n)}}function Es(e,t){var n=e.updateQueue,r=e.alternate;if(r!==null&&(r=r.updateQueue,n===r)){var l=null,u=null;if(n=n.firstBaseUpdate,n!==null){do{var i={eventTime:n.eventTime,lane:n.lane,tag:n.tag,payload:n.payload,callback:n.callback,next:null};u===null?l=u=i:u=u.next=i,n=n.next}while(n!==null);u===null?l=u=t:u=u.next=t}else l=u=t;n={baseState:r.baseState,firstBaseUpdate:l,lastBaseUpdate:u,shared:r.shared,effects:r.effects},e.updateQueue=n;return}e=n.lastBaseUpdate,e===null?n.firstBaseUpdate=t:e.next=t,n.lastBaseUpdate=t}function cl(e,t,n,r){var l=e.updateQueue;Wt=!1;var u=l.firstBaseUpdate,i=l.lastBaseUpdate,o=l.shared.pending;if(o!==null){l.shared.pending=null;var s=o,h=s.next;s.next=null,i===null?u=h:i.next=h,i=s;var S=e.alternate;S!==null&&(S=S.updateQueue,o=S.lastBaseUpdate,o!==i&&(o===null?S.firstBaseUpdate=h:o.next=h,S.lastBaseUpdate=s))}if(u!==null){var x=l.baseState;i=0,S=h=s=null,o=u;do{var g=o.lane,_=o.eventTime;if((r&g)===g){S!==null&&(S=S.next={eventTime:_,lane:0,tag:o.tag,payload:o.payload,callback:o.callback,next:null});e:{var j=e,z=o;switch(g=t,_=n,z.tag){case 1:if(j=z.payload,typeof j=="function"){x=j.call(_,x,g);break e}x=j;break e;case 3:j.flags=j.flags&-65537|128;case 0:if(j=z.payload,g=typeof j=="function"?j.call(_,x,g):j,g==null)break e;x=N({},x,g);break e;case 2:Wt=!0}}o.callback!==null&&o.lane!==0&&(e.flags|=64,g=l.effects,g===null?l.effects=[o]:g.push(o))}else _={eventTime:_,lane:g,tag:o.tag,payload:o.payload,callback:o.callback,next:null},S===null?(h=S=_,s=x):S=S.next=_,i|=g;if(o=o.next,o===null){if(o=l.shared.pending,o===null)break;g=o,o=g.next,g.next=null,l.lastBaseUpdate=g,l.shared.pending=null}}while(!0);if(S===null&&(s=x),l.baseState=s,l.firstBaseUpdate=h,l.lastBaseUpdate=S,t=l.shared.interleaved,t!==null){l=t;do i|=l.lane,l=l.next;while(l!==t)}else u===null&&(l.shared.lanes=0);sn|=i,e.lanes=i,e.memoizedState=x}}function _s(e,t,n){if(e=t.effects,t.effects=null,e!==null)for(t=0;t<e.length;t++){var r=e[t],l=r.callback;if(l!==null){if(r.callback=null,r=n,typeof l!="function")throw Error(c(191,l));l.call(r)}}}var dr={},St=Vt(dr),pr=Vt(dr),mr=Vt(dr);function un(e){if(e===dr)throw Error(c(174));return e}function Hu(e,t){switch(ce(mr,t),ce(pr,e),ce(St,dr),e=t.nodeType,e){case 9:case 11:t=(t=t.documentElement)?t.namespaceURI:Ql(null,"");break;default:e=e===8?t.parentNode:t,t=e.namespaceURI||null,e=e.tagName,t=Ql(t,e)}de(St),ce(St,t)}function On(){de(St),de(pr),de(mr)}function Cs(e){un(mr.current);var t=un(St.current),n=Ql(t,e.type);t!==n&&(ce(pr,e),ce(St,n))}function Qu(e){pr.current===e&&(de(St),de(pr))}var he=Vt(0);function fl(e){for(var t=e;t!==null;){if(t.tag===13){var n=t.memoizedState;if(n!==null&&(n=n.dehydrated,n===null||n.data==="$?"||n.data==="$!"))return t}else if(t.tag===19&&t.memoizedProps.revealOrder!==void 0){if((t.flags&128)!==0)return t}else if(t.child!==null){t.child.return=t,t=t.child;continue}if(t===e)break;for(;t.sibling===null;){if(t.return===null||t.return===e)return null;t=t.return}t.sibling.return=t.return,t=t.sibling}return null}var Ku=[];function Yu(){for(var e=0;e<Ku.length;e++)Ku[e]._workInProgressVersionPrimary=null;Ku.length=0}var dl=ie.ReactCurrentDispatcher,Xu=ie.ReactCurrentBatchConfig,on=0,ve=null,_e=null,Ne=null,pl=!1,hr=!1,vr=0,Nf=0;function De(){throw Error(c(321))}function Gu(e,t){if(t===null)return!1;for(var n=0;n<t.length&&n<e.length;n++)if(!ct(e[n],t[n]))return!1;return!0}function Zu(e,t,n,r,l,u){if(on=u,ve=t,t.memoizedState=null,t.updateQueue=null,t.lanes=0,dl.current=e===null||e.memoizedState===null?Rf:Lf,e=n(r,l),hr){u=0;do{if(hr=!1,vr=0,25<=u)throw Error(c(301));u+=1,Ne=_e=null,t.updateQueue=null,dl.current=Tf,e=n(r,l)}while(hr)}if(dl.current=vl,t=_e!==null&&_e.next!==null,on=0,Ne=_e=ve=null,pl=!1,t)throw Error(c(300));return e}function Ju(){var e=vr!==0;return vr=0,e}function xt(){var e={memoizedState:null,baseState:null,baseQueue:null,queue:null,next:null};return Ne===null?ve.memoizedState=Ne=e:Ne=Ne.next=e,Ne}function lt(){if(_e===null){var e=ve.alternate;e=e!==null?e.memoizedState:null}else e=_e.next;var t=Ne===null?ve.memoizedState:Ne.next;if(t!==null)Ne=t,_e=e;else{if(e===null)throw Error(c(310));_e=e,e={memoizedState:_e.memoizedState,baseState:_e.baseState,baseQueue:_e.baseQueue,queue:_e.queue,next:null},Ne===null?ve.memoizedState=Ne=e:Ne=Ne.next=e}return Ne}function yr(e,t){return typeof t=="function"?t(e):t}function qu(e){var t=lt(),n=t.queue;if(n===null)throw Error(c(311));n.lastRenderedReducer=e;var r=_e,l=r.baseQueue,u=n.pending;if(u!==null){if(l!==null){var i=l.next;l.next=u.next,u.next=i}r.baseQueue=l=u,n.pending=null}if(l!==null){u=l.next,r=r.baseState;var o=i=null,s=null,h=u;do{var S=h.lane;if((on&S)===S)s!==null&&(s=s.next={lane:0,action:h.action,hasEagerState:h.hasEagerState,eagerState:h.eagerState,next:null}),r=h.hasEagerState?h.eagerState:e(r,h.action);else{var x={lane:S,action:h.action,hasEagerState:h.hasEagerState,eagerState:h.eagerState,next:null};s===null?(o=s=x,i=r):s=s.next=x,ve.lanes|=S,sn|=S}h=h.next}while(h!==null&&h!==u);s===null?i=r:s.next=o,ct(r,t.memoizedState)||(Ke=!0),t.memoizedState=r,t.baseState=i,t.baseQueue=s,n.lastRenderedState=r}if(e=n.interleaved,e!==null){l=e;do u=l.lane,ve.lanes|=u,sn|=u,l=l.next;while(l!==e)}else l===null&&(n.lanes=0);return[t.memoizedState,n.dispatch]}function bu(e){var t=lt(),n=t.queue;if(n===null)throw Error(c(311));n.lastRenderedReducer=e;var r=n.dispatch,l=n.pending,u=t.memoizedState;if(l!==null){n.pending=null;var i=l=l.next;do u=e(u,i.action),i=i.next;while(i!==l);ct(u,t.memoizedState)||(Ke=!0),t.memoizedState=u,t.baseQueue===null&&(t.baseState=u),n.lastRenderedState=u}return[u,r]}function Ns(){}function Ps(e,t){var n=ve,r=lt(),l=t(),u=!ct(r.memoizedState,l);if(u&&(r.memoizedState=l,Ke=!0),r=r.queue,ei(Rs.bind(null,n,r,e),[e]),r.getSnapshot!==t||u||Ne!==null&&Ne.memoizedState.tag&1){if(n.flags|=2048,gr(9,zs.bind(null,n,r,l,t),void 0,null),Pe===null)throw Error(c(349));(on&30)!==0||js(n,t,l)}return l}function js(e,t,n){e.flags|=16384,e={getSnapshot:t,value:n},t=ve.updateQueue,t===null?(t={lastEffect:null,stores:null},ve.updateQueue=t,t.stores=[e]):(n=t.stores,n===null?t.stores=[e]:n.push(e))}function zs(e,t,n,r){t.value=n,t.getSnapshot=r,Ls(t)&&Ts(e)}function Rs(e,t,n){return n(function(){Ls(t)&&Ts(e)})}function Ls(e){var t=e.getSnapshot;e=e.value;try{var n=t();return!ct(e,n)}catch{return!0}}function Ts(e){var t=Pt(e,1);t!==null&&ht(t,e,1,-1)}function Os(e){var t=xt();return typeof e=="function"&&(e=e()),t.memoizedState=t.baseState=e,e={pending:null,interleaved:null,lanes:0,dispatch:null,lastRenderedReducer:yr,lastRenderedState:e},t.queue=e,e=e.dispatch=zf.bind(null,ve,e),[t.memoizedState,e]}function gr(e,t,n,r){return e={tag:e,create:t,destroy:n,deps:r,next:null},t=ve.updateQueue,t===null?(t={lastEffect:null,stores:null},ve.updateQueue=t,t.lastEffect=e.next=e):(n=t.lastEffect,n===null?t.lastEffect=e.next=e:(r=n.next,n.next=e,e.next=r,t.lastEffect=e)),e}function Is(){return lt().memoizedState}function ml(e,t,n,r){var l=xt();ve.flags|=e,l.memoizedState=gr(1|t,n,void 0,r===void 0?null:r)}function hl(e,t,n,r){var l=lt();r=r===void 0?null:r;var u=void 0;if(_e!==null){var i=_e.memoizedState;if(u=i.destroy,r!==null&&Gu(r,i.deps)){l.memoizedState=gr(t,n,u,r);return}}ve.flags|=e,l.memoizedState=gr(1|t,n,u,r)}function Ds(e,t){return ml(8390656,8,e,t)}function ei(e,t){return hl(2048,8,e,t)}function Ms(e,t){return hl(4,2,e,t)}function Fs(e,t){return hl(4,4,e,t)}function Us(e,t){if(typeof t=="function")return e=e(),t(e),function(){t(null)};if(t!=null)return e=e(),t.current=e,function(){t.current=null}}function As(e,t,n){return n=n!=null?n.concat([e]):null,hl(4,4,Us.bind(null,t,e),n)}function ti(){}function Vs(e,t){var n=lt();t=t===void 0?null:t;var r=n.memoizedState;return r!==null&&t!==null&&Gu(t,r[1])?r[0]:(n.memoizedState=[e,t],e)}function $s(e,t){var n=lt();t=t===void 0?null:t;var r=n.memoizedState;return r!==null&&t!==null&&Gu(t,r[1])?r[0]:(e=e(),n.memoizedState=[e,t],e)}function Bs(e,t,n){return(on&21)===0?(e.baseState&&(e.baseState=!1,Ke=!0),e.memoizedState=n):(ct(n,t)||(n=go(),ve.lanes|=n,sn|=n,e.baseState=!0),t)}function Pf(e,t){var n=ue;ue=n!==0&&4>n?n:4,e(!0);var r=Xu.transition;Xu.transition={};try{e(!1),t()}finally{ue=n,Xu.transition=r}}function Ws(){return lt().memoizedState}function jf(e,t,n){var r=Xt(e);if(n={lane:r,action:n,hasEagerState:!1,eagerState:null,next:null},Hs(e))Qs(t,n);else if(n=xs(e,t,n,r),n!==null){var l=$e();ht(n,e,r,l),Ks(n,t,r)}}function zf(e,t,n){var r=Xt(e),l={lane:r,action:n,hasEagerState:!1,eagerState:null,next:null};if(Hs(e))Qs(t,l);else{var u=e.alternate;if(e.lanes===0&&(u===null||u.lanes===0)&&(u=t.lastRenderedReducer,u!==null))try{var i=t.lastRenderedState,o=u(i,n);if(l.hasEagerState=!0,l.eagerState=o,ct(o,i)){var s=t.interleaved;s===null?(l.next=l,Bu(t)):(l.next=s.next,s.next=l),t.interleaved=l;return}}catch{}finally{}n=xs(e,t,l,r),n!==null&&(l=$e(),ht(n,e,r,l),Ks(n,t,r))}}function Hs(e){var t=e.alternate;return e===ve||t!==null&&t===ve}function Qs(e,t){hr=pl=!0;var n=e.pending;n===null?t.next=t:(t.next=n.next,n.next=t),e.pending=t}function Ks(e,t,n){if((n&4194240)!==0){var r=t.lanes;r&=e.pendingLanes,n|=r,t.lanes=n,ru(e,n)}}var vl={readContext:rt,useCallback:De,useContext:De,useEffect:De,useImperativeHandle:De,useInsertionEffect:De,useLayoutEffect:De,useMemo:De,useReducer:De,useRef:De,useState:De,useDebugValue:De,useDeferredValue:De,useTransition:De,useMutableSource:De,useSyncExternalStore:De,useId:De,unstable_isNewReconciler:!1},Rf={readContext:rt,useCallback:function(e,t){return xt().memoizedState=[e,t===void 0?null:t],e},useContext:rt,useEffect:Ds,useImperativeHandle:function(e,t,n){return n=n!=null?n.concat([e]):null,ml(4194308,4,Us.bind(null,t,e),n)},useLayoutEffect:function(e,t){return ml(4194308,4,e,t)},useInsertionEffect:function(e,t){return ml(4,2,e,t)},useMemo:function(e,t){var n=xt();return t=t===void 0?null:t,e=e(),n.memoizedState=[e,t],e},useReducer:function(e,t,n){var r=xt();return t=n!==void 0?n(t):t,r.memoizedState=r.baseState=t,e={pending:null,interleaved:null,lanes:0,dispatch:null,lastRenderedReducer:e,lastRenderedState:t},r.queue=e,e=e.dispatch=jf.bind(null,ve,e),[r.memoizedState,e]},useRef:function(e){var t=xt();return e={current:e},t.memoizedState=e},useState:Os,useDebugValue:ti,useDeferredValue:function(e){return xt().memoizedState=e},useTransition:function(){var e=Os(!1),t=e[0];return e=Pf.bind(null,e[1]),xt().memoizedState=e,[t,e]},useMutableSource:function(){},useSyncExternalStore:function(e,t,n){var r=ve,l=xt();if(me){if(n===void 0)throw Error(c(407));n=n()}else{if(n=t(),Pe===null)throw Error(c(349));(on&30)!==0||js(r,t,n)}l.memoizedState=n;var u={value:n,getSnapshot:t};return l.queue=u,Ds(Rs.bind(null,r,u,e),[e]),r.flags|=2048,gr(9,zs.bind(null,r,u,n,t),void 0,null),n},useId:function(){var e=xt(),t=Pe.identifierPrefix;if(me){var n=Nt,r=Ct;n=(r&~(1<<32-at(r)-1)).toString(32)+n,t=":"+t+"R"+n,n=vr++,0<n&&(t+="H"+n.toString(32)),t+=":"}else n=Nf++,t=":"+t+"r"+n.toString(32)+":";return e.memoizedState=t},unstable_isNewReconciler:!1},Lf={readContext:rt,useCallback:Vs,useContext:rt,useEffect:ei,useImperativeHandle:As,useInsertionEffect:Ms,useLayoutEffect:Fs,useMemo:$s,useReducer:qu,useRef:Is,useState:function(){return qu(yr)},useDebugValue:ti,useDeferredValue:function(e){var t=lt();return Bs(t,_e.memoizedState,e)},useTransition:function(){var e=qu(yr)[0],t=lt().memoizedState;return[e,t]},useMutableSource:Ns,useSyncExternalStore:Ps,useId:Ws,unstable_isNewReconciler:!1},Tf={readContext:rt,useCallback:Vs,useContext:rt,useEffect:ei,useImperativeHandle:As,useInsertionEffect:Ms,useLayoutEffect:Fs,useMemo:$s,useReducer:bu,useRef:Is,useState:function(){return bu(yr)},useDebugValue:ti,useDeferredValue:function(e){var t=lt();return _e===null?t.memoizedState=e:Bs(t,_e.memoizedState,e)},useTransition:function(){var e=bu(yr)[0],t=lt().memoizedState;return[e,t]},useMutableSource:Ns,useSyncExternalStore:Ps,useId:Ws,unstable_isNewReconciler:!1};function dt(e,t){if(e&&e.defaultProps){t=N({},t),e=e.defaultProps;for(var n in e)t[n]===void 0&&(t[n]=e[n]);return t}return t}function ni(e,t,n,r){t=e.memoizedState,n=n(r,t),n=n==null?t:N({},t,n),e.memoizedState=n,e.lanes===0&&(e.updateQueue.baseState=n)}var yl={isMounted:function(e){return(e=e._reactInternals)?bt(e)===e:!1},enqueueSetState:function(e,t,n){e=e._reactInternals;var r=$e(),l=Xt(e),u=jt(r,l);u.payload=t,n!=null&&(u.callback=n),t=Ht(e,u,l),t!==null&&(ht(t,e,l,r),al(t,e,l))},enqueueReplaceState:function(e,t,n){e=e._reactInternals;var r=$e(),l=Xt(e),u=jt(r,l);u.tag=1,u.payload=t,n!=null&&(u.callback=n),t=Ht(e,u,l),t!==null&&(ht(t,e,l,r),al(t,e,l))},enqueueForceUpdate:function(e,t){e=e._reactInternals;var n=$e(),r=Xt(e),l=jt(n,r);l.tag=2,t!=null&&(l.callback=t),t=Ht(e,l,r),t!==null&&(ht(t,e,r,n),al(t,e,r))}};function Ys(e,t,n,r,l,u,i){return e=e.stateNode,typeof e.shouldComponentUpdate=="function"?e.shouldComponentUpdate(r,u,i):t.prototype&&t.prototype.isPureReactComponent?!lr(n,r)||!lr(l,u):!0}function Xs(e,t,n){var r=!1,l=$t,u=t.contextType;return typeof u=="object"&&u!==null?u=rt(u):(l=Qe(t)?tn:Ie.current,r=t.contextTypes,u=(r=r!=null)?Nn(e,l):$t),t=new t(n,u),e.memoizedState=t.state!==null&&t.state!==void 0?t.state:null,t.updater=yl,e.stateNode=t,t._reactInternals=e,r&&(e=e.stateNode,e.__reactInternalMemoizedUnmaskedChildContext=l,e.__reactInternalMemoizedMaskedChildContext=u),t}function Gs(e,t,n,r){e=t.state,typeof t.componentWillReceiveProps=="function"&&t.componentWillReceiveProps(n,r),typeof t.UNSAFE_componentWillReceiveProps=="function"&&t.UNSAFE_componentWillReceiveProps(n,r),t.state!==e&&yl.enqueueReplaceState(t,t.state,null)}function ri(e,t,n,r){var l=e.stateNode;l.props=n,l.state=e.memoizedState,l.refs={},Wu(e);var u=t.contextType;typeof u=="object"&&u!==null?l.context=rt(u):(u=Qe(t)?tn:Ie.current,l.context=Nn(e,u)),l.state=e.memoizedState,u=t.getDerivedStateFromProps,typeof u=="function"&&(ni(e,t,u,n),l.state=e.memoizedState),typeof t.getDerivedStateFromProps=="function"||typeof l.getSnapshotBeforeUpdate=="function"||typeof l.UNSAFE_componentWillMount!="function"&&typeof l.componentWillMount!="function"||(t=l.state,typeof l.componentWillMount=="function"&&l.componentWillMount(),typeof l.UNSAFE_componentWillMount=="function"&&l.UNSAFE_componentWillMount(),t!==l.state&&yl.enqueueReplaceState(l,l.state,null),cl(e,n,l,r),l.state=e.memoizedState),typeof l.componentDidMount=="function"&&(e.flags|=4194308)}function In(e,t){try{var n="",r=t;do n+=b(r),r=r.return;while(r);var l=n}catch(u){l=`
+Error generating stack: `+u.message+`
+`+u.stack}return{value:e,source:t,stack:l,digest:null}}function li(e,t,n){return{value:e,source:null,stack:n??null,digest:t??null}}function ui(e,t){try{console.error(t.value)}catch(n){setTimeout(function(){throw n})}}var Of=typeof WeakMap=="function"?WeakMap:Map;function Zs(e,t,n){n=jt(-1,n),n.tag=3,n.payload={element:null};var r=t.value;return n.callback=function(){_l||(_l=!0,Si=r),ui(e,t)},n}function Js(e,t,n){n=jt(-1,n),n.tag=3;var r=e.type.getDerivedStateFromError;if(typeof r=="function"){var l=t.value;n.payload=function(){return r(l)},n.callback=function(){ui(e,t)}}var u=e.stateNode;return u!==null&&typeof u.componentDidCatch=="function"&&(n.callback=function(){ui(e,t),typeof r!="function"&&(Kt===null?Kt=new Set([this]):Kt.add(this));var i=t.stack;this.componentDidCatch(t.value,{componentStack:i!==null?i:""})}),n}function qs(e,t,n){var r=e.pingCache;if(r===null){r=e.pingCache=new Of;var l=new Set;r.set(t,l)}else l=r.get(t),l===void 0&&(l=new Set,r.set(t,l));l.has(n)||(l.add(n),e=Yf.bind(null,e,t,n),t.then(e,e))}function bs(e){do{var t;if((t=e.tag===13)&&(t=e.memoizedState,t=t!==null?t.dehydrated!==null:!0),t)return e;e=e.return}while(e!==null);return null}function ea(e,t,n,r,l){return(e.mode&1)===0?(e===t?e.flags|=65536:(e.flags|=128,n.flags|=131072,n.flags&=-52805,n.tag===1&&(n.alternate===null?n.tag=17:(t=jt(-1,1),t.tag=2,Ht(n,t,1))),n.lanes|=1),e):(e.flags|=65536,e.lanes=l,e)}var If=ie.ReactCurrentOwner,Ke=!1;function Ve(e,t,n,r){t.child=e===null?Ss(t,null,n,r):Rn(t,e.child,n,r)}function ta(e,t,n,r,l){n=n.render;var u=t.ref;return Tn(t,l),r=Zu(e,t,n,r,u,l),n=Ju(),e!==null&&!Ke?(t.updateQueue=e.updateQueue,t.flags&=-2053,e.lanes&=~l,zt(e,t,l)):(me&&n&&Ou(t),t.flags|=1,Ve(e,t,r,l),t.child)}function na(e,t,n,r,l){if(e===null){var u=n.type;return typeof u=="function"&&!Pi(u)&&u.defaultProps===void 0&&n.compare===null&&n.defaultProps===void 0?(t.tag=15,t.type=u,ra(e,t,u,r,l)):(e=Rl(n.type,null,r,t,t.mode,l),e.ref=t.ref,e.return=t,t.child=e)}if(u=e.child,(e.lanes&l)===0){var i=u.memoizedProps;if(n=n.compare,n=n!==null?n:lr,n(i,r)&&e.ref===t.ref)return zt(e,t,l)}return t.flags|=1,e=Zt(u,r),e.ref=t.ref,e.return=t,t.child=e}function ra(e,t,n,r,l){if(e!==null){var u=e.memoizedProps;if(lr(u,r)&&e.ref===t.ref)if(Ke=!1,t.pendingProps=r=u,(e.lanes&l)!==0)(e.flags&131072)!==0&&(Ke=!0);else return t.lanes=e.lanes,zt(e,t,l)}return ii(e,t,n,r,l)}function la(e,t,n){var r=t.pendingProps,l=r.children,u=e!==null?e.memoizedState:null;if(r.mode==="hidden")if((t.mode&1)===0)t.memoizedState={baseLanes:0,cachePool:null,transitions:null},ce(Mn,et),et|=n;else{if((n&1073741824)===0)return e=u!==null?u.baseLanes|n:n,t.lanes=t.childLanes=1073741824,t.memoizedState={baseLanes:e,cachePool:null,transitions:null},t.updateQueue=null,ce(Mn,et),et|=e,null;t.memoizedState={baseLanes:0,cachePool:null,transitions:null},r=u!==null?u.baseLanes:n,ce(Mn,et),et|=r}else u!==null?(r=u.baseLanes|n,t.memoizedState=null):r=n,ce(Mn,et),et|=r;return Ve(e,t,l,n),t.child}function ua(e,t){var n=t.ref;(e===null&&n!==null||e!==null&&e.ref!==n)&&(t.flags|=512,t.flags|=2097152)}function ii(e,t,n,r,l){var u=Qe(n)?tn:Ie.current;return u=Nn(t,u),Tn(t,l),n=Zu(e,t,n,r,u,l),r=Ju(),e!==null&&!Ke?(t.updateQueue=e.updateQueue,t.flags&=-2053,e.lanes&=~l,zt(e,t,l)):(me&&r&&Ou(t),t.flags|=1,Ve(e,t,n,l),t.child)}function ia(e,t,n,r,l){if(Qe(n)){var u=!0;tl(t)}else u=!1;if(Tn(t,l),t.stateNode===null)wl(e,t),Xs(t,n,r),ri(t,n,r,l),r=!0;else if(e===null){var i=t.stateNode,o=t.memoizedProps;i.props=o;var s=i.context,h=n.contextType;typeof h=="object"&&h!==null?h=rt(h):(h=Qe(n)?tn:Ie.current,h=Nn(t,h));var S=n.getDerivedStateFromProps,x=typeof S=="function"||typeof i.getSnapshotBeforeUpdate=="function";x||typeof i.UNSAFE_componentWillReceiveProps!="function"&&typeof i.componentWillReceiveProps!="function"||(o!==r||s!==h)&&Gs(t,i,r,h),Wt=!1;var g=t.memoizedState;i.state=g,cl(t,r,i,l),s=t.memoizedState,o!==r||g!==s||He.current||Wt?(typeof S=="function"&&(ni(t,n,S,r),s=t.memoizedState),(o=Wt||Ys(t,n,o,r,g,s,h))?(x||typeof i.UNSAFE_componentWillMount!="function"&&typeof i.componentWillMount!="function"||(typeof i.componentWillMount=="function"&&i.componentWillMount(),typeof i.UNSAFE_componentWillMount=="function"&&i.UNSAFE_componentWillMount()),typeof i.componentDidMount=="function"&&(t.flags|=4194308)):(typeof i.componentDidMount=="function"&&(t.flags|=4194308),t.memoizedProps=r,t.memoizedState=s),i.props=r,i.state=s,i.context=h,r=o):(typeof i.componentDidMount=="function"&&(t.flags|=4194308),r=!1)}else{i=t.stateNode,ks(e,t),o=t.memoizedProps,h=t.type===t.elementType?o:dt(t.type,o),i.props=h,x=t.pendingProps,g=i.context,s=n.contextType,typeof s=="object"&&s!==null?s=rt(s):(s=Qe(n)?tn:Ie.current,s=Nn(t,s));var _=n.getDerivedStateFromProps;(S=typeof _=="function"||typeof i.getSnapshotBeforeUpdate=="function")||typeof i.UNSAFE_componentWillReceiveProps!="function"&&typeof i.componentWillReceiveProps!="function"||(o!==x||g!==s)&&Gs(t,i,r,s),Wt=!1,g=t.memoizedState,i.state=g,cl(t,r,i,l);var j=t.memoizedState;o!==x||g!==j||He.current||Wt?(typeof _=="function"&&(ni(t,n,_,r),j=t.memoizedState),(h=Wt||Ys(t,n,h,r,g,j,s)||!1)?(S||typeof i.UNSAFE_componentWillUpdate!="function"&&typeof i.componentWillUpdate!="function"||(typeof i.componentWillUpdate=="function"&&i.componentWillUpdate(r,j,s),typeof i.UNSAFE_componentWillUpdate=="function"&&i.UNSAFE_componentWillUpdate(r,j,s)),typeof i.componentDidUpdate=="function"&&(t.flags|=4),typeof i.getSnapshotBeforeUpdate=="function"&&(t.flags|=1024)):(typeof i.componentDidUpdate!="function"||o===e.memoizedProps&&g===e.memoizedState||(t.flags|=4),typeof i.getSnapshotBeforeUpdate!="function"||o===e.memoizedProps&&g===e.memoizedState||(t.flags|=1024),t.memoizedProps=r,t.memoizedState=j),i.props=r,i.state=j,i.context=s,r=h):(typeof i.componentDidUpdate!="function"||o===e.memoizedProps&&g===e.memoizedState||(t.flags|=4),typeof i.getSnapshotBeforeUpdate!="function"||o===e.memoizedProps&&g===e.memoizedState||(t.flags|=1024),r=!1)}return oi(e,t,n,r,u,l)}function oi(e,t,n,r,l,u){ua(e,t);var i=(t.flags&128)!==0;if(!r&&!i)return l&&fs(t,n,!1),zt(e,t,u);r=t.stateNode,If.current=t;var o=i&&typeof n.getDerivedStateFromError!="function"?null:r.render();return t.flags|=1,e!==null&&i?(t.child=Rn(t,e.child,null,u),t.child=Rn(t,null,o,u)):Ve(e,t,o,u),t.memoizedState=r.state,l&&fs(t,n,!0),t.child}function oa(e){var t=e.stateNode;t.pendingContext?as(e,t.pendingContext,t.pendingContext!==t.context):t.context&&as(e,t.context,!1),Hu(e,t.containerInfo)}function sa(e,t,n,r,l){return zn(),Fu(l),t.flags|=256,Ve(e,t,n,r),t.child}var si={dehydrated:null,treeContext:null,retryLane:0};function ai(e){return{baseLanes:e,cachePool:null,transitions:null}}function aa(e,t,n){var r=t.pendingProps,l=he.current,u=!1,i=(t.flags&128)!==0,o;if((o=i)||(o=e!==null&&e.memoizedState===null?!1:(l&2)!==0),o?(u=!0,t.flags&=-129):(e===null||e.memoizedState!==null)&&(l|=1),ce(he,l&1),e===null)return Mu(t),e=t.memoizedState,e!==null&&(e=e.dehydrated,e!==null)?((t.mode&1)===0?t.lanes=1:e.data==="$!"?t.lanes=8:t.lanes=1073741824,null):(i=r.children,e=r.fallback,u?(r=t.mode,u=t.child,i={mode:"hidden",children:i},(r&1)===0&&u!==null?(u.childLanes=0,u.pendingProps=i):u=Ll(i,r,0,null),e=dn(e,r,n,null),u.return=t,e.return=t,u.sibling=e,t.child=u,t.child.memoizedState=ai(n),t.memoizedState=si,e):ci(t,i));if(l=e.memoizedState,l!==null&&(o=l.dehydrated,o!==null))return Df(e,t,i,r,o,l,n);if(u){u=r.fallback,i=t.mode,l=e.child,o=l.sibling;var s={mode:"hidden",children:r.children};return(i&1)===0&&t.child!==l?(r=t.child,r.childLanes=0,r.pendingProps=s,t.deletions=null):(r=Zt(l,s),r.subtreeFlags=l.subtreeFlags&14680064),o!==null?u=Zt(o,u):(u=dn(u,i,n,null),u.flags|=2),u.return=t,r.return=t,r.sibling=u,t.child=r,r=u,u=t.child,i=e.child.memoizedState,i=i===null?ai(n):{baseLanes:i.baseLanes|n,cachePool:null,transitions:i.transitions},u.memoizedState=i,u.childLanes=e.childLanes&~n,t.memoizedState=si,r}return u=e.child,e=u.sibling,r=Zt(u,{mode:"visible",children:r.children}),(t.mode&1)===0&&(r.lanes=n),r.return=t,r.sibling=null,e!==null&&(n=t.deletions,n===null?(t.deletions=[e],t.flags|=16):n.push(e)),t.child=r,t.memoizedState=null,r}function ci(e,t){return t=Ll({mode:"visible",children:t},e.mode,0,null),t.return=e,e.child=t}function gl(e,t,n,r){return r!==null&&Fu(r),Rn(t,e.child,null,n),e=ci(t,t.pendingProps.children),e.flags|=2,t.memoizedState=null,e}function Df(e,t,n,r,l,u,i){if(n)return t.flags&256?(t.flags&=-257,r=li(Error(c(422))),gl(e,t,i,r)):t.memoizedState!==null?(t.child=e.child,t.flags|=128,null):(u=r.fallback,l=t.mode,r=Ll({mode:"visible",children:r.children},l,0,null),u=dn(u,l,i,null),u.flags|=2,r.return=t,u.return=t,r.sibling=u,t.child=r,(t.mode&1)!==0&&Rn(t,e.child,null,i),t.child.memoizedState=ai(i),t.memoizedState=si,u);if((t.mode&1)===0)return gl(e,t,i,null);if(l.data==="$!"){if(r=l.nextSibling&&l.nextSibling.dataset,r)var o=r.dgst;return r=o,u=Error(c(419)),r=li(u,r,void 0),gl(e,t,i,r)}if(o=(i&e.childLanes)!==0,Ke||o){if(r=Pe,r!==null){switch(i&-i){case 4:l=2;break;case 16:l=8;break;case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:l=32;break;case 536870912:l=268435456;break;default:l=0}l=(l&(r.suspendedLanes|i))!==0?0:l,l!==0&&l!==u.retryLane&&(u.retryLane=l,Pt(e,l),ht(r,e,l,-1))}return Ni(),r=li(Error(c(421))),gl(e,t,i,r)}return l.data==="$?"?(t.flags|=128,t.child=e.child,t=Xf.bind(null,e),l._reactRetry=t,null):(e=u.treeContext,be=At(l.nextSibling),qe=t,me=!0,ft=null,e!==null&&(tt[nt++]=Ct,tt[nt++]=Nt,tt[nt++]=nn,Ct=e.id,Nt=e.overflow,nn=t),t=ci(t,r.children),t.flags|=4096,t)}function ca(e,t,n){e.lanes|=t;var r=e.alternate;r!==null&&(r.lanes|=t),$u(e.return,t,n)}function fi(e,t,n,r,l){var u=e.memoizedState;u===null?e.memoizedState={isBackwards:t,rendering:null,renderingStartTime:0,last:r,tail:n,tailMode:l}:(u.isBackwards=t,u.rendering=null,u.renderingStartTime=0,u.last=r,u.tail=n,u.tailMode=l)}function fa(e,t,n){var r=t.pendingProps,l=r.revealOrder,u=r.tail;if(Ve(e,t,r.children,n),r=he.current,(r&2)!==0)r=r&1|2,t.flags|=128;else{if(e!==null&&(e.flags&128)!==0)e:for(e=t.child;e!==null;){if(e.tag===13)e.memoizedState!==null&&ca(e,n,t);else if(e.tag===19)ca(e,n,t);else if(e.child!==null){e.child.return=e,e=e.child;continue}if(e===t)break e;for(;e.sibling===null;){if(e.return===null||e.return===t)break e;e=e.return}e.sibling.return=e.return,e=e.sibling}r&=1}if(ce(he,r),(t.mode&1)===0)t.memoizedState=null;else switch(l){case"forwards":for(n=t.child,l=null;n!==null;)e=n.alternate,e!==null&&fl(e)===null&&(l=n),n=n.sibling;n=l,n===null?(l=t.child,t.child=null):(l=n.sibling,n.sibling=null),fi(t,!1,l,n,u);break;case"backwards":for(n=null,l=t.child,t.child=null;l!==null;){if(e=l.alternate,e!==null&&fl(e)===null){t.child=l;break}e=l.sibling,l.sibling=n,n=l,l=e}fi(t,!0,n,null,u);break;case"together":fi(t,!1,null,null,void 0);break;default:t.memoizedState=null}return t.child}function wl(e,t){(t.mode&1)===0&&e!==null&&(e.alternate=null,t.alternate=null,t.flags|=2)}function zt(e,t,n){if(e!==null&&(t.dependencies=e.dependencies),sn|=t.lanes,(n&t.childLanes)===0)return null;if(e!==null&&t.child!==e.child)throw Error(c(153));if(t.child!==null){for(e=t.child,n=Zt(e,e.pendingProps),t.child=n,n.return=t;e.sibling!==null;)e=e.sibling,n=n.sibling=Zt(e,e.pendingProps),n.return=t;n.sibling=null}return t.child}function Mf(e,t,n){switch(t.tag){case 3:oa(t),zn();break;case 5:Cs(t);break;case 1:Qe(t.type)&&tl(t);break;case 4:Hu(t,t.stateNode.containerInfo);break;case 10:var r=t.type._context,l=t.memoizedProps.value;ce(ol,r._currentValue),r._currentValue=l;break;case 13:if(r=t.memoizedState,r!==null)return r.dehydrated!==null?(ce(he,he.current&1),t.flags|=128,null):(n&t.child.childLanes)!==0?aa(e,t,n):(ce(he,he.current&1),e=zt(e,t,n),e!==null?e.sibling:null);ce(he,he.current&1);break;case 19:if(r=(n&t.childLanes)!==0,(e.flags&128)!==0){if(r)return fa(e,t,n);t.flags|=128}if(l=t.memoizedState,l!==null&&(l.rendering=null,l.tail=null,l.lastEffect=null),ce(he,he.current),r)break;return null;case 22:case 23:return t.lanes=0,la(e,t,n)}return zt(e,t,n)}var da,di,pa,ma;da=function(e,t){for(var n=t.child;n!==null;){if(n.tag===5||n.tag===6)e.appendChild(n.stateNode);else if(n.tag!==4&&n.child!==null){n.child.return=n,n=n.child;continue}if(n===t)break;for(;n.sibling===null;){if(n.return===null||n.return===t)return;n=n.return}n.sibling.return=n.return,n=n.sibling}},di=function(){},pa=function(e,t,n,r){var l=e.memoizedProps;if(l!==r){e=t.stateNode,un(St.current);var u=null;switch(n){case"input":l=$l(e,l),r=$l(e,r),u=[];break;case"select":l=N({},l,{value:void 0}),r=N({},r,{value:void 0}),u=[];break;case"textarea":l=Hl(e,l),r=Hl(e,r),u=[];break;default:typeof l.onClick!="function"&&typeof r.onClick=="function"&&(e.onclick=qr)}Kl(n,r);var i;n=null;for(h in l)if(!r.hasOwnProperty(h)&&l.hasOwnProperty(h)&&l[h]!=null)if(h==="style"){var o=l[h];for(i in o)o.hasOwnProperty(i)&&(n||(n={}),n[i]="")}else h!=="dangerouslySetInnerHTML"&&h!=="children"&&h!=="suppressContentEditableWarning"&&h!=="suppressHydrationWarning"&&h!=="autoFocus"&&(R.hasOwnProperty(h)?u||(u=[]):(u=u||[]).push(h,null));for(h in r){var s=r[h];if(o=l!=null?l[h]:void 0,r.hasOwnProperty(h)&&s!==o&&(s!=null||o!=null))if(h==="style")if(o){for(i in o)!o.hasOwnProperty(i)||s&&s.hasOwnProperty(i)||(n||(n={}),n[i]="");for(i in s)s.hasOwnProperty(i)&&o[i]!==s[i]&&(n||(n={}),n[i]=s[i])}else n||(u||(u=[]),u.push(h,n)),n=s;else h==="dangerouslySetInnerHTML"?(s=s?s.__html:void 0,o=o?o.__html:void 0,s!=null&&o!==s&&(u=u||[]).push(h,s)):h==="children"?typeof s!="string"&&typeof s!="number"||(u=u||[]).push(h,""+s):h!=="suppressContentEditableWarning"&&h!=="suppressHydrationWarning"&&(R.hasOwnProperty(h)?(s!=null&&h==="onScroll"&&fe("scroll",e),u||o===s||(u=[])):(u=u||[]).push(h,s))}n&&(u=u||[]).push("style",n);var h=u;(t.updateQueue=h)&&(t.flags|=4)}},ma=function(e,t,n,r){n!==r&&(t.flags|=4)};function wr(e,t){if(!me)switch(e.tailMode){case"hidden":t=e.tail;for(var n=null;t!==null;)t.alternate!==null&&(n=t),t=t.sibling;n===null?e.tail=null:n.sibling=null;break;case"collapsed":n=e.tail;for(var r=null;n!==null;)n.alternate!==null&&(r=n),n=n.sibling;r===null?t||e.tail===null?e.tail=null:e.tail.sibling=null:r.sibling=null}}function Me(e){var t=e.alternate!==null&&e.alternate.child===e.child,n=0,r=0;if(t)for(var l=e.child;l!==null;)n|=l.lanes|l.childLanes,r|=l.subtreeFlags&14680064,r|=l.flags&14680064,l.return=e,l=l.sibling;else for(l=e.child;l!==null;)n|=l.lanes|l.childLanes,r|=l.subtreeFlags,r|=l.flags,l.return=e,l=l.sibling;return e.subtreeFlags|=r,e.childLanes=n,t}function Ff(e,t,n){var r=t.pendingProps;switch(Iu(t),t.tag){case 2:case 16:case 15:case 0:case 11:case 7:case 8:case 12:case 9:case 14:return Me(t),null;case 1:return Qe(t.type)&&el(),Me(t),null;case 3:return r=t.stateNode,On(),de(He),de(Ie),Yu(),r.pendingContext&&(r.context=r.pendingContext,r.pendingContext=null),(e===null||e.child===null)&&(ul(t)?t.flags|=4:e===null||e.memoizedState.isDehydrated&&(t.flags&256)===0||(t.flags|=1024,ft!==null&&(Ei(ft),ft=null))),di(e,t),Me(t),null;case 5:Qu(t);var l=un(mr.current);if(n=t.type,e!==null&&t.stateNode!=null)pa(e,t,n,r,l),e.ref!==t.ref&&(t.flags|=512,t.flags|=2097152);else{if(!r){if(t.stateNode===null)throw Error(c(166));return Me(t),null}if(e=un(St.current),ul(t)){r=t.stateNode,n=t.type;var u=t.memoizedProps;switch(r[wt]=t,r[ar]=u,e=(t.mode&1)!==0,n){case"dialog":fe("cancel",r),fe("close",r);break;case"iframe":case"object":case"embed":fe("load",r);break;case"video":case"audio":for(l=0;l<ir.length;l++)fe(ir[l],r);break;case"source":fe("error",r);break;case"img":case"image":case"link":fe("error",r),fe("load",r);break;case"details":fe("toggle",r);break;case"input":Yi(r,u),fe("invalid",r);break;case"select":r._wrapperState={wasMultiple:!!u.multiple},fe("invalid",r);break;case"textarea":Zi(r,u),fe("invalid",r)}Kl(n,u),l=null;for(var i in u)if(u.hasOwnProperty(i)){var o=u[i];i==="children"?typeof o=="string"?r.textContent!==o&&(u.suppressHydrationWarning!==!0&&Jr(r.textContent,o,e),l=["children",o]):typeof o=="number"&&r.textContent!==""+o&&(u.suppressHydrationWarning!==!0&&Jr(r.textContent,o,e),l=["children",""+o]):R.hasOwnProperty(i)&&o!=null&&i==="onScroll"&&fe("scroll",r)}switch(n){case"input":jr(r),Gi(r,u,!0);break;case"textarea":jr(r),qi(r);break;case"select":case"option":break;default:typeof u.onClick=="function"&&(r.onclick=qr)}r=l,t.updateQueue=r,r!==null&&(t.flags|=4)}else{i=l.nodeType===9?l:l.ownerDocument,e==="http://www.w3.org/1999/xhtml"&&(e=bi(n)),e==="http://www.w3.org/1999/xhtml"?n==="script"?(e=i.createElement("div"),e.innerHTML="<script><\/script>",e=e.removeChild(e.firstChild)):typeof r.is=="string"?e=i.createElement(n,{is:r.is}):(e=i.createElement(n),n==="select"&&(i=e,r.multiple?i.multiple=!0:r.size&&(i.size=r.size))):e=i.createElementNS(e,n),e[wt]=t,e[ar]=r,da(e,t,!1,!1),t.stateNode=e;e:{switch(i=Yl(n,r),n){case"dialog":fe("cancel",e),fe("close",e),l=r;break;case"iframe":case"object":case"embed":fe("load",e),l=r;break;case"video":case"audio":for(l=0;l<ir.length;l++)fe(ir[l],e);l=r;break;case"source":fe("error",e),l=r;break;case"img":case"image":case"link":fe("error",e),fe("load",e),l=r;break;case"details":fe("toggle",e),l=r;break;case"input":Yi(e,r),l=$l(e,r),fe("invalid",e);break;case"option":l=r;break;case"select":e._wrapperState={wasMultiple:!!r.multiple},l=N({},r,{value:void 0}),fe("invalid",e);break;case"textarea":Zi(e,r),l=Hl(e,r),fe("invalid",e);break;default:l=r}Kl(n,l),o=l;for(u in o)if(o.hasOwnProperty(u)){var s=o[u];u==="style"?no(e,s):u==="dangerouslySetInnerHTML"?(s=s?s.__html:void 0,s!=null&&eo(e,s)):u==="children"?typeof s=="string"?(n!=="textarea"||s!=="")&&$n(e,s):typeof s=="number"&&$n(e,""+s):u!=="suppressContentEditableWarning"&&u!=="suppressHydrationWarning"&&u!=="autoFocus"&&(R.hasOwnProperty(u)?s!=null&&u==="onScroll"&&fe("scroll",e):s!=null&&ke(e,u,s,i))}switch(n){case"input":jr(e),Gi(e,r,!1);break;case"textarea":jr(e),qi(e);break;case"option":r.value!=null&&e.setAttribute("value",""+le(r.value));break;case"select":e.multiple=!!r.multiple,u=r.value,u!=null?mn(e,!!r.multiple,u,!1):r.defaultValue!=null&&mn(e,!!r.multiple,r.defaultValue,!0);break;default:typeof l.onClick=="function"&&(e.onclick=qr)}switch(n){case"button":case"input":case"select":case"textarea":r=!!r.autoFocus;break e;case"img":r=!0;break e;default:r=!1}}r&&(t.flags|=4)}t.ref!==null&&(t.flags|=512,t.flags|=2097152)}return Me(t),null;case 6:if(e&&t.stateNode!=null)ma(e,t,e.memoizedProps,r);else{if(typeof r!="string"&&t.stateNode===null)throw Error(c(166));if(n=un(mr.current),un(St.current),ul(t)){if(r=t.stateNode,n=t.memoizedProps,r[wt]=t,(u=r.nodeValue!==n)&&(e=qe,e!==null))switch(e.tag){case 3:Jr(r.nodeValue,n,(e.mode&1)!==0);break;case 5:e.memoizedProps.suppressHydrationWarning!==!0&&Jr(r.nodeValue,n,(e.mode&1)!==0)}u&&(t.flags|=4)}else r=(n.nodeType===9?n:n.ownerDocument).createTextNode(r),r[wt]=t,t.stateNode=r}return Me(t),null;case 13:if(de(he),r=t.memoizedState,e===null||e.memoizedState!==null&&e.memoizedState.dehydrated!==null){if(me&&be!==null&&(t.mode&1)!==0&&(t.flags&128)===0)ys(),zn(),t.flags|=98560,u=!1;else if(u=ul(t),r!==null&&r.dehydrated!==null){if(e===null){if(!u)throw Error(c(318));if(u=t.memoizedState,u=u!==null?u.dehydrated:null,!u)throw Error(c(317));u[wt]=t}else zn(),(t.flags&128)===0&&(t.memoizedState=null),t.flags|=4;Me(t),u=!1}else ft!==null&&(Ei(ft),ft=null),u=!0;if(!u)return t.flags&65536?t:null}return(t.flags&128)!==0?(t.lanes=n,t):(r=r!==null,r!==(e!==null&&e.memoizedState!==null)&&r&&(t.child.flags|=8192,(t.mode&1)!==0&&(e===null||(he.current&1)!==0?Ce===0&&(Ce=3):Ni())),t.updateQueue!==null&&(t.flags|=4),Me(t),null);case 4:return On(),di(e,t),e===null&&or(t.stateNode.containerInfo),Me(t),null;case 10:return Vu(t.type._context),Me(t),null;case 17:return Qe(t.type)&&el(),Me(t),null;case 19:if(de(he),u=t.memoizedState,u===null)return Me(t),null;if(r=(t.flags&128)!==0,i=u.rendering,i===null)if(r)wr(u,!1);else{if(Ce!==0||e!==null&&(e.flags&128)!==0)for(e=t.child;e!==null;){if(i=fl(e),i!==null){for(t.flags|=128,wr(u,!1),r=i.updateQueue,r!==null&&(t.updateQueue=r,t.flags|=4),t.subtreeFlags=0,r=n,n=t.child;n!==null;)u=n,e=r,u.flags&=14680066,i=u.alternate,i===null?(u.childLanes=0,u.lanes=e,u.child=null,u.subtreeFlags=0,u.memoizedProps=null,u.memoizedState=null,u.updateQueue=null,u.dependencies=null,u.stateNode=null):(u.childLanes=i.childLanes,u.lanes=i.lanes,u.child=i.child,u.subtreeFlags=0,u.deletions=null,u.memoizedProps=i.memoizedProps,u.memoizedState=i.memoizedState,u.updateQueue=i.updateQueue,u.type=i.type,e=i.dependencies,u.dependencies=e===null?null:{lanes:e.lanes,firstContext:e.firstContext}),n=n.sibling;return ce(he,he.current&1|2),t.child}e=e.sibling}u.tail!==null&&Se()>Fn&&(t.flags|=128,r=!0,wr(u,!1),t.lanes=4194304)}else{if(!r)if(e=fl(i),e!==null){if(t.flags|=128,r=!0,n=e.updateQueue,n!==null&&(t.updateQueue=n,t.flags|=4),wr(u,!0),u.tail===null&&u.tailMode==="hidden"&&!i.alternate&&!me)return Me(t),null}else 2*Se()-u.renderingStartTime>Fn&&n!==1073741824&&(t.flags|=128,r=!0,wr(u,!1),t.lanes=4194304);u.isBackwards?(i.sibling=t.child,t.child=i):(n=u.last,n!==null?n.sibling=i:t.child=i,u.last=i)}return u.tail!==null?(t=u.tail,u.rendering=t,u.tail=t.sibling,u.renderingStartTime=Se(),t.sibling=null,n=he.current,ce(he,r?n&1|2:n&1),t):(Me(t),null);case 22:case 23:return Ci(),r=t.memoizedState!==null,e!==null&&e.memoizedState!==null!==r&&(t.flags|=8192),r&&(t.mode&1)!==0?(et&1073741824)!==0&&(Me(t),t.subtreeFlags&6&&(t.flags|=8192)):Me(t),null;case 24:return null;case 25:return null}throw Error(c(156,t.tag))}function Uf(e,t){switch(Iu(t),t.tag){case 1:return Qe(t.type)&&el(),e=t.flags,e&65536?(t.flags=e&-65537|128,t):null;case 3:return On(),de(He),de(Ie),Yu(),e=t.flags,(e&65536)!==0&&(e&128)===0?(t.flags=e&-65537|128,t):null;case 5:return Qu(t),null;case 13:if(de(he),e=t.memoizedState,e!==null&&e.dehydrated!==null){if(t.alternate===null)throw Error(c(340));zn()}return e=t.flags,e&65536?(t.flags=e&-65537|128,t):null;case 19:return de(he),null;case 4:return On(),null;case 10:return Vu(t.type._context),null;case 22:case 23:return Ci(),null;case 24:return null;default:return null}}var Sl=!1,Fe=!1,Af=typeof WeakSet=="function"?WeakSet:Set,C=null;function Dn(e,t){var n=e.ref;if(n!==null)if(typeof n=="function")try{n(null)}catch(r){we(e,t,r)}else n.current=null}function pi(e,t,n){try{n()}catch(r){we(e,t,r)}}var ha=!1;function Vf(e,t){if(Cu=Vr,e=Yo(),yu(e)){if("selectionStart"in e)var n={start:e.selectionStart,end:e.selectionEnd};else e:{n=(n=e.ownerDocument)&&n.defaultView||window;var r=n.getSelection&&n.getSelection();if(r&&r.rangeCount!==0){n=r.anchorNode;var l=r.anchorOffset,u=r.focusNode;r=r.focusOffset;try{n.nodeType,u.nodeType}catch{n=null;break e}var i=0,o=-1,s=-1,h=0,S=0,x=e,g=null;t:for(;;){for(var _;x!==n||l!==0&&x.nodeType!==3||(o=i+l),x!==u||r!==0&&x.nodeType!==3||(s=i+r),x.nodeType===3&&(i+=x.nodeValue.length),(_=x.firstChild)!==null;)g=x,x=_;for(;;){if(x===e)break t;if(g===n&&++h===l&&(o=i),g===u&&++S===r&&(s=i),(_=x.nextSibling)!==null)break;x=g,g=x.parentNode}x=_}n=o===-1||s===-1?null:{start:o,end:s}}else n=null}n=n||{start:0,end:0}}else n=null;for(Nu={focusedElem:e,selectionRange:n},Vr=!1,C=t;C!==null;)if(t=C,e=t.child,(t.subtreeFlags&1028)!==0&&e!==null)e.return=t,C=e;else for(;C!==null;){t=C;try{var j=t.alternate;if((t.flags&1024)!==0)switch(t.tag){case 0:case 11:case 15:break;case 1:if(j!==null){var z=j.memoizedProps,xe=j.memoizedState,d=t.stateNode,a=d.getSnapshotBeforeUpdate(t.elementType===t.type?z:dt(t.type,z),xe);d.__reactInternalSnapshotBeforeUpdate=a}break;case 3:var p=t.stateNode.containerInfo;p.nodeType===1?p.textContent="":p.nodeType===9&&p.documentElement&&p.removeChild(p.documentElement);break;case 5:case 6:case 4:case 17:break;default:throw Error(c(163))}}catch(k){we(t,t.return,k)}if(e=t.sibling,e!==null){e.return=t.return,C=e;break}C=t.return}return j=ha,ha=!1,j}function Sr(e,t,n){var r=t.updateQueue;if(r=r!==null?r.lastEffect:null,r!==null){var l=r=r.next;do{if((l.tag&e)===e){var u=l.destroy;l.destroy=void 0,u!==void 0&&pi(t,n,u)}l=l.next}while(l!==r)}}function xl(e,t){if(t=t.updateQueue,t=t!==null?t.lastEffect:null,t!==null){var n=t=t.next;do{if((n.tag&e)===e){var r=n.create;n.destroy=r()}n=n.next}while(n!==t)}}function mi(e){var t=e.ref;if(t!==null){var n=e.stateNode;switch(e.tag){case 5:e=n;break;default:e=n}typeof t=="function"?t(e):t.current=e}}function va(e){var t=e.alternate;t!==null&&(e.alternate=null,va(t)),e.child=null,e.deletions=null,e.sibling=null,e.tag===5&&(t=e.stateNode,t!==null&&(delete t[wt],delete t[ar],delete t[Ru],delete t[kf],delete t[Ef])),e.stateNode=null,e.return=null,e.dependencies=null,e.memoizedProps=null,e.memoizedState=null,e.pendingProps=null,e.stateNode=null,e.updateQueue=null}function ya(e){return e.tag===5||e.tag===3||e.tag===4}function ga(e){e:for(;;){for(;e.sibling===null;){if(e.return===null||ya(e.return))return null;e=e.return}for(e.sibling.return=e.return,e=e.sibling;e.tag!==5&&e.tag!==6&&e.tag!==18;){if(e.flags&2||e.child===null||e.tag===4)continue e;e.child.return=e,e=e.child}if(!(e.flags&2))return e.stateNode}}function hi(e,t,n){var r=e.tag;if(r===5||r===6)e=e.stateNode,t?n.nodeType===8?n.parentNode.insertBefore(e,t):n.insertBefore(e,t):(n.nodeType===8?(t=n.parentNode,t.insertBefore(e,n)):(t=n,t.appendChild(e)),n=n._reactRootContainer,n!=null||t.onclick!==null||(t.onclick=qr));else if(r!==4&&(e=e.child,e!==null))for(hi(e,t,n),e=e.sibling;e!==null;)hi(e,t,n),e=e.sibling}function vi(e,t,n){var r=e.tag;if(r===5||r===6)e=e.stateNode,t?n.insertBefore(e,t):n.appendChild(e);else if(r!==4&&(e=e.child,e!==null))for(vi(e,t,n),e=e.sibling;e!==null;)vi(e,t,n),e=e.sibling}var Le=null,pt=!1;function Qt(e,t,n){for(n=n.child;n!==null;)wa(e,t,n),n=n.sibling}function wa(e,t,n){if(gt&&typeof gt.onCommitFiberUnmount=="function")try{gt.onCommitFiberUnmount(Ir,n)}catch{}switch(n.tag){case 5:Fe||Dn(n,t);case 6:var r=Le,l=pt;Le=null,Qt(e,t,n),Le=r,pt=l,Le!==null&&(pt?(e=Le,n=n.stateNode,e.nodeType===8?e.parentNode.removeChild(n):e.removeChild(n)):Le.removeChild(n.stateNode));break;case 18:Le!==null&&(pt?(e=Le,n=n.stateNode,e.nodeType===8?zu(e.parentNode,n):e.nodeType===1&&zu(e,n),qn(e)):zu(Le,n.stateNode));break;case 4:r=Le,l=pt,Le=n.stateNode.containerInfo,pt=!0,Qt(e,t,n),Le=r,pt=l;break;case 0:case 11:case 14:case 15:if(!Fe&&(r=n.updateQueue,r!==null&&(r=r.lastEffect,r!==null))){l=r=r.next;do{var u=l,i=u.destroy;u=u.tag,i!==void 0&&((u&2)!==0||(u&4)!==0)&&pi(n,t,i),l=l.next}while(l!==r)}Qt(e,t,n);break;case 1:if(!Fe&&(Dn(n,t),r=n.stateNode,typeof r.componentWillUnmount=="function"))try{r.props=n.memoizedProps,r.state=n.memoizedState,r.componentWillUnmount()}catch(o){we(n,t,o)}Qt(e,t,n);break;case 21:Qt(e,t,n);break;case 22:n.mode&1?(Fe=(r=Fe)||n.memoizedState!==null,Qt(e,t,n),Fe=r):Qt(e,t,n);break;default:Qt(e,t,n)}}function Sa(e){var t=e.updateQueue;if(t!==null){e.updateQueue=null;var n=e.stateNode;n===null&&(n=e.stateNode=new Af),t.forEach(function(r){var l=Gf.bind(null,e,r);n.has(r)||(n.add(r),r.then(l,l))})}}function mt(e,t){var n=t.deletions;if(n!==null)for(var r=0;r<n.length;r++){var l=n[r];try{var u=e,i=t,o=i;e:for(;o!==null;){switch(o.tag){case 5:Le=o.stateNode,pt=!1;break e;case 3:Le=o.stateNode.containerInfo,pt=!0;break e;case 4:Le=o.stateNode.containerInfo,pt=!0;break e}o=o.return}if(Le===null)throw Error(c(160));wa(u,i,l),Le=null,pt=!1;var s=l.alternate;s!==null&&(s.return=null),l.return=null}catch(h){we(l,t,h)}}if(t.subtreeFlags&12854)for(t=t.child;t!==null;)xa(t,e),t=t.sibling}function xa(e,t){var n=e.alternate,r=e.flags;switch(e.tag){case 0:case 11:case 14:case 15:if(mt(t,e),kt(e),r&4){try{Sr(3,e,e.return),xl(3,e)}catch(z){we(e,e.return,z)}try{Sr(5,e,e.return)}catch(z){we(e,e.return,z)}}break;case 1:mt(t,e),kt(e),r&512&&n!==null&&Dn(n,n.return);break;case 5:if(mt(t,e),kt(e),r&512&&n!==null&&Dn(n,n.return),e.flags&32){var l=e.stateNode;try{$n(l,"")}catch(z){we(e,e.return,z)}}if(r&4&&(l=e.stateNode,l!=null)){var u=e.memoizedProps,i=n!==null?n.memoizedProps:u,o=e.type,s=e.updateQueue;if(e.updateQueue=null,s!==null)try{o==="input"&&u.type==="radio"&&u.name!=null&&Xi(l,u),Yl(o,i);var h=Yl(o,u);for(i=0;i<s.length;i+=2){var S=s[i],x=s[i+1];S==="style"?no(l,x):S==="dangerouslySetInnerHTML"?eo(l,x):S==="children"?$n(l,x):ke(l,S,x,h)}switch(o){case"input":Bl(l,u);break;case"textarea":Ji(l,u);break;case"select":var g=l._wrapperState.wasMultiple;l._wrapperState.wasMultiple=!!u.multiple;var _=u.value;_!=null?mn(l,!!u.multiple,_,!1):g!==!!u.multiple&&(u.defaultValue!=null?mn(l,!!u.multiple,u.defaultValue,!0):mn(l,!!u.multiple,u.multiple?[]:"",!1))}l[ar]=u}catch(z){we(e,e.return,z)}}break;case 6:if(mt(t,e),kt(e),r&4){if(e.stateNode===null)throw Error(c(162));l=e.stateNode,u=e.memoizedProps;try{l.nodeValue=u}catch(z){we(e,e.return,z)}}break;case 3:if(mt(t,e),kt(e),r&4&&n!==null&&n.memoizedState.isDehydrated)try{qn(t.containerInfo)}catch(z){we(e,e.return,z)}break;case 4:mt(t,e),kt(e);break;case 13:mt(t,e),kt(e),l=e.child,l.flags&8192&&(u=l.memoizedState!==null,l.stateNode.isHidden=u,!u||l.alternate!==null&&l.alternate.memoizedState!==null||(wi=Se())),r&4&&Sa(e);break;case 22:if(S=n!==null&&n.memoizedState!==null,e.mode&1?(Fe=(h=Fe)||S,mt(t,e),Fe=h):mt(t,e),kt(e),r&8192){if(h=e.memoizedState!==null,(e.stateNode.isHidden=h)&&!S&&(e.mode&1)!==0)for(C=e,S=e.child;S!==null;){for(x=C=S;C!==null;){switch(g=C,_=g.child,g.tag){case 0:case 11:case 14:case 15:Sr(4,g,g.return);break;case 1:Dn(g,g.return);var j=g.stateNode;if(typeof j.componentWillUnmount=="function"){r=g,n=g.return;try{t=r,j.props=t.memoizedProps,j.state=t.memoizedState,j.componentWillUnmount()}catch(z){we(r,n,z)}}break;case 5:Dn(g,g.return);break;case 22:if(g.memoizedState!==null){_a(x);continue}}_!==null?(_.return=g,C=_):_a(x)}S=S.sibling}e:for(S=null,x=e;;){if(x.tag===5){if(S===null){S=x;try{l=x.stateNode,h?(u=l.style,typeof u.setProperty=="function"?u.setProperty("display","none","important"):u.display="none"):(o=x.stateNode,s=x.memoizedProps.style,i=s!=null&&s.hasOwnProperty("display")?s.display:null,o.style.display=to("display",i))}catch(z){we(e,e.return,z)}}}else if(x.tag===6){if(S===null)try{x.stateNode.nodeValue=h?"":x.memoizedProps}catch(z){we(e,e.return,z)}}else if((x.tag!==22&&x.tag!==23||x.memoizedState===null||x===e)&&x.child!==null){x.child.return=x,x=x.child;continue}if(x===e)break e;for(;x.sibling===null;){if(x.return===null||x.return===e)break e;S===x&&(S=null),x=x.return}S===x&&(S=null),x.sibling.return=x.return,x=x.sibling}}break;case 19:mt(t,e),kt(e),r&4&&Sa(e);break;case 21:break;default:mt(t,e),kt(e)}}function kt(e){var t=e.flags;if(t&2){try{e:{for(var n=e.return;n!==null;){if(ya(n)){var r=n;break e}n=n.return}throw Error(c(160))}switch(r.tag){case 5:var l=r.stateNode;r.flags&32&&($n(l,""),r.flags&=-33);var u=ga(e);vi(e,u,l);break;case 3:case 4:var i=r.stateNode.containerInfo,o=ga(e);hi(e,o,i);break;default:throw Error(c(161))}}catch(s){we(e,e.return,s)}e.flags&=-3}t&4096&&(e.flags&=-4097)}function $f(e,t,n){C=e,ka(e)}function ka(e,t,n){for(var r=(e.mode&1)!==0;C!==null;){var l=C,u=l.child;if(l.tag===22&&r){var i=l.memoizedState!==null||Sl;if(!i){var o=l.alternate,s=o!==null&&o.memoizedState!==null||Fe;o=Sl;var h=Fe;if(Sl=i,(Fe=s)&&!h)for(C=l;C!==null;)i=C,s=i.child,i.tag===22&&i.memoizedState!==null?Ca(l):s!==null?(s.return=i,C=s):Ca(l);for(;u!==null;)C=u,ka(u),u=u.sibling;C=l,Sl=o,Fe=h}Ea(e)}else(l.subtreeFlags&8772)!==0&&u!==null?(u.return=l,C=u):Ea(e)}}function Ea(e){for(;C!==null;){var t=C;if((t.flags&8772)!==0){var n=t.alternate;try{if((t.flags&8772)!==0)switch(t.tag){case 0:case 11:case 15:Fe||xl(5,t);break;case 1:var r=t.stateNode;if(t.flags&4&&!Fe)if(n===null)r.componentDidMount();else{var l=t.elementType===t.type?n.memoizedProps:dt(t.type,n.memoizedProps);r.componentDidUpdate(l,n.memoizedState,r.__reactInternalSnapshotBeforeUpdate)}var u=t.updateQueue;u!==null&&_s(t,u,r);break;case 3:var i=t.updateQueue;if(i!==null){if(n=null,t.child!==null)switch(t.child.tag){case 5:n=t.child.stateNode;break;case 1:n=t.child.stateNode}_s(t,i,n)}break;case 5:var o=t.stateNode;if(n===null&&t.flags&4){n=o;var s=t.memoizedProps;switch(t.type){case"button":case"input":case"select":case"textarea":s.autoFocus&&n.focus();break;case"img":s.src&&(n.src=s.src)}}break;case 6:break;case 4:break;case 12:break;case 13:if(t.memoizedState===null){var h=t.alternate;if(h!==null){var S=h.memoizedState;if(S!==null){var x=S.dehydrated;x!==null&&qn(x)}}}break;case 19:case 17:case 21:case 22:case 23:case 25:break;default:throw Error(c(163))}Fe||t.flags&512&&mi(t)}catch(g){we(t,t.return,g)}}if(t===e){C=null;break}if(n=t.sibling,n!==null){n.return=t.return,C=n;break}C=t.return}}function _a(e){for(;C!==null;){var t=C;if(t===e){C=null;break}var n=t.sibling;if(n!==null){n.return=t.return,C=n;break}C=t.return}}function Ca(e){for(;C!==null;){var t=C;try{switch(t.tag){case 0:case 11:case 15:var n=t.return;try{xl(4,t)}catch(s){we(t,n,s)}break;case 1:var r=t.stateNode;if(typeof r.componentDidMount=="function"){var l=t.return;try{r.componentDidMount()}catch(s){we(t,l,s)}}var u=t.return;try{mi(t)}catch(s){we(t,u,s)}break;case 5:var i=t.return;try{mi(t)}catch(s){we(t,i,s)}}}catch(s){we(t,t.return,s)}if(t===e){C=null;break}var o=t.sibling;if(o!==null){o.return=t.return,C=o;break}C=t.return}}var Bf=Math.ceil,kl=ie.ReactCurrentDispatcher,yi=ie.ReactCurrentOwner,ut=ie.ReactCurrentBatchConfig,J=0,Pe=null,Ee=null,Te=0,et=0,Mn=Vt(0),Ce=0,xr=null,sn=0,El=0,gi=0,kr=null,Ye=null,wi=0,Fn=1/0,Rt=null,_l=!1,Si=null,Kt=null,Cl=!1,Yt=null,Nl=0,Er=0,xi=null,Pl=-1,jl=0;function $e(){return(J&6)!==0?Se():Pl!==-1?Pl:Pl=Se()}function Xt(e){return(e.mode&1)===0?1:(J&2)!==0&&Te!==0?Te&-Te:Cf.transition!==null?(jl===0&&(jl=go()),jl):(e=ue,e!==0||(e=window.event,e=e===void 0?16:Po(e.type)),e)}function ht(e,t,n,r){if(50<Er)throw Er=0,xi=null,Error(c(185));Yn(e,n,r),((J&2)===0||e!==Pe)&&(e===Pe&&((J&2)===0&&(El|=n),Ce===4&&Gt(e,Te)),Xe(e,r),n===1&&J===0&&(t.mode&1)===0&&(Fn=Se()+500,nl&&Bt()))}function Xe(e,t){var n=e.callbackNode;_c(e,t);var r=Fr(e,e===Pe?Te:0);if(r===0)n!==null&&ho(n),e.callbackNode=null,e.callbackPriority=0;else if(t=r&-r,e.callbackPriority!==t){if(n!=null&&ho(n),t===1)e.tag===0?_f(Pa.bind(null,e)):ds(Pa.bind(null,e)),Sf(function(){(J&6)===0&&Bt()}),n=null;else{switch(wo(r)){case 1:n=eu;break;case 4:n=vo;break;case 16:n=Or;break;case 536870912:n=yo;break;default:n=Or}n=Da(n,Na.bind(null,e))}e.callbackPriority=t,e.callbackNode=n}}function Na(e,t){if(Pl=-1,jl=0,(J&6)!==0)throw Error(c(327));var n=e.callbackNode;if(Un()&&e.callbackNode!==n)return null;var r=Fr(e,e===Pe?Te:0);if(r===0)return null;if((r&30)!==0||(r&e.expiredLanes)!==0||t)t=zl(e,r);else{t=r;var l=J;J|=2;var u=za();(Pe!==e||Te!==t)&&(Rt=null,Fn=Se()+500,cn(e,t));do try{Qf();break}catch(o){ja(e,o)}while(!0);Au(),kl.current=u,J=l,Ee!==null?t=0:(Pe=null,Te=0,t=Ce)}if(t!==0){if(t===2&&(l=tu(e),l!==0&&(r=l,t=ki(e,l))),t===1)throw n=xr,cn(e,0),Gt(e,r),Xe(e,Se()),n;if(t===6)Gt(e,r);else{if(l=e.current.alternate,(r&30)===0&&!Wf(l)&&(t=zl(e,r),t===2&&(u=tu(e),u!==0&&(r=u,t=ki(e,u))),t===1))throw n=xr,cn(e,0),Gt(e,r),Xe(e,Se()),n;switch(e.finishedWork=l,e.finishedLanes=r,t){case 0:case 1:throw Error(c(345));case 2:fn(e,Ye,Rt);break;case 3:if(Gt(e,r),(r&130023424)===r&&(t=wi+500-Se(),10<t)){if(Fr(e,0)!==0)break;if(l=e.suspendedLanes,(l&r)!==r){$e(),e.pingedLanes|=e.suspendedLanes&l;break}e.timeoutHandle=ju(fn.bind(null,e,Ye,Rt),t);break}fn(e,Ye,Rt);break;case 4:if(Gt(e,r),(r&4194240)===r)break;for(t=e.eventTimes,l=-1;0<r;){var i=31-at(r);u=1<<i,i=t[i],i>l&&(l=i),r&=~u}if(r=l,r=Se()-r,r=(120>r?120:480>r?480:1080>r?1080:1920>r?1920:3e3>r?3e3:4320>r?4320:1960*Bf(r/1960))-r,10<r){e.timeoutHandle=ju(fn.bind(null,e,Ye,Rt),r);break}fn(e,Ye,Rt);break;case 5:fn(e,Ye,Rt);break;default:throw Error(c(329))}}}return Xe(e,Se()),e.callbackNode===n?Na.bind(null,e):null}function ki(e,t){var n=kr;return e.current.memoizedState.isDehydrated&&(cn(e,t).flags|=256),e=zl(e,t),e!==2&&(t=Ye,Ye=n,t!==null&&Ei(t)),e}function Ei(e){Ye===null?Ye=e:Ye.push.apply(Ye,e)}function Wf(e){for(var t=e;;){if(t.flags&16384){var n=t.updateQueue;if(n!==null&&(n=n.stores,n!==null))for(var r=0;r<n.length;r++){var l=n[r],u=l.getSnapshot;l=l.value;try{if(!ct(u(),l))return!1}catch{return!1}}}if(n=t.child,t.subtreeFlags&16384&&n!==null)n.return=t,t=n;else{if(t===e)break;for(;t.sibling===null;){if(t.return===null||t.return===e)return!0;t=t.return}t.sibling.return=t.return,t=t.sibling}}return!0}function Gt(e,t){for(t&=~gi,t&=~El,e.suspendedLanes|=t,e.pingedLanes&=~t,e=e.expirationTimes;0<t;){var n=31-at(t),r=1<<n;e[n]=-1,t&=~r}}function Pa(e){if((J&6)!==0)throw Error(c(327));Un();var t=Fr(e,0);if((t&1)===0)return Xe(e,Se()),null;var n=zl(e,t);if(e.tag!==0&&n===2){var r=tu(e);r!==0&&(t=r,n=ki(e,r))}if(n===1)throw n=xr,cn(e,0),Gt(e,t),Xe(e,Se()),n;if(n===6)throw Error(c(345));return e.finishedWork=e.current.alternate,e.finishedLanes=t,fn(e,Ye,Rt),Xe(e,Se()),null}function _i(e,t){var n=J;J|=1;try{return e(t)}finally{J=n,J===0&&(Fn=Se()+500,nl&&Bt())}}function an(e){Yt!==null&&Yt.tag===0&&(J&6)===0&&Un();var t=J;J|=1;var n=ut.transition,r=ue;try{if(ut.transition=null,ue=1,e)return e()}finally{ue=r,ut.transition=n,J=t,(J&6)===0&&Bt()}}function Ci(){et=Mn.current,de(Mn)}function cn(e,t){e.finishedWork=null,e.finishedLanes=0;var n=e.timeoutHandle;if(n!==-1&&(e.timeoutHandle=-1,wf(n)),Ee!==null)for(n=Ee.return;n!==null;){var r=n;switch(Iu(r),r.tag){case 1:r=r.type.childContextTypes,r!=null&&el();break;case 3:On(),de(He),de(Ie),Yu();break;case 5:Qu(r);break;case 4:On();break;case 13:de(he);break;case 19:de(he);break;case 10:Vu(r.type._context);break;case 22:case 23:Ci()}n=n.return}if(Pe=e,Ee=e=Zt(e.current,null),Te=et=t,Ce=0,xr=null,gi=El=sn=0,Ye=kr=null,ln!==null){for(t=0;t<ln.length;t++)if(n=ln[t],r=n.interleaved,r!==null){n.interleaved=null;var l=r.next,u=n.pending;if(u!==null){var i=u.next;u.next=l,r.next=i}n.pending=r}ln=null}return e}function ja(e,t){do{var n=Ee;try{if(Au(),dl.current=vl,pl){for(var r=ve.memoizedState;r!==null;){var l=r.queue;l!==null&&(l.pending=null),r=r.next}pl=!1}if(on=0,Ne=_e=ve=null,hr=!1,vr=0,yi.current=null,n===null||n.return===null){Ce=1,xr=t,Ee=null;break}e:{var u=e,i=n.return,o=n,s=t;if(t=Te,o.flags|=32768,s!==null&&typeof s=="object"&&typeof s.then=="function"){var h=s,S=o,x=S.tag;if((S.mode&1)===0&&(x===0||x===11||x===15)){var g=S.alternate;g?(S.updateQueue=g.updateQueue,S.memoizedState=g.memoizedState,S.lanes=g.lanes):(S.updateQueue=null,S.memoizedState=null)}var _=bs(i);if(_!==null){_.flags&=-257,ea(_,i,o,u,t),_.mode&1&&qs(u,h,t),t=_,s=h;var j=t.updateQueue;if(j===null){var z=new Set;z.add(s),t.updateQueue=z}else j.add(s);break e}else{if((t&1)===0){qs(u,h,t),Ni();break e}s=Error(c(426))}}else if(me&&o.mode&1){var xe=bs(i);if(xe!==null){(xe.flags&65536)===0&&(xe.flags|=256),ea(xe,i,o,u,t),Fu(In(s,o));break e}}u=s=In(s,o),Ce!==4&&(Ce=2),kr===null?kr=[u]:kr.push(u),u=i;do{switch(u.tag){case 3:u.flags|=65536,t&=-t,u.lanes|=t;var d=Zs(u,s,t);Es(u,d);break e;case 1:o=s;var a=u.type,p=u.stateNode;if((u.flags&128)===0&&(typeof a.getDerivedStateFromError=="function"||p!==null&&typeof p.componentDidCatch=="function"&&(Kt===null||!Kt.has(p)))){u.flags|=65536,t&=-t,u.lanes|=t;var k=Js(u,o,t);Es(u,k);break e}}u=u.return}while(u!==null)}La(n)}catch(L){t=L,Ee===n&&n!==null&&(Ee=n=n.return);continue}break}while(!0)}function za(){var e=kl.current;return kl.current=vl,e===null?vl:e}function Ni(){(Ce===0||Ce===3||Ce===2)&&(Ce=4),Pe===null||(sn&268435455)===0&&(El&268435455)===0||Gt(Pe,Te)}function zl(e,t){var n=J;J|=2;var r=za();(Pe!==e||Te!==t)&&(Rt=null,cn(e,t));do try{Hf();break}catch(l){ja(e,l)}while(!0);if(Au(),J=n,kl.current=r,Ee!==null)throw Error(c(261));return Pe=null,Te=0,Ce}function Hf(){for(;Ee!==null;)Ra(Ee)}function Qf(){for(;Ee!==null&&!hc();)Ra(Ee)}function Ra(e){var t=Ia(e.alternate,e,et);e.memoizedProps=e.pendingProps,t===null?La(e):Ee=t,yi.current=null}function La(e){var t=e;do{var n=t.alternate;if(e=t.return,(t.flags&32768)===0){if(n=Ff(n,t,et),n!==null){Ee=n;return}}else{if(n=Uf(n,t),n!==null){n.flags&=32767,Ee=n;return}if(e!==null)e.flags|=32768,e.subtreeFlags=0,e.deletions=null;else{Ce=6,Ee=null;return}}if(t=t.sibling,t!==null){Ee=t;return}Ee=t=e}while(t!==null);Ce===0&&(Ce=5)}function fn(e,t,n){var r=ue,l=ut.transition;try{ut.transition=null,ue=1,Kf(e,t,n,r)}finally{ut.transition=l,ue=r}return null}function Kf(e,t,n,r){do Un();while(Yt!==null);if((J&6)!==0)throw Error(c(327));n=e.finishedWork;var l=e.finishedLanes;if(n===null)return null;if(e.finishedWork=null,e.finishedLanes=0,n===e.current)throw Error(c(177));e.callbackNode=null,e.callbackPriority=0;var u=n.lanes|n.childLanes;if(Cc(e,u),e===Pe&&(Ee=Pe=null,Te=0),(n.subtreeFlags&2064)===0&&(n.flags&2064)===0||Cl||(Cl=!0,Da(Or,function(){return Un(),null})),u=(n.flags&15990)!==0,(n.subtreeFlags&15990)!==0||u){u=ut.transition,ut.transition=null;var i=ue;ue=1;var o=J;J|=4,yi.current=null,Vf(e,n),xa(n,e),df(Nu),Vr=!!Cu,Nu=Cu=null,e.current=n,$f(n),vc(),J=o,ue=i,ut.transition=u}else e.current=n;if(Cl&&(Cl=!1,Yt=e,Nl=l),u=e.pendingLanes,u===0&&(Kt=null),wc(n.stateNode),Xe(e,Se()),t!==null)for(r=e.onRecoverableError,n=0;n<t.length;n++)l=t[n],r(l.value,{componentStack:l.stack,digest:l.digest});if(_l)throw _l=!1,e=Si,Si=null,e;return(Nl&1)!==0&&e.tag!==0&&Un(),u=e.pendingLanes,(u&1)!==0?e===xi?Er++:(Er=0,xi=e):Er=0,Bt(),null}function Un(){if(Yt!==null){var e=wo(Nl),t=ut.transition,n=ue;try{if(ut.transition=null,ue=16>e?16:e,Yt===null)var r=!1;else{if(e=Yt,Yt=null,Nl=0,(J&6)!==0)throw Error(c(331));var l=J;for(J|=4,C=e.current;C!==null;){var u=C,i=u.child;if((C.flags&16)!==0){var o=u.deletions;if(o!==null){for(var s=0;s<o.length;s++){var h=o[s];for(C=h;C!==null;){var S=C;switch(S.tag){case 0:case 11:case 15:Sr(8,S,u)}var x=S.child;if(x!==null)x.return=S,C=x;else for(;C!==null;){S=C;var g=S.sibling,_=S.return;if(va(S),S===h){C=null;break}if(g!==null){g.return=_,C=g;break}C=_}}}var j=u.alternate;if(j!==null){var z=j.child;if(z!==null){j.child=null;do{var xe=z.sibling;z.sibling=null,z=xe}while(z!==null)}}C=u}}if((u.subtreeFlags&2064)!==0&&i!==null)i.return=u,C=i;else e:for(;C!==null;){if(u=C,(u.flags&2048)!==0)switch(u.tag){case 0:case 11:case 15:Sr(9,u,u.return)}var d=u.sibling;if(d!==null){d.return=u.return,C=d;break e}C=u.return}}var a=e.current;for(C=a;C!==null;){i=C;var p=i.child;if((i.subtreeFlags&2064)!==0&&p!==null)p.return=i,C=p;else e:for(i=a;C!==null;){if(o=C,(o.flags&2048)!==0)try{switch(o.tag){case 0:case 11:case 15:xl(9,o)}}catch(L){we(o,o.return,L)}if(o===i){C=null;break e}var k=o.sibling;if(k!==null){k.return=o.return,C=k;break e}C=o.return}}if(J=l,Bt(),gt&&typeof gt.onPostCommitFiberRoot=="function")try{gt.onPostCommitFiberRoot(Ir,e)}catch{}r=!0}return r}finally{ue=n,ut.transition=t}}return!1}function Ta(e,t,n){t=In(n,t),t=Zs(e,t,1),e=Ht(e,t,1),t=$e(),e!==null&&(Yn(e,1,t),Xe(e,t))}function we(e,t,n){if(e.tag===3)Ta(e,e,n);else for(;t!==null;){if(t.tag===3){Ta(t,e,n);break}else if(t.tag===1){var r=t.stateNode;if(typeof t.type.getDerivedStateFromError=="function"||typeof r.componentDidCatch=="function"&&(Kt===null||!Kt.has(r))){e=In(n,e),e=Js(t,e,1),t=Ht(t,e,1),e=$e(),t!==null&&(Yn(t,1,e),Xe(t,e));break}}t=t.return}}function Yf(e,t,n){var r=e.pingCache;r!==null&&r.delete(t),t=$e(),e.pingedLanes|=e.suspendedLanes&n,Pe===e&&(Te&n)===n&&(Ce===4||Ce===3&&(Te&130023424)===Te&&500>Se()-wi?cn(e,0):gi|=n),Xe(e,t)}function Oa(e,t){t===0&&((e.mode&1)===0?t=1:(t=Mr,Mr<<=1,(Mr&130023424)===0&&(Mr=4194304)));var n=$e();e=Pt(e,t),e!==null&&(Yn(e,t,n),Xe(e,n))}function Xf(e){var t=e.memoizedState,n=0;t!==null&&(n=t.retryLane),Oa(e,n)}function Gf(e,t){var n=0;switch(e.tag){case 13:var r=e.stateNode,l=e.memoizedState;l!==null&&(n=l.retryLane);break;case 19:r=e.stateNode;break;default:throw Error(c(314))}r!==null&&r.delete(t),Oa(e,n)}var Ia;Ia=function(e,t,n){if(e!==null)if(e.memoizedProps!==t.pendingProps||He.current)Ke=!0;else{if((e.lanes&n)===0&&(t.flags&128)===0)return Ke=!1,Mf(e,t,n);Ke=(e.flags&131072)!==0}else Ke=!1,me&&(t.flags&1048576)!==0&&ps(t,ll,t.index);switch(t.lanes=0,t.tag){case 2:var r=t.type;wl(e,t),e=t.pendingProps;var l=Nn(t,Ie.current);Tn(t,n),l=Zu(null,t,r,e,l,n);var u=Ju();return t.flags|=1,typeof l=="object"&&l!==null&&typeof l.render=="function"&&l.$$typeof===void 0?(t.tag=1,t.memoizedState=null,t.updateQueue=null,Qe(r)?(u=!0,tl(t)):u=!1,t.memoizedState=l.state!==null&&l.state!==void 0?l.state:null,Wu(t),l.updater=yl,t.stateNode=l,l._reactInternals=t,ri(t,r,e,n),t=oi(null,t,r,!0,u,n)):(t.tag=0,me&&u&&Ou(t),Ve(null,t,l,n),t=t.child),t;case 16:r=t.elementType;e:{switch(wl(e,t),e=t.pendingProps,l=r._init,r=l(r._payload),t.type=r,l=t.tag=Jf(r),e=dt(r,e),l){case 0:t=ii(null,t,r,e,n);break e;case 1:t=ia(null,t,r,e,n);break e;case 11:t=ta(null,t,r,e,n);break e;case 14:t=na(null,t,r,dt(r.type,e),n);break e}throw Error(c(306,r,""))}return t;case 0:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:dt(r,l),ii(e,t,r,l,n);case 1:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:dt(r,l),ia(e,t,r,l,n);case 3:e:{if(oa(t),e===null)throw Error(c(387));r=t.pendingProps,u=t.memoizedState,l=u.element,ks(e,t),cl(t,r,null,n);var i=t.memoizedState;if(r=i.element,u.isDehydrated)if(u={element:r,isDehydrated:!1,cache:i.cache,pendingSuspenseBoundaries:i.pendingSuspenseBoundaries,transitions:i.transitions},t.updateQueue.baseState=u,t.memoizedState=u,t.flags&256){l=In(Error(c(423)),t),t=sa(e,t,r,n,l);break e}else if(r!==l){l=In(Error(c(424)),t),t=sa(e,t,r,n,l);break e}else for(be=At(t.stateNode.containerInfo.firstChild),qe=t,me=!0,ft=null,n=Ss(t,null,r,n),t.child=n;n;)n.flags=n.flags&-3|4096,n=n.sibling;else{if(zn(),r===l){t=zt(e,t,n);break e}Ve(e,t,r,n)}t=t.child}return t;case 5:return Cs(t),e===null&&Mu(t),r=t.type,l=t.pendingProps,u=e!==null?e.memoizedProps:null,i=l.children,Pu(r,l)?i=null:u!==null&&Pu(r,u)&&(t.flags|=32),ua(e,t),Ve(e,t,i,n),t.child;case 6:return e===null&&Mu(t),null;case 13:return aa(e,t,n);case 4:return Hu(t,t.stateNode.containerInfo),r=t.pendingProps,e===null?t.child=Rn(t,null,r,n):Ve(e,t,r,n),t.child;case 11:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:dt(r,l),ta(e,t,r,l,n);case 7:return Ve(e,t,t.pendingProps,n),t.child;case 8:return Ve(e,t,t.pendingProps.children,n),t.child;case 12:return Ve(e,t,t.pendingProps.children,n),t.child;case 10:e:{if(r=t.type._context,l=t.pendingProps,u=t.memoizedProps,i=l.value,ce(ol,r._currentValue),r._currentValue=i,u!==null)if(ct(u.value,i)){if(u.children===l.children&&!He.current){t=zt(e,t,n);break e}}else for(u=t.child,u!==null&&(u.return=t);u!==null;){var o=u.dependencies;if(o!==null){i=u.child;for(var s=o.firstContext;s!==null;){if(s.context===r){if(u.tag===1){s=jt(-1,n&-n),s.tag=2;var h=u.updateQueue;if(h!==null){h=h.shared;var S=h.pending;S===null?s.next=s:(s.next=S.next,S.next=s),h.pending=s}}u.lanes|=n,s=u.alternate,s!==null&&(s.lanes|=n),$u(u.return,n,t),o.lanes|=n;break}s=s.next}}else if(u.tag===10)i=u.type===t.type?null:u.child;else if(u.tag===18){if(i=u.return,i===null)throw Error(c(341));i.lanes|=n,o=i.alternate,o!==null&&(o.lanes|=n),$u(i,n,t),i=u.sibling}else i=u.child;if(i!==null)i.return=u;else for(i=u;i!==null;){if(i===t){i=null;break}if(u=i.sibling,u!==null){u.return=i.return,i=u;break}i=i.return}u=i}Ve(e,t,l.children,n),t=t.child}return t;case 9:return l=t.type,r=t.pendingProps.children,Tn(t,n),l=rt(l),r=r(l),t.flags|=1,Ve(e,t,r,n),t.child;case 14:return r=t.type,l=dt(r,t.pendingProps),l=dt(r.type,l),na(e,t,r,l,n);case 15:return ra(e,t,t.type,t.pendingProps,n);case 17:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:dt(r,l),wl(e,t),t.tag=1,Qe(r)?(e=!0,tl(t)):e=!1,Tn(t,n),Xs(t,r,l),ri(t,r,l,n),oi(null,t,r,!0,e,n);case 19:return fa(e,t,n);case 22:return la(e,t,n)}throw Error(c(156,t.tag))};function Da(e,t){return mo(e,t)}function Zf(e,t,n,r){this.tag=e,this.key=n,this.sibling=this.child=this.return=this.stateNode=this.type=this.elementType=null,this.index=0,this.ref=null,this.pendingProps=t,this.dependencies=this.memoizedState=this.updateQueue=this.memoizedProps=null,this.mode=r,this.subtreeFlags=this.flags=0,this.deletions=null,this.childLanes=this.lanes=0,this.alternate=null}function it(e,t,n,r){return new Zf(e,t,n,r)}function Pi(e){return e=e.prototype,!(!e||!e.isReactComponent)}function Jf(e){if(typeof e=="function")return Pi(e)?1:0;if(e!=null){if(e=e.$$typeof,e===Be)return 11;if(e===yt)return 14}return 2}function Zt(e,t){var n=e.alternate;return n===null?(n=it(e.tag,t,e.key,e.mode),n.elementType=e.elementType,n.type=e.type,n.stateNode=e.stateNode,n.alternate=e,e.alternate=n):(n.pendingProps=t,n.type=e.type,n.flags=0,n.subtreeFlags=0,n.deletions=null),n.flags=e.flags&14680064,n.childLanes=e.childLanes,n.lanes=e.lanes,n.child=e.child,n.memoizedProps=e.memoizedProps,n.memoizedState=e.memoizedState,n.updateQueue=e.updateQueue,t=e.dependencies,n.dependencies=t===null?null:{lanes:t.lanes,firstContext:t.firstContext},n.sibling=e.sibling,n.index=e.index,n.ref=e.ref,n}function Rl(e,t,n,r,l,u){var i=2;if(r=e,typeof e=="function")Pi(e)&&(i=1);else if(typeof e=="string")i=5;else e:switch(e){case se:return dn(n.children,l,u,t);case ze:i=8,l|=8;break;case vt:return e=it(12,n,t,l|2),e.elementType=vt,e.lanes=u,e;case Ae:return e=it(13,n,t,l),e.elementType=Ae,e.lanes=u,e;case Re:return e=it(19,n,t,l),e.elementType=Re,e.lanes=u,e;case ge:return Ll(n,l,u,t);default:if(typeof e=="object"&&e!==null)switch(e.$$typeof){case st:i=10;break e;case M:i=9;break e;case Be:i=11;break e;case yt:i=14;break e;case We:i=16,r=null;break e}throw Error(c(130,e==null?e:typeof e,""))}return t=it(i,n,t,l),t.elementType=e,t.type=r,t.lanes=u,t}function dn(e,t,n,r){return e=it(7,e,r,t),e.lanes=n,e}function Ll(e,t,n,r){return e=it(22,e,r,t),e.elementType=ge,e.lanes=n,e.stateNode={isHidden:!1},e}function ji(e,t,n){return e=it(6,e,null,t),e.lanes=n,e}function zi(e,t,n){return t=it(4,e.children!==null?e.children:[],e.key,t),t.lanes=n,t.stateNode={containerInfo:e.containerInfo,pendingChildren:null,implementation:e.implementation},t}function qf(e,t,n,r,l){this.tag=t,this.containerInfo=e,this.finishedWork=this.pingCache=this.current=this.pendingChildren=null,this.timeoutHandle=-1,this.callbackNode=this.pendingContext=this.context=null,this.callbackPriority=0,this.eventTimes=nu(0),this.expirationTimes=nu(-1),this.entangledLanes=this.finishedLanes=this.mutableReadLanes=this.expiredLanes=this.pingedLanes=this.suspendedLanes=this.pendingLanes=0,this.entanglements=nu(0),this.identifierPrefix=r,this.onRecoverableError=l,this.mutableSourceEagerHydrationData=null}function Ri(e,t,n,r,l,u,i,o,s){return e=new qf(e,t,n,o,s),t===1?(t=1,u===!0&&(t|=8)):t=0,u=it(3,null,null,t),e.current=u,u.stateNode=e,u.memoizedState={element:r,isDehydrated:n,cache:null,transitions:null,pendingSuspenseBoundaries:null},Wu(u),e}function bf(e,t,n){var r=3<arguments.length&&arguments[3]!==void 0?arguments[3]:null;return{$$typeof:re,key:r==null?null:""+r,children:e,containerInfo:t,implementation:n}}function Ma(e){if(!e)return $t;e=e._reactInternals;e:{if(bt(e)!==e||e.tag!==1)throw Error(c(170));var t=e;do{switch(t.tag){case 3:t=t.stateNode.context;break e;case 1:if(Qe(t.type)){t=t.stateNode.__reactInternalMemoizedMergedChildContext;break e}}t=t.return}while(t!==null);throw Error(c(171))}if(e.tag===1){var n=e.type;if(Qe(n))return cs(e,n,t)}return t}function Fa(e,t,n,r,l,u,i,o,s){return e=Ri(n,r,!0,e,l,u,i,o,s),e.context=Ma(null),n=e.current,r=$e(),l=Xt(n),u=jt(r,l),u.callback=t??null,Ht(n,u,l),e.current.lanes=l,Yn(e,l,r),Xe(e,r),e}function Tl(e,t,n,r){var l=t.current,u=$e(),i=Xt(l);return n=Ma(n),t.context===null?t.context=n:t.pendingContext=n,t=jt(u,i),t.payload={element:e},r=r===void 0?null:r,r!==null&&(t.callback=r),e=Ht(l,t,i),e!==null&&(ht(e,l,i,u),al(e,l,i)),i}function Ol(e){if(e=e.current,!e.child)return null;switch(e.child.tag){case 5:return e.child.stateNode;default:return e.child.stateNode}}function Ua(e,t){if(e=e.memoizedState,e!==null&&e.dehydrated!==null){var n=e.retryLane;e.retryLane=n!==0&&n<t?n:t}}function Li(e,t){Ua(e,t),(e=e.alternate)&&Ua(e,t)}function ed(){return null}var Aa=typeof reportError=="function"?reportError:function(e){console.error(e)};function Ti(e){this._internalRoot=e}Il.prototype.render=Ti.prototype.render=function(e){var t=this._internalRoot;if(t===null)throw Error(c(409));Tl(e,t,null,null)},Il.prototype.unmount=Ti.prototype.unmount=function(){var e=this._internalRoot;if(e!==null){this._internalRoot=null;var t=e.containerInfo;an(function(){Tl(null,e,null,null)}),t[Et]=null}};function Il(e){this._internalRoot=e}Il.prototype.unstable_scheduleHydration=function(e){if(e){var t=ko();e={blockedOn:null,target:e,priority:t};for(var n=0;n<Mt.length&&t!==0&&t<Mt[n].priority;n++);Mt.splice(n,0,e),n===0&&Co(e)}};function Oi(e){return!(!e||e.nodeType!==1&&e.nodeType!==9&&e.nodeType!==11)}function Dl(e){return!(!e||e.nodeType!==1&&e.nodeType!==9&&e.nodeType!==11&&(e.nodeType!==8||e.nodeValue!==" react-mount-point-unstable "))}function Va(){}function td(e,t,n,r,l){if(l){if(typeof r=="function"){var u=r;r=function(){var h=Ol(i);u.call(h)}}var i=Fa(t,r,e,0,null,!1,!1,"",Va);return e._reactRootContainer=i,e[Et]=i.current,or(e.nodeType===8?e.parentNode:e),an(),i}for(;l=e.lastChild;)e.removeChild(l);if(typeof r=="function"){var o=r;r=function(){var h=Ol(s);o.call(h)}}var s=Ri(e,0,!1,null,null,!1,!1,"",Va);return e._reactRootContainer=s,e[Et]=s.current,or(e.nodeType===8?e.parentNode:e),an(function(){Tl(t,s,n,r)}),s}function Ml(e,t,n,r,l){var u=n._reactRootContainer;if(u){var i=u;if(typeof l=="function"){var o=l;l=function(){var s=Ol(i);o.call(s)}}Tl(t,i,e,l)}else i=td(n,t,e,l,r);return Ol(i)}So=function(e){switch(e.tag){case 3:var t=e.stateNode;if(t.current.memoizedState.isDehydrated){var n=Kn(t.pendingLanes);n!==0&&(ru(t,n|1),Xe(t,Se()),(J&6)===0&&(Fn=Se()+500,Bt()))}break;case 13:an(function(){var r=Pt(e,1);if(r!==null){var l=$e();ht(r,e,1,l)}}),Li(e,1)}},lu=function(e){if(e.tag===13){var t=Pt(e,134217728);if(t!==null){var n=$e();ht(t,e,134217728,n)}Li(e,134217728)}},xo=function(e){if(e.tag===13){var t=Xt(e),n=Pt(e,t);if(n!==null){var r=$e();ht(n,e,t,r)}Li(e,t)}},ko=function(){return ue},Eo=function(e,t){var n=ue;try{return ue=e,t()}finally{ue=n}},Zl=function(e,t,n){switch(t){case"input":if(Bl(e,n),t=n.name,n.type==="radio"&&t!=null){for(n=e;n.parentNode;)n=n.parentNode;for(n=n.querySelectorAll("input[name="+JSON.stringify(""+t)+'][type="radio"]'),t=0;t<n.length;t++){var r=n[t];if(r!==e&&r.form===e.form){var l=br(r);if(!l)throw Error(c(90));Ki(r),Bl(r,l)}}}break;case"textarea":Ji(e,n);break;case"select":t=n.value,t!=null&&mn(e,!!n.multiple,t,!1)}},io=_i,oo=an;var nd={usingClientEntryPoint:!1,Events:[cr,_n,br,lo,uo,_i]},_r={findFiberByHostInstance:en,bundleType:0,version:"18.3.1",rendererPackageName:"react-dom"},rd={bundleType:_r.bundleType,version:_r.version,rendererPackageName:_r.rendererPackageName,rendererConfig:_r.rendererConfig,overrideHookState:null,overrideHookStateDeletePath:null,overrideHookStateRenamePath:null,overrideProps:null,overridePropsDeletePath:null,overridePropsRenamePath:null,setErrorHandler:null,setSuspenseHandler:null,scheduleUpdate:null,currentDispatcherRef:ie.ReactCurrentDispatcher,findHostInstanceByFiber:function(e){return e=fo(e),e===null?null:e.stateNode},findFiberByHostInstance:_r.findFiberByHostInstance||ed,findHostInstancesForRefresh:null,scheduleRefresh:null,scheduleRoot:null,setRefreshHandler:null,getCurrentFiber:null,reconcilerVersion:"18.3.1-next-f1338f8080-20240426"};if(typeof __REACT_DEVTOOLS_GLOBAL_HOOK__<"u"){var Fl=__REACT_DEVTOOLS_GLOBAL_HOOK__;if(!Fl.isDisabled&&Fl.supportsFiber)try{Ir=Fl.inject(rd),gt=Fl}catch{}}return Ge.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED=nd,Ge.createPortal=function(e,t){var n=2<arguments.length&&arguments[2]!==void 0?arguments[2]:null;if(!Oi(t))throw Error(c(200));return bf(e,t,null,n)},Ge.createRoot=function(e,t){if(!Oi(e))throw Error(c(299));var n=!1,r="",l=Aa;return t!=null&&(t.unstable_strictMode===!0&&(n=!0),t.identifierPrefix!==void 0&&(r=t.identifierPrefix),t.onRecoverableError!==void 0&&(l=t.onRecoverableError)),t=Ri(e,1,!1,null,null,n,!1,r,l),e[Et]=t.current,or(e.nodeType===8?e.parentNode:e),new Ti(t)},Ge.findDOMNode=function(e){if(e==null)return null;if(e.nodeType===1)return e;var t=e._reactInternals;if(t===void 0)throw typeof e.render=="function"?Error(c(188)):(e=Object.keys(e).join(","),Error(c(268,e)));return e=fo(t),e=e===null?null:e.stateNode,e},Ge.flushSync=function(e){return an(e)},Ge.hydrate=function(e,t,n){if(!Dl(t))throw Error(c(200));return Ml(null,e,t,!0,n)},Ge.hydrateRoot=function(e,t,n){if(!Oi(e))throw Error(c(405));var r=n!=null&&n.hydratedSources||null,l=!1,u="",i=Aa;if(n!=null&&(n.unstable_strictMode===!0&&(l=!0),n.identifierPrefix!==void 0&&(u=n.identifierPrefix),n.onRecoverableError!==void 0&&(i=n.onRecoverableError)),t=Fa(t,null,e,1,n??null,l,!1,u,i),e[Et]=t.current,or(e),r)for(e=0;e<r.length;e++)n=r[e],l=n._getVersion,l=l(n._source),t.mutableSourceEagerHydrationData==null?t.mutableSourceEagerHydrationData=[n,l]:t.mutableSourceEagerHydrationData.push(n,l);return new Il(t)},Ge.render=function(e,t,n){if(!Dl(t))throw Error(c(200));return Ml(null,e,t,!1,n)},Ge.unmountComponentAtNode=function(e){if(!Dl(e))throw Error(c(40));return e._reactRootContainer?(an(function(){Ml(null,null,e,!1,function(){e._reactRootContainer=null,e[Et]=null})}),!0):!1},Ge.unstable_batchedUpdates=_i,Ge.unstable_renderSubtreeIntoContainer=function(e,t,n,r){if(!Dl(n))throw Error(c(200));if(e==null||e._reactInternals===void 0)throw Error(c(38));return Ml(e,t,n,!1,r)},Ge.version="18.3.1-next-f1338f8080-20240426",Ge}var Xa;function fd(){if(Xa)return Mi.exports;Xa=1;function m(){if(!(typeof __REACT_DEVTOOLS_GLOBAL_HOOK__>"u"||typeof __REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE!="function"))try{__REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE(m)}catch(y){console.error(y)}}return m(),Mi.exports=cd(),Mi.exports}var Ga;function dd(){if(Ga)return Ul;Ga=1;var m=fd();return Ul.createRoot=m.createRoot,Ul.hydrateRoot=m.hydrateRoot,Ul}var pd=dd();const md=Qi(pd);function Ue(m,y,c,P){if(typeof y=="function"?m!==y||!P:!y.has(m))throw new TypeError("Cannot read private member from an object whose class did not declare it");return c==="m"?P:c==="a"?P.call(m):P?P.value:y.get(m)}function Nr(m,y,c,P,R){if(typeof y=="function"?m!==y||!0:!y.has(m))throw new TypeError("Cannot write private member to an object whose class did not declare it");return y.set(m,c),c}var pn,ot,An,Al;const Za="__TAURI_TO_IPC_KEY__";function hd(m,y=!1){return window.__TAURI_INTERNALS__.transformCallback(m,y)}class Ja{constructor(y){pn.set(this,void 0),ot.set(this,0),An.set(this,[]),Al.set(this,void 0),Nr(this,pn,y||(()=>{})),this.id=hd(c=>{const P=c.index;if("end"in c){P==Ue(this,ot,"f")?this.cleanupCallback():Nr(this,Al,P);return}const R=c.message;if(P==Ue(this,ot,"f")){for(Ue(this,pn,"f").call(this,R),Nr(this,ot,Ue(this,ot,"f")+1);Ue(this,ot,"f")in Ue(this,An,"f");){const V=Ue(this,An,"f")[Ue(this,ot,"f")];Ue(this,pn,"f").call(this,V),delete Ue(this,An,"f")[Ue(this,ot,"f")],Nr(this,ot,Ue(this,ot,"f")+1)}Ue(this,ot,"f")===Ue(this,Al,"f")&&this.cleanupCallback()}else Ue(this,An,"f")[P]=R})}cleanupCallback(){window.__TAURI_INTERNALS__.unregisterCallback(this.id)}set onmessage(y){Nr(this,pn,y)}get onmessage(){return Ue(this,pn,"f")}[(pn=new WeakMap,ot=new WeakMap,An=new WeakMap,Al=new WeakMap,Za)](){return`__CHANNEL__:${this.id}`}toJSON(){return this[Za]()}}async function vd(m,y={},c){return window.__TAURI_INTERNALS__.invoke(m,y,c)}const yd="plugin:pytauri|pyfunc",gd="pyfunc";async function Lt(m,y,c){let P;if(y===void 0||y instanceof ArrayBuffer||y instanceof Uint8Array)P=y;else{const V=JSON.stringify(y);P=new TextEncoder().encode(V)}const R=new Headers(c==null?void 0:c.headers);return R.set(gd,m),await vd(yd,P,{...c,headers:R})}const qt={detectProviders:()=>Lt("detect_providers",{}),isRunning:()=>Lt("is_running",{}),startRun:(m,y,c)=>Lt("start_run",{config:m,progressChannel:y,logChannel:c}),stopRun:()=>Lt("stop_run",{}),shutdownBrowser:()=>Lt("shutdown_browser",{}),getEnvFields:()=>Lt("get_env_fields",{}),writeEnvFields:m=>Lt("write_env_fields",{updates:m}),getLetters:(m=200)=>Lt("get_letters",{limit:m}),getTelemetrySummary:()=>Lt("get_telemetry_summary",{})},wd={},qa=m=>{let y;const c=new Set,P=(B,O)=>{const T=typeof B=="function"?B(y):B;if(!Object.is(T,y)){const Y=y;y=O??(typeof T!="object"||T===null)?T:Object.assign({},y,T),c.forEach(G=>G(y,Y))}},R=()=>y,$={setState:P,getState:R,getInitialState:()=>H,subscribe:B=>(c.add(B),()=>c.delete(B)),destroy:()=>{(wd?"production":void 0)!=="production"&&console.warn("[DEPRECATED] The `destroy` method will be unsupported in a future version. Instead use unsubscribe function returned by subscribe. Everything will be garbage-collected if store is garbage-collected."),c.clear()}},H=y=m(P,R,$);return $},Sd=m=>m?qa(m):qa;var Ai={exports:{}},Vi={},$i={exports:{}},Bi={};/**
+ * @license React
+ * use-sync-external-store-shim.production.js
+ *
+ * Copyright (c) Meta Platforms, Inc. and affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */var ba;function xd(){if(ba)return Bi;ba=1;var m=Pr();function y(O,T){return O===T&&(O!==0||1/O===1/T)||O!==O&&T!==T}var c=typeof Object.is=="function"?Object.is:y,P=m.useState,R=m.useEffect,V=m.useLayoutEffect,q=m.useDebugValue;function ne(O,T){var Y=T(),G=P({inst:{value:Y,getSnapshot:T}}),A=G[0].inst,W=G[1];return V(function(){A.value=Y,A.getSnapshot=T,$(A)&&W({inst:A})},[O,Y,T]),R(function(){return $(A)&&W({inst:A}),O(function(){$(A)&&W({inst:A})})},[O]),q(Y),Y}function $(O){var T=O.getSnapshot;O=O.value;try{var Y=T();return!c(O,Y)}catch{return!0}}function H(O,T){return T()}var B=typeof window>"u"||typeof window.document>"u"||typeof window.document.createElement>"u"?H:ne;return Bi.useSyncExternalStore=m.useSyncExternalStore!==void 0?m.useSyncExternalStore:B,Bi}var ec;function kd(){return ec||(ec=1,$i.exports=xd()),$i.exports}/**
+ * @license React
+ * use-sync-external-store-shim/with-selector.production.js
+ *
+ * Copyright (c) Meta Platforms, Inc. and affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */var tc;function Ed(){if(tc)return Vi;tc=1;var m=Pr(),y=kd();function c(H,B){return H===B&&(H!==0||1/H===1/B)||H!==H&&B!==B}var P=typeof Object.is=="function"?Object.is:c,R=y.useSyncExternalStore,V=m.useRef,q=m.useEffect,ne=m.useMemo,$=m.useDebugValue;return Vi.useSyncExternalStoreWithSelector=function(H,B,O,T,Y){var G=V(null);if(G.current===null){var A={hasValue:!1,value:null};G.current=A}else A=G.current;G=ne(function(){function Z(re){if(!ye){if(ye=!0,ke=re,re=T(re),Y!==void 0&&A.hasValue){var se=A.value;if(Y(se,re))return ie=se}return ie=re}if(se=ie,P(ke,re))return se;var ze=T(re);return Y!==void 0&&Y(se,ze)?(ke=re,se):(ke=re,ie=ze)}var ye=!1,ke,ie,Oe=O===void 0?null:O;return[function(){return Z(B())},Oe===null?void 0:function(){return Z(Oe())}]},[B,O,T,Y]);var W=R(H,G[0],G[1]);return q(function(){A.hasValue=!0,A.value=W},[W]),$(W),W},Vi}var nc;function _d(){return nc||(nc=1,Ai.exports=Ed()),Ai.exports}var Cd=_d();const Nd=Qi(Cd),oc={},{useDebugValue:Pd}=ic,{useSyncExternalStoreWithSelector:jd}=Nd;let rc=!1;const zd=m=>m;function Rd(m,y=zd,c){(oc?"production":void 0)!=="production"&&c&&!rc&&(console.warn("[DEPRECATED] Use `createWithEqualityFn` instead of `create` or use `useStoreWithEqualityFn` instead of `useStore`. They can be imported from 'zustand/traditional'. https://github.com/pmndrs/zustand/discussions/1937"),rc=!0);const P=jd(m.subscribe,m.getState,m.getServerState||m.getInitialState,y,c);return Pd(P),P}const lc=m=>{(oc?"production":void 0)!=="production"&&typeof m!="function"&&console.warn("[DEPRECATED] Passing a vanilla store will be unsupported in a future version. Instead use `import { useStore } from 'zustand'`.");const y=typeof m=="function"?Sd(m):m,c=(P,R)=>Rd(y,P,R);return Object.assign(c,y),c},Ld=m=>m?lc(m):lc,Tt=Ld(m=>({running:!1,events:[],logs:[],currentIndex:null,setRunning:y=>m({running:y}),pushEvent:y=>m(c=>{const P=[...c.events,y].slice(-500);let R=c.currentIndex;y.kind==="job_found"&&typeof y.payload.index=="number"&&(R=y.payload.index);let V=c.running;return y.kind==="loop_ended"&&(V=!1),{events:P,currentIndex:R,running:V}}),pushLog:y=>m(c=>({logs:[...c.logs,y].slice(-1e3)})),clear:()=>m({events:[],logs:[],currentIndex:null})}));function Td(){const[m,y]=oe.useState([]),[c,P]=oe.useState(null),[R,V]=oe.useState(""),[q,ne]=oe.useState(""),[$,H]=oe.useState(""),[B,O]=oe.useState(!0),[T,Y]=oe.useState(!1),G=Tt(M=>M.running),A=Tt(M=>M.events),W=Tt(M=>M.logs),Z=Tt(M=>M.currentIndex),ye=Tt(M=>M.setRunning),ke=Tt(M=>M.pushEvent),ie=Tt(M=>M.pushLog),Oe=Tt(M=>M.clear),re=oe.useRef(null),se=oe.useRef(null);oe.useEffect(()=>{qt.detectProviders().then(({providers:M})=>{y(M),M.length>0&&H(M[0])}).catch(M=>P(String(M)))},[]),oe.useEffect(()=>{re.current&&(re.current.scrollTop=re.current.scrollHeight)},[W.length]),oe.useEffect(()=>{se.current&&(se.current.scrollTop=se.current.scrollHeight)},[A.length]);async function ze(){if(!R.trim()){alert("请填用户名");return}if(!$){alert("没有可用的 provider，先去配置 tab 填 API key");return}Y(!0),Oe(),ye(!0);const M=new Ja(Re=>ke(Re)),Be=new Ja(Re=>ie(Re)),Ae={usrName:R.trim(),label:q.trim(),provider:$,dryRun:B};try{await qt.startRun(Ae,M,Be)}catch(Re){ie(`[start_run 失败] ${Re}`),ye(!1)}finally{Y(!1)}}async function vt(){Y(!0);try{await qt.stopRun()}catch(M){ie(`[stop_run 失败] ${M}`)}finally{Y(!1)}}async function st(){Y(!0);try{await qt.shutdownBrowser(),ie("[已关 Chrome，下次启动会重新打开]")}catch(M){ie(`[shutdown_browser 失败] ${M}`)}finally{Y(!1)}}return v.jsxs("div",{className:"max-w-5xl mx-auto space-y-4",children:[v.jsxs("div",{className:"bg-white p-4 rounded-lg shadow-sm border border-slate-200",children:[v.jsxs("div",{className:"grid grid-cols-2 gap-4",children:[v.jsx(Vl,{label:"你的名字（招呼语署名）",children:v.jsx("input",{type:"text",value:R,onChange:M=>V(M.target.value),disabled:G,className:"input",placeholder:"必填"})}),v.jsx(Vl,{label:"求职 tag",children:v.jsx("input",{type:"text",value:q,onChange:M=>ne(M.target.value),disabled:G,className:"input",placeholder:"留空走 BOSS 推荐 feed"})}),v.jsx(Vl,{label:"Provider",children:c?v.jsxs("span",{className:"text-red-600 text-sm",children:["检测失败: ",c]}):m.length===0?v.jsx("span",{className:"text-amber-600 text-sm",children:"没检测到 API key —— 去配置 tab 填一个"}):v.jsx("select",{value:$,onChange:M=>H(M.target.value),disabled:G,className:"input",children:m.map(M=>v.jsx("option",{value:M,children:M},M))})}),v.jsx(Vl,{label:"",children:v.jsxs("label",{className:"inline-flex items-center gap-2 text-sm",children:[v.jsx("input",{type:"checkbox",checked:B,onChange:M=>O(M.target.checked),disabled:G}),"DRY-RUN（只生成不发送）"]})})]}),v.jsxs("div",{className:"flex items-center gap-3 mt-4",children:[v.jsx("button",{onClick:ze,disabled:G||T,className:"btn-primary",children:"开始"}),v.jsx("button",{onClick:vt,disabled:!G||T,className:"btn-danger",children:"停止"}),v.jsx("button",{onClick:st,disabled:G||T,className:"btn-secondary",children:"重置 Chrome"}),v.jsx("span",{className:"ml-auto text-sm text-slate-600",children:G?Z!==null?`运行中 · 当前 job #${Z}`:"运行中...":"idle"})]})]}),v.jsxs("div",{className:"grid grid-cols-1 lg:grid-cols-2 gap-4",children:[v.jsx(uc,{title:"进度事件",children:v.jsx("div",{ref:se,className:"bg-slate-50 rounded p-3 text-xs font-mono h-72 overflow-y-auto",children:A.length===0?v.jsx("div",{className:"text-slate-400",children:"还没事件"}):A.map((M,Be)=>v.jsx(Id,{ev:M},Be))})}),v.jsx(uc,{title:"日志",children:v.jsx("pre",{ref:re,className:"bg-slate-900 text-slate-100 rounded p-3 text-xs font-mono h-72 overflow-y-auto whitespace-pre-wrap",children:W.length===0?v.jsx("span",{className:"text-slate-500",children:"还没日志"}):W.join(`
+`)})})]}),v.jsx("style",{children:`
+        .input {
+          padding: 0.4rem 0.6rem; font-size: 14px;
+          border: 1px solid #cbd5e1; border-radius: 0.375rem;
+          background: white; outline: none; transition: border-color 0.15s;
+          width: 100%;
+        }
+        .input:focus { border-color: #64748b; }
+        .input:disabled { background: #f1f5f9; color: #64748b; }
+        .btn-primary {
+          padding: 0.5rem 1.25rem; font-size: 14px; font-weight: 500;
+          background: #16a34a; color: white; border-radius: 0.375rem;
+          transition: background 0.15s;
+        }
+        .btn-primary:hover:not(:disabled) { background: #15803d; }
+        .btn-danger {
+          padding: 0.5rem 1.25rem; font-size: 14px; font-weight: 500;
+          background: #dc2626; color: white; border-radius: 0.375rem;
+        }
+        .btn-danger:hover:not(:disabled) { background: #b91c1c; }
+        .btn-secondary {
+          padding: 0.5rem 1.25rem; font-size: 14px; font-weight: 500;
+          background: white; color: #475569; border: 1px solid #cbd5e1;
+          border-radius: 0.375rem;
+        }
+        .btn-secondary:hover:not(:disabled) { background: #f1f5f9; }
+        .btn-primary:disabled, .btn-danger:disabled, .btn-secondary:disabled {
+          opacity: 0.5; cursor: not-allowed;
+        }
+      `})]})}function Vl({label:m,children:y}){return v.jsxs("div",{children:[m&&v.jsx("label",{className:"block text-sm text-slate-600 mb-1",children:m}),y]})}function uc({title:m,children:y}){return v.jsxs("div",{className:"bg-white p-4 rounded-lg shadow-sm border border-slate-200",children:[v.jsx("h3",{className:"font-medium text-sm text-slate-700 mb-2",children:m}),y]})}const Od={browser_started:"text-slate-500",login_ok:"text-emerald-700",job_found:"text-blue-700",job_skipped:"text-slate-500",letter_sent:"text-emerald-700 font-medium",feed_exhausted:"text-purple-700 font-medium",loop_ended:"text-purple-700 font-medium",error:"text-red-700 font-medium"};function Id({ev:m}){const y=Object.entries(m.payload).map(([c,P])=>`${c}=${JSON.stringify(P)}`).join(" ");return v.jsxs("div",{className:`py-0.5 ${Od[m.kind]??"text-slate-700"}`,children:[v.jsxs("span",{className:"opacity-70",children:["[",m.kind,"]"]})," ",y]})}function Dd(){const[m,y]=oe.useState([]),[c,P]=oe.useState({}),[R,V]=oe.useState(!0),[q,ne]=oe.useState(!1),[$,H]=oe.useState(!1),[B,O]=oe.useState(null);oe.useEffect(()=>{T()},[]);async function T(){V(!0),O(null);try{const{fields:Z}=await qt.getEnvFields();y(Z),P({})}catch(Z){O(String(Z))}finally{V(!1)}}function Y(Z,ye){H(!1),P(ke=>({...ke,[Z]:ye}))}function G(Z,ye){return Z in c?c[Z]:ye}async function A(){ne(!0),O(null);try{await qt.writeEnvFields(c),H(!0),await T()}catch(Z){O(String(Z))}finally{ne(!1)}}const W=Object.keys(c).length>0;return v.jsx("div",{className:"max-w-3xl mx-auto",children:v.jsxs("div",{className:"bg-white p-6 rounded-lg shadow-sm border border-slate-200",children:[v.jsxs("div",{className:"flex items-center justify-between mb-4",children:[v.jsx("h2",{className:"font-semibold",children:"配置 (.env)"}),v.jsxs("div",{className:"flex items-center gap-3",children:[$&&!W&&v.jsx("span",{className:"text-sm text-emerald-600",children:"✓ 已保存"}),v.jsx("button",{onClick:A,disabled:!W||q,className:"px-4 py-1.5 bg-slate-900 text-white rounded text-sm disabled:opacity-50",children:q?"保存中...":"保存"})]})]}),B&&v.jsx("div",{className:"bg-red-50 text-red-700 border border-red-200 rounded p-3 mb-4 text-sm",children:B}),R?v.jsx("div",{className:"text-slate-500",children:"加载中..."}):v.jsx("div",{className:"space-y-3",children:m.map(Z=>v.jsxs("div",{children:[v.jsxs("label",{className:"block text-sm text-slate-600 mb-1",children:[v.jsx("span",{className:"font-mono text-xs bg-slate-100 px-1.5 py-0.5 rounded mr-2",children:Z.key}),Z.label]}),v.jsx("input",{type:Z.isSecret?"password":"text",value:G(Z.key,Z.value),onChange:ye=>Y(Z.key,ye.target.value),className:"w-full px-3 py-1.5 border border-slate-300 rounded text-sm font-mono",placeholder:Z.isSecret?"（已设，留空覆盖为删除）":""})]},Z.key))}),v.jsx("p",{className:"mt-4 text-xs text-slate-500",children:"保存后 .env 立刻更新；但 LLM client 在程序启动时已经读过 API key， 要让新值生效请退出 App 重新打开。"})]})})}function Md(){const[m,y]=oe.useState([]),[c,P]=oe.useState(null),[R,V]=oe.useState(!0),[q,ne]=oe.useState(null);oe.useEffect(()=>{$()},[]);async function $(){V(!0),ne(null);try{const[{letters:H},{summary:B}]=await Promise.all([qt.getLetters(200),qt.getTelemetrySummary()]);y(H.reverse()),P(B)}catch(H){ne(String(H))}finally{V(!1)}}return v.jsxs("div",{className:"max-w-6xl mx-auto space-y-4",children:[v.jsxs("div",{className:"flex items-center justify-between",children:[v.jsx("h2",{className:"font-semibold",children:"历史"}),v.jsx("button",{onClick:$,disabled:R,className:"px-3 py-1 text-sm border border-slate-300 rounded hover:bg-slate-100",children:R?"...":"刷新"})]}),q&&v.jsx("div",{className:"bg-red-50 text-red-700 border border-red-200 rounded p-3 text-sm",children:q}),c&&v.jsx(Fd,{summary:c}),v.jsx("div",{className:"bg-white rounded-lg shadow-sm border border-slate-200 overflow-hidden",children:v.jsxs("table",{className:"w-full text-sm",children:[v.jsx("thead",{className:"bg-slate-50 border-b border-slate-200",children:v.jsxs("tr",{className:"text-left text-xs text-slate-600",children:[v.jsx("th",{className:"px-3 py-2",children:"时间"}),v.jsx("th",{className:"px-3 py-2",children:"Provider"}),v.jsx("th",{className:"px-3 py-2",children:"Model"}),v.jsx("th",{className:"px-3 py-2 text-center",children:"校验"}),v.jsx("th",{className:"px-3 py-2 text-center",children:"发送"}),v.jsx("th",{className:"px-3 py-2",children:"招呼语"})]})}),v.jsx("tbody",{children:m.length===0?v.jsx("tr",{children:v.jsx("td",{colSpan:6,className:"px-3 py-4 text-center text-slate-400",children:R?"加载中...":"logs/letters.jsonl 是空的——还没生成过招呼语"})}):m.map((H,B)=>v.jsx(Ud,{l:H},B))})]})})]})}function Fd({summary:m}){const y=Object.entries(m.by_provider);return v.jsxs("div",{className:"bg-white rounded-lg shadow-sm border border-slate-200 p-4",children:[v.jsx("h3",{className:"font-medium text-sm text-slate-700 mb-3",children:"LLM 调用成本（最近 1000 次）"}),v.jsxs("div",{className:"grid grid-cols-2 md:grid-cols-4 gap-3 text-sm",children:[v.jsx(Wi,{label:"总调用次数",value:m.total_calls.toString()}),v.jsx(Wi,{label:"总成本 (¥)",value:m.total_cost_cny.toFixed(4)}),y.map(([c,P])=>v.jsx(Wi,{label:`${c} (${P.calls} 次)`,value:`¥ ${P.cost_cny.toFixed(4)}`,sub:`in: ${P.input_tokens}  out: ${P.output_tokens}`},c))]})]})}function Wi({label:m,value:y,sub:c}){return v.jsxs("div",{className:"bg-slate-50 rounded p-2",children:[v.jsx("div",{className:"text-xs text-slate-500",children:m}),v.jsx("div",{className:"font-mono text-base mt-0.5",children:y}),c&&v.jsx("div",{className:"text-xs text-slate-400 mt-0.5",children:c})]})}function Ud({l:m}){const[y,c]=oe.useState(!1),P=new Date(m.ts).toLocaleString();return v.jsxs(v.Fragment,{children:[v.jsxs("tr",{className:"border-b border-slate-100 hover:bg-slate-50 cursor-pointer",onClick:()=>c(!y),children:[v.jsx("td",{className:"px-3 py-2 text-xs text-slate-600 whitespace-nowrap",children:P}),v.jsx("td",{className:"px-3 py-2 text-xs",children:m.provider}),v.jsx("td",{className:"px-3 py-2 text-xs",children:m.model}),v.jsx("td",{className:"px-3 py-2 text-center",children:m.validation_ok?v.jsx("span",{className:"text-emerald-600",children:"✓"}):v.jsx("span",{className:"text-red-600",title:m.validation_reasons.join(", "),children:"✗"})}),v.jsx("td",{className:"px-3 py-2 text-center text-xs",children:m.sent?"已发":m.dry_run?"DRY":"—"}),v.jsx("td",{className:"px-3 py-2 text-xs",children:y?m.letter:m.letter.slice(0,60)+(m.letter.length>60?"...":"")})]}),y&&v.jsx("tr",{className:"border-b border-slate-200 bg-slate-50",children:v.jsxs("td",{colSpan:6,className:"px-3 py-3 text-xs",children:[v.jsxs("div",{className:"mb-2",children:[v.jsx("span",{className:"font-medium text-slate-700",children:"JD 预览："}),v.jsxs("span",{className:"text-slate-600 ml-1",children:[m.job_description.slice(0,300),m.job_description.length>300&&"..."]})]}),!m.validation_ok&&v.jsxs("div",{className:"text-red-700",children:[v.jsx("span",{className:"font-medium",children:"校验失败："}),m.validation_reasons.join(", ")]})]})})]})}function Ad(){const[m,y]=oe.useState("run"),c=Tt(P=>P.running);return v.jsxs("div",{className:"min-h-screen flex flex-col",children:[v.jsxs("header",{className:"bg-white border-b border-slate-200 px-6 py-3 flex items-center gap-6",children:[v.jsx("h1",{className:"text-lg font-semibold",children:"BOSS Zhipin Helper"}),v.jsxs("nav",{className:"flex gap-1",children:[v.jsxs(Hi,{current:m,value:"run",onClick:y,children:["运行 ",c&&v.jsx("span",{className:"ml-1 inline-block w-2 h-2 bg-green-500 rounded-full animate-pulse"})]}),v.jsx(Hi,{current:m,value:"config",onClick:y,children:"配置"}),v.jsx(Hi,{current:m,value:"history",onClick:y,children:"历史"})]}),v.jsx("span",{className:"ml-auto text-xs text-slate-500",children:"自动化 Chrome 在另一个窗口，请勿手动关闭"})]}),v.jsxs("main",{className:"flex-1 p-6 overflow-auto",children:[m==="run"&&v.jsx(Td,{}),m==="config"&&v.jsx(Dd,{}),m==="history"&&v.jsx(Md,{})]})]})}function Hi({current:m,value:y,onClick:c,children:P}){const R=m===y;return v.jsx("button",{onClick:()=>c(y),className:`px-3 py-1.5 rounded-md text-sm transition-colors ${R?"bg-slate-900 text-white":"text-slate-600 hover:bg-slate-100"}`,children:P})}md.createRoot(document.getElementById("root")).render(v.jsx(ic.StrictMode,{children:v.jsx(Ad,{})}));
