+++ src/boss_zhipin/tauri/frontend/index.html
@@ -1,0 +1,13 @@
+<!doctype html>
+<html lang="zh-CN">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>BOSS Zhipin Helper</title>
+    <script type="module" crossorigin src="/assets/index-sINFgi6y.js"></script>
+    <link rel="stylesheet" crossorigin href="/assets/index-Ci4qnSmv.css">
+  </head>
+  <body class="bg-slate-50">
+    <div id="root"></div>
+  </body>
+</html>
