+++ src/boss_zhipin/vectorization.py
