+++ src/boss_zhipin_job_search.egg-info/SOURCES.txt
@@ -1,0 +1,37 @@
+README.md
+pyproject.toml
+src/boss_zhipin/__init__.py
+src/boss_zhipin/__main__.py
+src/boss_zhipin/cli.py
+src/boss_zhipin/vectorization.py
+src/boss_zhipin/audit/__init__.py
+src/boss_zhipin/audit/telemetry.py
+src/boss_zhipin/gui/__init__.py
+src/boss_zhipin/gui/env_io.py
+src/boss_zhipin/gui/events.py
+src/boss_zhipin/gui/history.py
+src/boss_zhipin/gui/log_bridge.py
+src/boss_zhipin/gui/runner.py
+src/boss_zhipin/models/job_matcher.py
+src/boss_zhipin/models/llm.py
+src/boss_zhipin/models/openai_assistant.py
+src/boss_zhipin/models/prompts.py
+src/boss_zhipin/tauri/__init__.py
+src/boss_zhipin/tauri/__main__.py
+src/boss_zhipin/utils/__init__.py
+src/boss_zhipin/utils/retry.py
+src/boss_zhipin/website_oper/finding_jobs.py
+src/boss_zhipin/website_oper/write_response.py
+src/boss_zhipin_job_search.egg-info/PKG-INFO
+src/boss_zhipin_job_search.egg-info/SOURCES.txt
+src/boss_zhipin_job_search.egg-info/dependency_links.txt
+src/boss_zhipin_job_search.egg-info/entry_points.txt
+src/boss_zhipin_job_search.egg-info/requires.txt
+src/boss_zhipin_job_search.egg-info/top_level.txt
+tests/test_audit.py
+tests/test_job_matcher.py
+tests/test_main_helpers.py
+tests/test_retry.py
+tests/test_runner.py
+tests/test_telemetry.py
+tests/test_vectorization.py
