+++ src/boss_zhipin_job_search.egg-info/entry_points.txt
@@ -1,0 +1,2 @@
+[console_scripts]
+boss-zhipin = boss_zhipin.cli:_cli_main
