+++ src/boss_zhipin_job_search.egg-info/requires.txt
@@ -1,0 +1,13 @@
+openai>=1.70.0
+packaging>=24.2
+python-dotenv>=1.1.0
+pypdf>=5.4.0
+chromadb>=0.6.3
+sentence-transformers>=3.0.0
+nodriver>=0.48.1
+
+[tauri]
+pydantic>=2.0
+anyio>=4.0
+pytauri==0.8.*
+pytauri-wheel==0.8.*
