+++ src/boss_zhipin_job_search.egg-info/top_level.txt
@@ -1,0 +1,1 @@
+boss_zhipin
