+++ tauri-ui/vite.config.ts
@@ -4,7 +4,7 @@ import react from "@vitejs/plugin-react";
 export default defineConfig({
   plugins: [react()],
   build: {
-    outDir: "../boss_tauri/frontend",
+    outDir: "../src/boss_zhipin/tauri/frontend",
     emptyOutDir: true,
   },
   server: {
